% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_attr}
\alias{add_attr}
\title{add_attr}
\usage{
add_attr(node, obj_parameter, attr_name)
}
\arguments{
\item{node}{The node the attribute should be added to}

\item{obj_parameter}{The value of the attribute to be added}

\item{attr_name}{The name of the attribute to be added}
}
\description{
Adds an attribute to a node attribute list
}
