% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_child}
\alias{add_child}
\title{add_child}
\usage{
add_child(node, obj_parameter, child_name = NULL)
}
\arguments{
\item{node}{The node the child should be added to}

\item{obj_parameter}{The value of the child to be added}

\item{child_name}{Optional: If it's a child node instead of just a value, the name of the child to be added}
}
\description{
Adds a child to a node child list
}
