% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_children}
\alias{add_children}
\title{add_children}
\usage{
add_children(node, children)
}
\arguments{
\item{node}{The node the children should be added to}

\item{children}{The children to be added (a partially named list)}
}
\description{
Adds one or more children to a node child list
}
