% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_opt_attr}
\alias{add_opt_attr}
\title{add_opt_attr}
\usage{
add_opt_attr(node, obj_parameter, attr_name)
}
\arguments{
\item{node}{The node the optional attribute should be added to}

\item{obj_parameter}{The value of the attribute to be added}

\item{attr_name}{The name of the attribute to be added}
}
\description{
Adds an optional attribute to a node attribute list
}
