% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addable_prj_components}
\alias{addable_prj_components}
\title{addable_prj_components}
\usage{
addable_prj_components()
}
\value{
character: Named vector of `OGS6` top level .prj components (names)
along with their respective class names (values)
}
\description{
Returns all possible `OGS6` top level .prj components (names)
along with their respective class names (values).
Change this if you implement new `OGS6` .prj components or delete old ones!
If you implement a new components, you add the following to the `class_names`
vector: `name_of_prj_component = "name_of_your_class"`
}
