% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adopt_nodes}
\alias{adopt_nodes}
\title{adopt_nodes}
\usage{
adopt_nodes(parent_name, obj_list)
}
\arguments{
\item{parent_name}{The name of the new parent node}

\item{obj_list}{A list of class objects (class should have method for generic function as_node)}
}
\description{
Takes a homogenous list of r2ogs6_* objects and creates a wrapper node
using the generic function as_node
}
