% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_null_or_numbers}
\alias{are_null_or_numbers}
\title{are_null_or_numbers}
\usage{
are_null_or_numbers(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if objects are either \code{NULL} or numbers
}
