% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_null_or_numeric}
\alias{are_null_or_numeric}
\title{are_null_or_numeric}
\usage{
are_null_or_numeric(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if objects are either NULL or numeric
}
