% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_null_or_string_flags}
\alias{are_null_or_string_flags}
\title{are_null_or_string_flags}
\usage{
are_null_or_string_flags(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if objects are either NULL or strings reading either
"true" or "false"
}
