% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_null_or_strings}
\alias{are_null_or_strings}
\title{are_null_or_strings}
\usage{
are_null_or_strings(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if objects are either NULL or strings
}
