% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_string_flags}
\alias{are_string_flags}
\title{are_string_flags}
\usage{
are_string_flags(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if objects are strings reading either "true" or "false"
}
