% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_vtu.R
\name{as_node.data_array}
\alias{as_node.data_array}
\title{as_node.data_array}
\usage{
\method{as_node}{data_array}(obj)
}
\description{
Implementation of the generic function as_node for data_array class objects
}
