% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_node.default}
\alias{as_node.default}
\title{as_node.default}
\usage{
\method{as_node}{default}(x, ...)
}
\arguments{
\item{x}{Any object}

\item{...}{Required: The name of the new node (string), Optional: A list of attributes}
}
\description{
Default implementation for S3 generic function as_node, to be used with
atomic objects. So far only returns the value it was given (will be altered later,
at the moment this is needed for the add_children function to work)
}
