% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_node.list}
\alias{as_node.list}
\title{as_node.list}
\usage{
\method{as_node}{list}(x, ...)
}
\arguments{
\item{x}{A named list}

\item{...}{Required: The name of the new node (string), Optional: A vector of attribute flags}
}
\description{
Implementation for S3 generic function as_node for lists. Works only with
named lists containing atomic values and wraps the list so that an element will
become an XML child to a parent node. The name of the list
(= the name of the new parent node) must be supplied.
}
