% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variables.R
\name{as_node.r2ogs6_boundary_condition}
\alias{as_node.r2ogs6_boundary_condition}
\title{as_node.r2ogs6_boundary_condition}
\usage{
\method{as_node}{r2ogs6_boundary_condition}(x)
}
\arguments{
\item{x}{A r2ogs6_boundary_condition class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_boundary_condition
}
