% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_linear_solvers.R
\name{as_node.r2ogs6_linear_solver}
\alias{as_node.r2ogs6_linear_solver}
\title{as_node.r2ogs6_linear_solver}
\usage{
\method{as_node}{r2ogs6_linear_solver}(x)
}
\arguments{
\item{x}{A r2ogs6_linear_solver class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_linear_solver
}
