% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{as_node.r2ogs6_medium}
\alias{as_node.r2ogs6_medium}
\title{as_node.r2ogs6_medium}
\usage{
\method{as_node}{r2ogs6_medium}(x)
}
\arguments{
\item{x}{A r2ogs6_medium class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_medium
}
