% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{as_node.r2ogs6_medium_property}
\alias{as_node.r2ogs6_medium_property}
\title{as_node.r2ogs6_medium_property}
\usage{
\method{as_node}{r2ogs6_medium_property}(obj)
}
\arguments{
\item{obj}{A r2ogs6_medium_property class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_medium_property
}
