% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{as_node.r2ogs6_mesh}
\alias{as_node.r2ogs6_mesh}
\title{as_node.r2ogs6_curve}
\usage{
\method{as_node}{r2ogs6_mesh}(x)
}
\arguments{
\item{x}{A r2ogs6_mesh class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_mesh
}
