% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_nonlinear_solvers.R
\name{as_node.r2ogs6_nonlinear_solver}
\alias{as_node.r2ogs6_nonlinear_solver}
\title{as_node.r2ogs6_nonlinear_solver}
\usage{
\method{as_node}{r2ogs6_nonlinear_solver}(x)
}
\arguments{
\item{x}{A r2ogs6_nonlinear_solver class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_nonlinear_solver
}
