% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_parameters.R
\name{as_node.r2ogs6_parameter}
\alias{as_node.r2ogs6_parameter}
\title{as_node.r2ogs6_parameter}
\usage{
\method{as_node}{r2ogs6_parameter}(x)
}
\arguments{
\item{x}{A r2ogs6_parameter class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_parameter
}
