% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_prj.R
\name{as_node.r2ogs6_prj_medium_phase}
\alias{as_node.r2ogs6_prj_medium_phase}
\title{Method for coercing an r2ogs_prj_medium_phase class object into the structure expected by xml2}
\usage{
\method{as_node}{r2ogs6_prj_medium_phase}(obj)
}
\description{
Method for coercing an r2ogs_prj_medium_phase class object into the structure expected by xml2
}
