% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_prj.R
\name{as_node.r2ogs6_prj_property}
\alias{as_node.r2ogs6_prj_property}
\title{Method for coercing an r2ogs6_prj_property class object into the structure expected by xml2}
\usage{
\method{as_node}{r2ogs6_prj_property}(obj)
}
\description{
Method for coercing an r2ogs6_prj_property class object into the structure expected by xml2
}
