% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{as_node.r2ogs6_time_loop}
\alias{as_node.r2ogs6_time_loop}
\title{as_node.r2ogs6_time_loop}
\usage{
\method{as_node}{r2ogs6_time_loop}(x)
}
\arguments{
\item{x}{A r2ogs6_time_loop class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_time_loop
}
