% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{as_node.r2ogs6_tl_output}
\alias{as_node.r2ogs6_tl_output}
\title{as_node.r2ogs6_tl_output}
\usage{
\method{as_node}{r2ogs6_tl_output}(x)
}
\arguments{
\item{x}{A r2ogs6_tl_output class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_tl_output
}
