% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_test_definition.R
\name{as_node.r2ogs6_vtkdiff}
\alias{as_node.r2ogs6_vtkdiff}
\title{as_node.r2ogs6_vtkdiff}
\usage{
\method{as_node}{r2ogs6_vtkdiff}(x)
}
\arguments{
\item{x}{A r2ogs6_vtkdiff class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_vtkdiff
}
