% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_node}
\alias{as_node}
\title{as_node}
\usage{
as_node(obj)
}
\arguments{
\item{obj}{A class object (must have implementation for as_node)}
}
\description{
S3 generic function, use to turn class data into XML friendly format
}
