% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{build_redux_tree}
\alias{build_redux_tree}
\title{build_redux_tree}
\usage{
build_redux_tree(path, pattern, xpath, required)
}
\arguments{
\item{path}{string: See ?analyse_xml}

\item{pattern}{string: See ?analyse_xml}

\item{xpath}{string: See ?analyse_xml}

\item{required}{flag: Recursion utility}
}
\description{
Builds an XML tree based on the findings of analyse_xml.
This is a recursive function. Handle with care.
}
