% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{check_file_extension}
\alias{check_file_extension}
\title{check_file_extension}
\usage{
check_file_extension(file, expected_extension)
}
\arguments{
\item{file}{A file}

\item{expected_extension}{The expected file extension}
}
\description{
Helper function to check the extension of a file
}
