% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_input_of_name}
\alias{check_for_input_of_name}
\title{check_for_input_of_name}
\usage{
check_for_input_of_name(
  ogs6_obj,
  input_name,
  input_expected,
  do_stop,
  init_func_name = NULL
)
}
\arguments{
\item{ogs6_obj}{The ogs6 object to check}

\item{input_name}{The name of the input to check for}

\item{input_expected}{Should the input exist?}

\item{do_stop}{Should execution be halted or should a boolean be returned?}

\item{init_func_name}{Optional: The name of the function the object can be initialized with}
}
\description{
Checks if a object with the given name was already defined for a ogs6 object and if not,
tells user to initialize one
}
