% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_obj_of_name}
\alias{check_for_obj_of_name}
\title{check_for_obj_of_name}
\usage{
check_for_obj_of_name(ogs6_obj, obj_name)
}
\arguments{
\item{ogs6_obj}{The ogs6 object to check}

\item{obj_name}{The name of the object to check for}
}
\description{
Checks if a object with the given name was already defined for a ogs6 object and if not,
tells user to initialize one
}
