% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_imported_list}
\alias{clean_imported_list}
\title{clean_imported_list}
\usage{
clean_imported_list(list)
}
\arguments{
\item{list}{list: A list}
}
\description{
Cleans an imported list because sometimes strings containing
only newline characters and spaces get imported in
}
