% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_up_imported_list}
\alias{clean_up_imported_list}
\title{clean_up_imported_list}
\usage{
clean_up_imported_list(imported_list)
}
\arguments{
\item{imported_list}{list: A list}
}
\description{
Cleans an imported list because sometimes strings containing
only newline characters and spaces get imported in
}
