% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_names}
\alias{coerce_names}
\title{coerce_names}
\usage{
coerce_names(vector, names)
}
\arguments{
\item{vector}{vector: Vector of parameters}

\item{names}{character: How the vector elements will be named as
per default}
}
\value{
vector: Named vector where the names correspond to \code{names}
}
\description{
Validator function for a parameter vector
}
