% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_xml.R
\name{compare_xml}
\alias{compare_xml}
\title{compare_xml}
\usage{
compare_xml(xml_node_x, xml_node_y)
}
\arguments{
\item{xml_node_x}{xml2::xml_node: Node to compare}

\item{xml_node_y}{xml2::xml_node: Node to compare to}
}
\description{
Helper function to check if 2 XML documents have the same
content. This is a recursive function. Strictly WIP!
}
