% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{create_gml_from_obj}
\alias{create_gml_from_obj}
\title{Creates a .gml XML file based on an GML object}
\usage{
create_gml_from_obj(gml_object, file_name)
}
\arguments{
\item{gml_object}{The specified GML object (of class GMLObject)}

\item{file_name}{The name of the file the XML will be written to}
}
\value{
WIP
}
\description{
Creates a .gml XML file based on an GML object
}
