% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_template_prj}
\alias{create_template_prj}
\title{Creates a template project file based on the XML structure and attributes of an existing project file.}
\usage{
create_template_prj(existing_prj, template_prj)
}
\arguments{
\item{existing_prj}{An existing .prj file to use as a template for the structure and attributes}

\item{template_prj}{The path where the template should be saved}
}
\value{
The newly created template project file
}
\description{
Creates a template project file based on the XML structure and attributes of an existing project file.
}
