% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml.R
\name{data_to_xml}
\alias{data_to_xml}
\title{Wrapper function to create a XML document based on the user input data}
\usage{
data_to_xml(geo_name, points_tibble, polylines_list, surfaces_list)
}
\arguments{
\item{geo_name}{The name of the geometry specified by the user}

\item{points_tibble}{A tibble containing points}

\item{polylines_list}{A list containing polylines}

\item{surfaces_list}{A list containing surfaces}
}
\value{
A XML document ready for export to a file
}
\description{
Wrapper function to create a XML document based on the user input data
}
