% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_vtu.R
\name{decode_appended_data}
\alias{decode_appended_data}
\title{decode_appended_data}
\usage{
decode_appended_data(appended_data, compressor = "")
}
\arguments{
\item{appended_data}{character: AppendedData}

\item{compressor}{string: Optional: How the data was compressed, this is the
`compressor` attribute of *VTKFile*}
}
\value{
list: DataArrays with `data` element which is the decoded data
}
\description{
Decodes AppendedData
}
