% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{delete_keywords_from_str}
\alias{delete_keywords_from_str}
\title{delete_keywords_from_str}
\usage{
delete_keywords_from_str(string)
}
\arguments{
\item{string}{string}
}
\description{
Utility function to delete keywords from a string,
this is important because there is a <repeat> tag in <time_loop> and
"repeat" is a reserved word in R (extend this function if you find more
reserved words)
}
