% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{delete_nulls_from_str}
\alias{delete_nulls_from_str}
\title{delete_nulls_from_str}
\usage{
delete_nulls_from_str(string)
}
\arguments{
\item{string}{A string}
}
\description{
Utility function to delete "param_name = NULL" from a string,
this isn't necessary for functionality of generate_benchmark_script but will
make generated scripts way more readable.
}
