% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_benchmarks.R
\name{download_benchmark}
\alias{download_benchmark}
\title{download_benchmark}
\usage{
download_benchmark(prj_url, path)
}
\arguments{
\item{prj_url}{The URL of a .prj file}

\item{path}{The path the benchmark files should be saved to}
}
\description{
Downloads a single benchmark (consisting of one .prj file and
either one .gml and one .vtu file OR multiple .vtu files)
}
