% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{export_all_sim_files}
\alias{export_all_sim_files}
\title{export_all_sim_files}
\usage{
export_all_sim_files(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}
}
\description{
Exports and / or copies all simulation files to
`ogs6_obj$sim_path`
}
