% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gml.R
\name{export_gml}
\alias{export_gml}
\title{export_gml}
\usage{
export_gml(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object to export r2ogs6_gml object from}
}
\description{
Creates a .gml XML document based on  user input data
}
