% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_prj.R
\name{export_prj}
\alias{export_prj}
\title{export_prj}
\usage{
export_prj(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{...}
}
\description{
Wrapper function to create a .prj XML document based on the user input data
}
