% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_xml_to_file}
\alias{export_xml_to_file}
\title{export_xml_to_file}
\usage{
export_xml_to_file(xml_data, file_name)
}
\arguments{
\item{xml_data}{The data to be exported (already in XML friendly format)}

\item{file_name}{The name of the file to be written}
}
\description{
Export function
}
