% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_invalid_xml}
\alias{filter_invalid_xml}
\title{filter_invalid_xml}
\usage{
filter_invalid_xml(paths, encoding = "ISO-8859-1", print_messages = TRUE)
}
\arguments{
\item{paths}{character: Vector of (maybe-)XML paths}

\item{encoding}{string: Optional: XML encoding. Defaults to ISO-8859-1}

\item{print_messages}{flag: Optional: Print error messages? Defaults to
\code{TRUE}}
}
\value{
character: Vector of invalid XML paths
}
\description{
Filters invalid XML paths out of a vector
}
