% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{find_read_in_func_call}
\alias{find_read_in_func_call}
\title{find_read_in_func_call}
\usage{
find_read_in_func_call(node_name)
}
\arguments{
\item{node_name}{The name of an XML node}
}
\value{
A ready-to-call string if a corresponding function was found, an
empty string otherwise
}
\description{
Checks if a custom read_in function was defined for nodes
with the given name
}
