% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_class.R
\name{flags_to_r6_active_field_str}
\alias{flags_to_r6_active_field_str}
\title{flags_to_r6_active_field_str}
\usage{
flags_to_r6_active_field_str(flags, mutable = TRUE, print_result = FALSE)
}
\arguments{
\item{flags}{vector: Flags}

\item{mutable}{flag: On per default, turn off if parameters are static}

\item{print_result}{flag: Should the result be printed to the console?}
}
\description{
Helper function to generate a string out of a flag vector
}
