% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_class.R
\name{flags_to_r6_init_str}
\alias{flags_to_r6_init_str}
\title{flags_to_r6_init_str}
\usage{
flags_to_r6_init_str(flags, print_result = FALSE)
}
\arguments{
\item{flags}{vector: Flags}

\item{print_result}{flag: Should the result be printed to the console?}
}
\description{
Helper function to generate a string out of a flag vector
}
