% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_class.R
\name{generate_active_field}
\alias{generate_active_field}
\title{generate_active_field}
\usage{
generate_active_field(parameter_name)
}
\arguments{
\item{parameter_name}{The name of the OGS6 class parameter}
}
\description{
Helper function to generate an R6 active field for a OGS6 class
parameter
}
