% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_class.R
\name{generate_add_method}
\alias{generate_add_method}
\title{generate_add_method}
\usage{
generate_add_method(tag_name, parent_tag_name)
}
\arguments{
\item{tag_name}{The tag name of the XML element represented by the class
object}

\item{parent_tag_name}{The tag name of the parent of the XML element
represented by the class object}
}
\description{
Helper function to generate an R6 \code{add_*} method for a
r2ogs6 class object
}
