% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{generate_all_benchmark_scripts}
\alias{generate_all_benchmark_scripts}
\title{generate_all_benchmark_scripts}
\usage{
generate_all_benchmark_scripts(
  path,
  scripts_path = "D:/OGS_scripts/",
  starting_from_prj_path = "",
  skip_prj_paths = character(),
  test_mode = FALSE
)
}
\arguments{
\item{path}{string: Path to a benchmark directory to generate scripts from}

\item{scripts_path}{string: Optional: Path where benchmark scripts will be
saved. Change this to fit your system!}

\item{starting_from_prj_path}{string: Optional:}

\item{skip_prj_paths}{character: Optional: .prj paths to skip}

\item{test_mode}{flag: Optional: In test mode, if `path` is missing,
internal function `get_default_benchmark_path()` will be called}
}
\description{
Wrapper function to generate benchmark scripts from all .prj
files in a directory
}
