% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{generate_all_benchmarks}
\alias{generate_all_benchmarks}
\title{generate_all_benchmarks}
\usage{
generate_all_benchmarks(
  path = "extdata/benchmarks/",
  dest_dir = "extdata/benchmark_scripts/"
)
}
\arguments{
\item{path}{Optional: A path to a benchmark directory to generate scripts
from (defaults to 'extdata/benchmarks' package folder)}

\item{dest_dir}{Optional: The directory to write the scripts to
(defaults to 'extdata/benchmark_scripts' package folder)}
}
\description{
Collector function to generate benchmarks from all .prj files in
a directory
}
