% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_xml.R
\name{generate_as_node_from_element}
\alias{generate_as_node_from_element}
\title{generate_as_node_from_element}
\usage{
generate_as_node_from_element(element_name, attrs = NULL, children = NULL)
}
\arguments{
\item{element_name}{The name of the XML element to base the function on}

\item{attrs}{Optional: A list of element attributes}

\item{children}{Optional: A list of element children}
}
\description{
Generates a method for the generic function as_node based on an XML element
}
