% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_functions.R
\name{generate_as_node_func}
\alias{generate_as_node_func}
\title{generate_as_node_func}
\usage{
generate_as_node_func(
  file,
  element_name,
  subclasses = NULL,
  show_result = TRUE
)
}
\arguments{
\item{file}{The XML file to parse}

\item{element_name}{The name of the XML element to base the method on}

\item{subclasses}{Optional: A named vector of subclasses}

\item{show_result}{Should the generated function be printed to the console?}
}
\description{
Generates a method for the generic function as_node based on an XML element
}
