% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_functions.R
\name{generate_as_node_method_content}
\alias{generate_as_node_method_content}
\title{generate_as_node_method_content}
\usage{
generate_as_node_method_content(
  element,
  iteration_depth = 0,
  subclasses = NULL
)
}
\arguments{
\item{element}{The XML element to base the method on}

\item{iteration_depth}{Utility parameter for the recursion, just leave this alone!}

\item{subclasses}{Optional: A named vector of subclasses}
}
\description{
...
}
