% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_class.R
\name{generate_helper}
\alias{generate_helper}
\title{generate_helper}
\usage{
generate_helper(tag_name, param_flags, prefix = "", print_result = FALSE)
}
\arguments{
\item{tag_name}{The name of the XML element the helper will be based on}

\item{param_flags}{The parameters for the class and if they are required or
not (i.e. 'c(a = TRUE, b = FALSE)')}

\item{prefix}{Optional: For subclasses whose represented elements have
the same tag name as an element for which a class was already specified,
a prefix must be appended to the class name}

\item{print_result}{flag: Should the result be printed to the console?}
}
\description{
Helper function to generate a helper out of a tag name
and a flag vector
}
