% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_functions.R
\name{generate_simple_read_in}
\alias{generate_simple_read_in}
\title{generate_simple_read_in}
\usage{
generate_simple_read_in(
  element_name,
  child_name,
  has_name_tag = TRUE,
  subclasses_names = NULL
)
}
\arguments{
\item{element_name}{The name of the .prj element to be read from (wrapper element, e.g. 'processes')}

\item{child_name}{The name of the element children (e.g. 'process')}

\item{has_name_tag}{Do the child elements have a child element with the name 'name'?}

\item{subclasses_names}{Optional: A character vector containing the names of r2ogs6_*
subclasses (r2ogs6_* classes without a method for input_add)}
}
\description{
Assuming function read_in gets good enough results, this could save code later.
}
