% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{generate_structured_mesh}
\alias{generate_structured_mesh}
\title{generate_structured_mesh}
\usage{
generate_structured_mesh(args_str, ogs_bin_path)
}
\arguments{
\item{args_str}{string: The arguments the script will be called with}

\item{ogs_bin_path}{string: Optional: Path to OpenGeoSys6 bin folder.
Defaults to options("r2ogs6.default_ogs_bin_path").}
}
\value{
string: .vtu file path
}
\description{
Wrapper function to call generateStructuredMesh.exe
(VTK mesh generator). For full documentation see
https://www.opengeosys.org/docs/tools/meshing/structured-mesh-generation/
}
