% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{get_benchmark_paths}
\alias{get_benchmark_paths}
\title{get_benchmark_paths}
\usage{
get_benchmark_paths(ogs6_processlib_path)
}
\arguments{
\item{ogs6_processlib_path}{string: Path to OpenGeoSys 6 ProcessLib folder
which contains relevant Tests.cmake files}
}
\description{
Gets paths to all benchmarks that should work
}
