% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{get_class_args}
\alias{get_class_args}
\title{get_class_args}
\usage{
get_class_args(class_name)
}
\arguments{
\item{class_name}{string: The name of a class}
}
\value{
character: Named vector of class arguments
}
\description{
Gets class arguments
}
