% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_class_from_xpath}
\alias{get_class_from_xpath}
\title{get_class_from_xpath}
\usage{
get_class_from_xpath(xpath)
}
\arguments{
\item{xpath}{string: An xpath expression. Works for path-like xpaths only}
}
\value{
string: The class name.
}
\description{
Gets r2ogs6 class name from an xpath-like expression
}
