% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_class_tag_name}
\alias{get_class_tag_name}
\title{get_class_tag_name}
\usage{
get_class_tag_name(class_name)
}
\arguments{
\item{class_name}{string: The name of a r2ogs6 class}
}
\value{
string: The tag name corresponding to class_name
}
\description{
Utility function, returns the tag name of a r2ogs6 class
}
