% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{get_default_benchmark_path}
\alias{get_default_benchmark_path}
\title{get_default_benchmark_path}
\usage{
get_default_benchmark_path()
}
\description{
Utility function for testing, change this to fit your system!
}
