% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{get_grandchild_length_vector}
\alias{get_grandchild_length_vector}
\title{get_grandchild_length_vector}
\usage{
get_grandchild_length_vector(xml_node)
}
\arguments{
\item{xml_node}{An XML node}
}
\value{
A numeric vector containing the number of children of each child of xml_node
}
\description{
Helper function to check the number of children of children of an XML node
(i.e. grandchildren)
}
