% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_implemented_classes}
\alias{get_implemented_classes}
\title{get_implemented_classes}
\usage{
get_implemented_classes()
}
\description{
Utility function, returns the names of all classes implemented
so far. Change this if you implement new classes or delete old ones!
If you implement a new class, you add the following to the character vector:
<name_of_corresponding_OGS6_parameter> = <name_of_your_class>
}
