% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_list_status}
\alias{get_list_status}
\title{get_list_status}
\usage{
get_list_status(flag, obj_list, element_type = "list element", is_opt = FALSE)
}
\arguments{
\item{flag}{Boolean flag to keep track of missing components}

\item{obj_list}{The specified list}

\item{element_type}{Optional: What kind of elements are in the list?}

\item{is_opt}{flag: Does the list need at least one element?}
}
\description{
Helper function for get_status() to check if a list has at least
one element.
}
