% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nonstandard_tag_names}
\alias{get_nonstandard_tag_names}
\title{get_nonstandard_tag_names}
\usage{
get_nonstandard_tag_names()
}
\value{
character: The tag names of classes that are not named after the
convention r2ogs6_<tag name> because there already is a class with that name.
If you as a dev create new classes like that, just add them to the list :)
}
\description{
Utility function, returns nonstandard tag names
}
