% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_obj_status}
\alias{get_obj_status}
\title{get_obj_status}
\usage{
get_obj_status(flag, obj)
}
\arguments{
\item{flag}{flag: To keep track of missing components}

\item{obj}{OGS6 parameter: Either a wrapper list or an r2ogs6 class object}
}
\description{
Helper function for get_status() to check the status of an OGS6
parameter
}
