% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_path_sublist}
\alias{get_path_sublist}
\title{get_path_sublist}
\usage{
get_path_sublist(prj_paths, starting_from_prj_path)
}
\arguments{
\item{prj_paths}{character: .prj paths}

\item{starting_from_prj_path}{string: .prj path to start from}
}
\value{

}
\description{
Helper function to speed up tests
}
