% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_xml.R
\name{get_required}
\alias{get_required}
\title{get_required}
\usage{
get_required(names, occurence_probabilities)
}
\arguments{
\item{names}{A vector of names}

\item{occurence_probabilities}{A vector of occurrence probabilities}
}
\description{
Helper function to mark required attributes or children from a vector of names and a
vector of occurrence probabilities
}
