% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subclass_names}
\alias{get_subclass_names}
\title{get_subclass_names}
\usage{
get_subclass_names(class_name, file_type = ".prj")
}
\arguments{
\item{class_name}{string: The name of a r2ogs6 class}
}
\value{
character: The names of the subclasses as a character vector
(empty if there are none)
}
\description{
Utility function, returns the names of the subclasses
of a r2ogs6 class
}
