% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tag_class_name}
\alias{get_tag_class_name}
\title{get_tag_class_name}
\usage{
get_tag_class_name(tag_name)
}
\arguments{
\item{tag_name}{string: An XML tag}
}
\value{
string: The class name corresponding to tag_name
}
\description{
Utility function, returns the class name of an XML tag
}
