% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tag_from_class}
\alias{get_tag_from_class}
\title{get_tag_from_class}
\usage{
get_tag_from_class(class_name)
}
\arguments{
\item{class_name}{string: The name of a r2ogs6 class}
}
\value{
string: The tag name corresponding to `class_name`
}
\description{
Utility function, returns the tag name of a r2ogs6 class
}
