% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tag_from_xpath}
\alias{get_tag_from_xpath}
\title{get_tag_from_xpath}
\usage{
get_tag_from_xpath(xpath)
}
\arguments{
\item{xpath}{string: An xpath expression. Works for path-like xpaths only}
}
\value{
string: The XML tag name
}
\description{
Gets the XML tag name from an xpath expression
}
\examples{
  get_tag_from_xpath("processes/process")
}
