% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_vtu_categories}
\alias{get_valid_vtu_categories}
\title{get_valid_vtu_categories}
\usage{
get_valid_vtu_categories()
}
\value{
character: `OGS6_vtu` categories
}
\description{
Returns all valid `OGS6_vtu` categories.
}
