% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_value_types}
\alias{get_value_types}
\title{get_value_types}
\usage{
get_value_types(xml_node)
}
\arguments{
\item{xml_node}{An XML node (of class xml2::xml_node)}
}
\description{
Gets the type of an XML value based on the documentation
(per default, XML values are read in as a string, but for many elements,
we want to coerce them to double)
}
