% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_vtu.R
\name{get_vtu_data_arrays_from_file}
\alias{get_vtu_data_arrays_from_file}
\title{get_vtu_data_arrays_from_file}
\usage{
get_vtu_data_arrays_from_file(
  vtu_path,
  Names = character(),
  categories = character()
)
}
\arguments{
\item{vtu_path}{string: .vtu file path}

\item{Names}{character: Optional: Select `DataArray` elements by `Name`
attribute}

\item{categories}{character: Optional: One or more of `FieldData`, `PointData`,
`CellData`, `Points` or `Cells`. If left empty, will get `DataArray`
elements from whole XML document.}
}
\description{
Reads DataArray elements from a .vtu file
}
