% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_xml_encoding}
\alias{get_xml_encoding}
\title{Reads the encoding of an XML file}
\usage{
get_xml_encoding(input_file_name)
}
\arguments{
\item{input_file_name}{An XML (.xml or, in the case of OpenGeoSys, .prj) file}
}
\value{
The encoding of the input file
}
\description{
Reads the encoding of an XML file
}
