% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gml.R
\name{gml_data_to_xml}
\alias{gml_data_to_xml}
\title{gml_data_to_xml}
\usage{
gml_data_to_xml(gml_obj)
}
\arguments{
\item{gml_obj}{A gml class object}
}
\value{
A XML document ready for export to a file
}
\description{
Wrapper function to create a .gml XML document based on the user input data
}
