% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{GMLPoints}
\alias{GMLPoints}
\title{Class describing a set of points}
\description{
Class describing a set of points

Class describing a set of points
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLPoints$new()}}
\item \href{#method-validate_ids}{\code{GMLPoints$validate_ids()}}
\item \href{#method-validate_dim}{\code{GMLPoints$validate_dim()}}
\item \href{#method-validate_coordinates}{\code{GMLPoints$validate_coordinates()}}
\item \href{#method-as_list}{\code{GMLPoints$as_list()}}
\item \href{#method-as_node}{\code{GMLPoints$as_node()}}
\item \href{#method-print}{\code{GMLPoints$print()}}
\item \href{#method-clone}{\code{GMLPoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$new(gml_points)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_ids"></a>}}
\if{latex}{\out{\hypertarget{method-validate_ids}{}}}
\subsection{Method \code{validate_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$validate_ids(gml_points)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_dim"></a>}}
\if{latex}{\out{\hypertarget{method-validate_dim}{}}}
\subsection{Method \code{validate_dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$validate_dim(gml_points)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_coordinates"></a>}}
\if{latex}{\out{\hypertarget{method-validate_coordinates}{}}}
\subsection{Method \code{validate_coordinates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$validate_coordinates(gml_points)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$as_list(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_node"></a>}}
\if{latex}{\out{\hypertarget{method-as_node}{}}}
\subsection{Method \code{as_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$as_node(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
