% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{GMLPolyline}
\alias{GMLPolyline}
\title{Class describing a single polyline}
\description{
Class describing a single polyline

Class describing a single polyline
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLPolyline$new()}}
\item \href{#method-as_node}{\code{GMLPolyline$as_node()}}
\item \href{#method-print}{\code{GMLPolyline$print()}}
\item \href{#method-clone}{\code{GMLPolyline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolyline$new(id = 0, name = "front_left", point1 = 0, point2 = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_node"></a>}}
\if{latex}{\out{\hypertarget{method-as_node}{}}}
\subsection{Method \code{as_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolyline$as_node(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolyline$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolyline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
