% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{GMLPolylines}
\alias{GMLPolylines}
\title{Class describing a set of polylines}
\description{
Class describing a set of polylines

Class describing a set of polylines
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLPolylines$new()}}
\item \href{#method-validate_ids}{\code{GMLPolylines$validate_ids()}}
\item \href{#method-validate_polylines}{\code{GMLPolylines$validate_polylines()}}
\item \href{#method-as_node}{\code{GMLPolylines$as_node()}}
\item \href{#method-clone}{\code{GMLPolylines$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolylines$new(gml_polylines)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_ids"></a>}}
\if{latex}{\out{\hypertarget{method-validate_ids}{}}}
\subsection{Method \code{validate_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolylines$validate_ids(gml_polylines)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_polylines"></a>}}
\if{latex}{\out{\hypertarget{method-validate_polylines}{}}}
\subsection{Method \code{validate_polylines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolylines$validate_polylines(gml_polylines)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_node"></a>}}
\if{latex}{\out{\hypertarget{method-as_node}{}}}
\subsection{Method \code{as_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolylines$as_node(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLPolylines$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
