% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{GMLSurface}
\alias{GMLSurface}
\title{Class describing a single surface (using 2 GMLSurfaceElements)}
\description{
Class describing a single surface (using 2 GMLSurfaceElements)

Class describing a single surface (using 2 GMLSurfaceElements)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLSurface$new()}}
\item \href{#method-validate_elements}{\code{GMLSurface$validate_elements()}}
\item \href{#method-as_node}{\code{GMLSurface$as_node()}}
\item \href{#method-clone}{\code{GMLSurface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurface$new(
  id = 0,
  name = "left",
  gml_surface_element1 = GMLSurfaceElement$new(),
  gml_surface_element2 = GMLSurfaceElement$new(p3 = 3)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_elements"></a>}}
\if{latex}{\out{\hypertarget{method-validate_elements}{}}}
\subsection{Method \code{validate_elements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurface$validate_elements(gml_surface_element1, gml_surface_element2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_node"></a>}}
\if{latex}{\out{\hypertarget{method-as_node}{}}}
\subsection{Method \code{as_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurface$as_node()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
