% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{GMLSurfaceElement}
\alias{GMLSurfaceElement}
\title{Class describing a single surface element (using 3 point IDs)}
\description{
Class describing a single surface element (using 3 point IDs)

Class describing a single surface element (using 3 point IDs)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLSurfaceElement$new()}}
\item \href{#method-as_list}{\code{GMLSurfaceElement$as_list()}}
\item \href{#method-as_node}{\code{GMLSurfaceElement$as_node()}}
\item \href{#method-clone}{\code{GMLSurfaceElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaceElement$new(p1 = 0, p2 = 1, p3 = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaceElement$as_list(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_node"></a>}}
\if{latex}{\out{\hypertarget{method-as_node}{}}}
\subsection{Method \code{as_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaceElement$as_node(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaceElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
