% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml_R6.R
\name{GMLSurfaces}
\alias{GMLSurfaces}
\title{Class describing a set of surfaces}
\description{
Class describing a set of surfaces

Class describing a set of surfaces
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GMLSurfaces$new()}}
\item \href{#method-validate_ids}{\code{GMLSurfaces$validate_ids()}}
\item \href{#method-as_node}{\code{GMLSurfaces$as_node()}}
\item \href{#method-clone}{\code{GMLSurfaces$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaces$new(gml_surfaces)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_ids"></a>}}
\if{latex}{\out{\hypertarget{method-validate_ids}{}}}
\subsection{Method \code{validate_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaces$validate_ids(gml_surfaces)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_node"></a>}}
\if{latex}{\out{\hypertarget{method-as_node}{}}}
\subsection{Method \code{as_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaces$as_node(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMLSurfaces$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
