% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_curves.R
\name{input_add.r2ogs6_curve}
\alias{input_add.r2ogs6_curve}
\title{input_add.r2ogs6_curve}
\usage{
\method{input_add}{r2ogs6_curve}(obj, ogs6_obj)
}
\arguments{
\item{obj}{A r2ogs6_curve class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_curve
}
