% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{input_add.r2ogs6_gml}
\alias{input_add.r2ogs6_gml}
\title{input_add.r2ogs6_gml}
\usage{
\method{input_add}{r2ogs6_gml}(x, ogs6_obj)
}
\arguments{
\item{x}{A r2ogs6_gml class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_gml
}
