% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{input_add.r2ogs6_medium}
\alias{input_add.r2ogs6_medium}
\title{input_add.r2ogs6_medium}
\usage{
\method{input_add}{r2ogs6_medium}(x, ogs6_obj)
}
\arguments{
\item{x}{A r2ogs6_medium class object}

\item{ogs6_obj}{An OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_medium
}
