% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_nonlinear_solvers.R
\name{input_add.r2ogs6_nonlinear_solver}
\alias{input_add.r2ogs6_nonlinear_solver}
\title{input_add.r2ogs6_nonlinear_solver}
\usage{
\method{input_add}{r2ogs6_nonlinear_solver}(obj, ogs6_obj)
}
\arguments{
\item{obj}{A r2ogs6_nonlinear_solver class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_nonlinear_solver
}
