% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_parameters.R
\name{input_add.r2ogs6_parameter}
\alias{input_add.r2ogs6_parameter}
\title{input_add.r2ogs6_parameter}
\usage{
\method{input_add}{r2ogs6_parameter}(x, ogs6_obj)
}
\arguments{
\item{x}{A r2ogs6_parameter class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_parameter
}
