% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_processes.R
\name{input_add.r2ogs6_process}
\alias{input_add.r2ogs6_process}
\title{input_add.r2ogs6_process}
\usage{
\method{input_add}{r2ogs6_process}(x, ogs6_obj)
}
\arguments{
\item{x}{A r2ogs6_process class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_process
}
