% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variables.R
\name{input_add.r2ogs6_process_variable}
\alias{input_add.r2ogs6_process_variable}
\title{input_add.r2ogs6_process_variable}
\usage{
\method{input_add}{r2ogs6_process_variable}(obj, ogs6_obj)
}
\arguments{
\item{obj}{A r2ogs6_process_variable class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_process_variable
}
