% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_test_definition.R
\name{input_add.r2ogs6_vtkdiff}
\alias{input_add.r2ogs6_vtkdiff}
\title{input_add.r2ogs6_vtkdiff}
\usage{
\method{input_add}{r2ogs6_vtkdiff}(x, ogs6_obj)
}
\arguments{
\item{x}{A r2ogs6_vtkdiff class object}

\item{ogs6_obj}{A OGS6 class object}
}
\description{
Implementation of generic function input_add for S3 class r2ogs6_vtkdiff
}
