% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gml.R
\name{input_add_gml_obj}
\alias{input_add_gml_obj}
\title{Adds an empty gml class object to a ogs6 class object input list}
\usage{
input_add_gml_obj(ogs6_obj, geometry_name)
}
\arguments{
\item{ogs6_obj}{The ogs6 object the gml class object should be added to}

\item{geometry_name}{The name of the geometry specified by the gml class object}
}
\description{
Adds an empty gml class object to a ogs6 class object input list
}
