% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gml.R
\name{input_add_gml_points}
\alias{input_add_gml_points}
\title{Adds gml points to a ogs6 class object}
\usage{
input_add_gml_points(ogs6_obj, gml_points)
}
\arguments{
\item{ogs6_obj}{A ogs6 class object}

\item{gml_points}{A tibble of gml points}
}
\description{
Adds gml points to a ogs6 class object
}
