% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gml.R
\name{input_add_gml_polylines}
\alias{input_add_gml_polylines}
\title{Adds gml polylines to a ogs6 class object}
\usage{
input_add_gml_polylines(ogs6_obj, gml_polylines)
}
\arguments{
\item{ogs6_obj}{A ogs6 class object}

\item{gml_polylines}{A list of gml polylines}
}
\description{
Adds gml polylines to a ogs6 class object
}
