% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gml.R
\name{input_add_gml_surfaces}
\alias{input_add_gml_surfaces}
\title{Adds gml surfaces to a ogs6 class object}
\usage{
input_add_gml_surfaces(ogs6_obj, gml_surfaces)
}
\arguments{
\item{ogs6_obj}{A ogs6 class object}

\item{gml_surfaces}{A list of gml surfaces}
}
\description{
Adds gml surfaces to a ogs6 class object
}
