% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_prj.R
\name{input_add_prj_obj}
\alias{input_add_prj_obj}
\title{input_add_prj_obj}
\usage{
input_add_prj_obj(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{The ogs6 object the prj class object should be added to}
}
\description{
Adds an empty prj class object to a ogs6 class object input list
}
