% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_vtu.R
\name{input_add_vtu_appended_data}
\alias{input_add_vtu_appended_data}
\title{input_add_vtu_appended_data}
\usage{
input_add_vtu_appended_data(ogs6_obj, encoding, data)
}
\arguments{
\item{ogs6_obj}{The ogs6 object the .vtu AppendedData should be added to}

\item{encoding}{How the AppendedData is encoded}

\item{data}{The actual data in the specified encoding}
}
\description{
Adds .vtu AppendedData to a ogs6 class object
}
