% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_vtu.R
\name{input_add_vtu_obj}
\alias{input_add_vtu_obj}
\title{input_add_vtu_obj}
\usage{
input_add_vtu_obj(
  ogs6_obj,
  type,
  version,
  byte_order,
  header_type,
  compressor = NULL
)
}
\arguments{
\item{ogs6_obj}{The ogs6 object the vtu class object should be added to}

\item{type}{The type of the .vtu file (e. g. "UnstructuredGrid")}

\item{version}{The version of the .vtu file (e. g. 1.0)}

\item{byte_order}{Either "LittleEndian" or "BigEndian"}

\item{header_type}{The header type (e. g. "UInt32")}

\item{compressor}{Optional: The compressor to be used}
}
\description{
Adds an empty vtu class object to a ogs6 class object input.
}
