% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_vtu.R
\name{input_add_vtu_unstructured_grid}
\alias{input_add_vtu_unstructured_grid}
\title{input_add_vtu_unstructured_grid}
\usage{
input_add_vtu_unstructured_grid(ogs6_obj, piece, field_data = NULL)
}
\arguments{
\item{ogs6_obj}{The ogs6 object the .vtu UnstructuredGrid should be added to}

\item{piece}{A Piece (WIP)}

\item{field_data}{FieldData (WIP)}
}
\description{
Adds .vtu UnstructuredGrid to a ogs6 class object
}
