% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_null_or_coerce_names}
\alias{is_null_or_coerce_names}
\title{is_null_or_coerce_names}
\usage{
is_null_or_coerce_names(obj, names)
}
\arguments{
\item{obj}{A list (or vector) of parameters}

\item{names}{How the list elements will be named as per default}
}
\description{
Validator function for a parameter list or vector or NULL
}
