% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_null_or_has_class}
\alias{is_null_or_has_class}
\title{is_null_or_has_class}
\usage{
is_null_or_has_class(obj, class_name)
}
\arguments{
\item{obj}{The object to check}

\item{class_name}{The name of the expected class}
}
\description{
Checks if an object is either \code{NULL} or a class object of
class \code{class_name}
}
