% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_null_or_wrapper_list}
\alias{is_null_or_wrapper_list}
\title{is_null_or_wrapper_list}
\usage{
is_null_or_wrapper_list(obj, element_class)
}
\arguments{
\item{obj}{list | NULL: Object to check}

\item{element_class}{string: Class each element of `obj` should have}
}
\description{
Checks if an object is either NULL or a list of elements
of class `element_class`
}
