% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_optional_sim_component}
\alias{is_optional_sim_component}
\title{is_optional_sim_component}
\usage{
is_optional_sim_component(ogs6_parameter_name)
}
\arguments{
\item{ogs6_parameter_name}{string: Name of a OGS6 parameter}
}
\description{
Checks if a simulation component is optional
}
