% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wrapper}
\alias{is_wrapper}
\title{is_wrapper}
\usage{
is_wrapper(ogs6_param_name)
}
\arguments{
\item{ogs6_param_name}{string: Name of an OGS6 parameter}
}
\description{
Helper function, checks if an OGS6 parameter is a wrapper
}
