% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wrapper_list}
\alias{is_wrapper_list}
\title{is_wrapper_list}
\usage{
is_wrapper_list(list, element_class)
}
\arguments{
\item{list}{list: List to check}

\item{element_class}{string: Class each element of `list` should have}
}
\description{
Checks if a list consists only of elements of class
`element_class`
}
