% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_has_element}
\alias{list_has_element}
\title{list_has_element}
\usage{
list_has_element(obj_list, element_type = "list element")
}
\arguments{
\item{obj_list}{The specified list}

\item{element_type}{Optional: What kind of elements are in the list?}
}
\description{
Helper function for get_status() to check if a list has at least one element.
}
