% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{make_call_from_parameters}
\alias{make_call_from_parameters}
\title{make_call_from_parameters}
\usage{
make_call_from_parameters(class_name, parameters)
}
\arguments{
\item{class_name}{Which class constructor should be called?}

\item{parameters}{A named list of parameters}
}
\description{
Helper function to construct a call from a class name and a
list of class parameters
}
