% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_vtu.R
\name{new_data_array}
\alias{new_data_array}
\title{new_data_array}
\usage{
new_data_array(
  type,
  name,
  format,
  range_min,
  range_max,
  number_of_components = NULL,
  number_of_tuples = NULL,
  offset = NULL,
  data = NULL
)
}
\description{
S3 class constructor, returns S3 class describing a .vtu DataArray element
}
