% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gml.R
\name{new_gml}
\alias{new_gml}
\title{Class describing the gml file and any parameters defined in one}
\usage{
new_gml(geometry_name = NULL, points = NULL, polylines = NULL, surfaces = NULL)
}
\description{
Class describing the gml file and any parameters defined in one
}
