% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{new_ogs6}
\alias{new_ogs6}
\title{Constructor for the ogs6 base class}
\usage{
new_ogs6(
  sim_io = list(input = list(), output = list()),
  sim_name = character(),
  sim_id = integer(),
  sim_path = character()
)
}
\arguments{
\item{sim_io}{...}

\item{sim_name}{A string value representing the simulation name}

\item{sim_id}{An integer value representing the simulation ID}

\item{sim_path}{A string value describing the path where the IO files will be saved}
}
\description{
Constructor for the ogs6 base class
}
