% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variables.R
\name{new_r2ogs6_boundary_condition}
\alias{new_r2ogs6_boundary_condition}
\title{new_r2ogs6_boundary_condition}
\usage{
new_r2ogs6_boundary_condition(
  type,
  parameter,
  component = NULL,
  mesh = NULL,
  geometrical_set = NULL,
  geometry = NULL
)
}
\description{
Constructor for S3 class r2ogs6_boundary_condition
}
