% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_curve.R
\name{new_r2ogs6_curve}
\alias{new_r2ogs6_curve}
\title{new_r2ogs6_curve}
\usage{
new_r2ogs6_curve(name, coords, values)
}
\arguments{
\item{name}{The name of the curve}

\item{coords}{Coordinates at which the curve's values are given}

\item{values}{Values of the curve at the given coordinates}
}
\description{
Constructor for S3 class r2ogs6_curve
}
