% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{new_r2ogs6_gml}
\alias{new_r2ogs6_gml}
\title{new_r2ogs6_gml}
\usage{
new_r2ogs6_gml(name, points, polylines = NULL, surfaces = NULL)
}
\arguments{
\item{name}{The name of the geometry}

\item{points}{A tibble of points}

\item{polylines}{Optional: A list of polylines}

\item{surfaces}{Optional: A list of surfaces}
}
\description{
Constructor for S3 class new_r2ogs6_gml
}
