% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_bar}
\alias{new_r2ogs6_gml_bar}
\title{Constructor based on HydroMechanics/StaggeredScheme/InjectionProduction1D/bar.gml}
\usage{
new_r2ogs6_gml_bar()
}
\description{
Constructor based on HydroMechanics/StaggeredScheme/InjectionProduction1D/bar.gml
}
