% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_circle_1_axi}
\alias{new_r2ogs6_gml_circle_1_axi}
\title{Constructor based on ogs6_benchmarks/Elliptic/circle_radius_1/circle_1_axi.gml}
\usage{
new_r2ogs6_gml_circle_1_axi()
}
\description{
Constructor based on ogs6_benchmarks/Elliptic/circle_radius_1/circle_1_axi.gml
}
