% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_cube_1x1x1}
\alias{new_r2ogs6_gml_cube_1x1x1}
\title{Constructor based on HydroMechanics/IdealGas/flow_free_expansion/cube_1x1x1.gml}
\usage{
new_r2ogs6_gml_cube_1x1x1()
}
\description{
Constructor based on HydroMechanics/IdealGas/flow_free_expansion/cube_1x1x1.gml
}
