% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_hm1_1Dbeam}
\alias{new_r2ogs6_gml_hm1_1Dbeam}
\title{Constructor based on HydroMechanics/Verification/hm1_1Dbeam.gml}
\usage{
new_r2ogs6_gml_hm1_1Dbeam()
}
\description{
Constructor based on HydroMechanics/Verification/hm1_1Dbeam.gml
}
