% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_hm1_2Dsquare}
\alias{new_r2ogs6_gml_hm1_2Dsquare}
\title{Constructor based on HydroMechanics/Verification/hm1_2Dsquare.gml}
\usage{
new_r2ogs6_gml_hm1_2Dsquare()
}
\description{
Constructor based on HydroMechanics/Verification/hm1_2Dsquare.gml
}
