% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_hm1_3Dcube}
\alias{new_r2ogs6_gml_hm1_3Dcube}
\title{Constructor based on HydroMechanics/Verification/hm1_3Dcube.gml}
\usage{
new_r2ogs6_gml_hm1_3Dcube()
}
\description{
Constructor based on HydroMechanics/Verification/hm1_3Dcube.gml
}
