% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_quad_1x10}
\alias{new_r2ogs6_gml_quad_1x10}
\title{Constructor based on HydroMechanics/IdealGas/flow_pressure_boundary/quad_1x10.gml}
\usage{
new_r2ogs6_gml_quad_1x10()
}
\description{
Constructor based on HydroMechanics/IdealGas/flow_pressure_boundary/quad_1x10.gml
}
