% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml_prefabs.R
\name{new_r2ogs6_gml_square_1x1}
\alias{new_r2ogs6_gml_square_1x1}
\title{Constructor based on HydroMechanics/IdealGas/flow_no_strain/square_1x1.gml}
\usage{
new_r2ogs6_gml_square_1x1()
}
\description{
Constructor based on HydroMechanics/IdealGas/flow_no_strain/square_1x1.gml
}
