% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_linear_solvers.R
\name{new_r2ogs6_linear_solver}
\alias{new_r2ogs6_linear_solver}
\title{new_r2ogs6_linear_solver}
\usage{
new_r2ogs6_linear_solver(name, eigen, lis = NULL, petsc = NULL)
}
\description{
Constructor for S3 class r2ogs6_linear_solver
}
