% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{new_r2ogs6_medium}
\alias{new_r2ogs6_medium}
\title{new_r2ogs6_medium}
\usage{
new_r2ogs6_medium(phases, properties, id = NULL)
}
\arguments{
\item{phases}{A list of medium phases (r2ogs6_medium_phase objects)}

\item{properties}{A list of medium properties (r2ogs6_medium_property objects)}

\item{id}{Optional: The medium ID}
}
\description{
new_r2ogs6_medium
}
