% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{new_r2ogs6_medium_property}
\alias{new_r2ogs6_medium_property}
\title{new_r2ogs6_medium_property}
\usage{
new_r2ogs6_medium_property(name, type, value = NULL, ...)
}
\description{
Constructor for S3 class r2ogs6_medium_property
}
