% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_nonlinear_solvers.R
\name{new_r2ogs6_nonlinear_solver}
\alias{new_r2ogs6_nonlinear_solver}
\title{new_r2ogs6_nonlinear_solver}
\usage{
new_r2ogs6_nonlinear_solver(name, type, max_iter, linear_solver)
}
\description{
Constructor for S3 class r2ogs6_nonlinear_solver
}
