% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_prj.R
\name{new_r2ogs6_prj_medium_phase}
\alias{new_r2ogs6_prj_medium_phase}
\title{Defines a .prj medium phase element}
\usage{
new_r2ogs6_prj_medium_phase(type = character(), properties = list())
}
\arguments{
\item{type}{A string specifying the medium type (valid types: ...)}

\item{properties}{A list of properties (see ?r2ogs6_property for more info)}
}
\description{
Defines a .prj medium phase element
}
