% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variables.R
\name{new_r2ogs6_process_variable}
\alias{new_r2ogs6_process_variable}
\title{new_r2ogs6_process_variable}
\usage{
new_r2ogs6_process_variable(
  name,
  components,
  order,
  initial_condition,
  boundary_conditions
)
}
\description{
Constructor for S3 class r2ogs6_process_variable
}
