% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{new_r2ogs6_time_loop}
\alias{new_r2ogs6_time_loop}
\title{new_r2ogs6_time_loop}
\usage{
new_r2ogs6_time_loop(processes, output, global_processes_coupling = NULL)
}
\description{
Constructor for S3 class r2ogs6_time_loop
}
