% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_test_definition.R
\name{new_r2ogs6_vtkdiff}
\alias{new_r2ogs6_vtkdiff}
\title{new_r2ogs6_vtkdiff}
\usage{
new_r2ogs6_vtkdiff(regex, field, absolute_tolerance, relative_tolerance)
}
\description{
Constructor for S3 class r2ogs6_vtkdiff
}
