% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_vtu.R
\name{new_vtu}
\alias{new_vtu}
\title{new_vtu}
\usage{
new_vtu(type, version, byte_order, header_type, compressor = NULL)
}
\arguments{
\item{type}{The type of the vtu file (e. g. "UnstructuredGrid")}

\item{version}{The version of the .vtu file (e. g. 1.0)}

\item{byte_order}{Either "LittleEndian" or "BigEndian"}

\item{header_type}{The header type (e. g. "UInt32")}

\item{compressor}{Optional: The compressor to be used}
}
\description{
S3 class constructor, returns S3 class describing the .vtu file and any parameters defined in it
}
