% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{node_to_object}
\alias{node_to_object}
\title{node_to_object}
\usage{
node_to_object(xml_node, xpath = "")
}
\arguments{
\item{xml_node}{xml2::xml_node: XML node}

\item{xpath}{string: XPath expression (for subclass differentiation)}
}
\description{
Returns representation of an XML node. This is a recursive
function.
ASSUMPTIONS:
1) Leaf nodes will never be prj_* objects
2) Wrapper nodes are represented as lists
3) Parent nodes whose children have no children are represented as lists
}
