% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{node_to_prj_class_object}
\alias{node_to_prj_class_object}
\title{node_to_prj_class_object}
\usage{
node_to_prj_class_object(xml_node, xpath)
}
\arguments{
\item{xml_node}{xml2::xml_node: XML node}

\item{xpath}{string: XPath expression (for subclass differentiation)}
}
\description{
Takes an XML node and turns it into a class object
}
