% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{node_to_r2ogs6_class_object}
\alias{node_to_r2ogs6_class_object}
\title{node_to_r2ogs6_class_object}
\usage{
node_to_r2ogs6_class_object(
  xml_node,
  xpath_expr,
  subclasses_names = character()
)
}
\arguments{
\item{xml_node}{xml2::xml_node: XML node}

\item{xpath_expr}{string: XPath expression (for subclass differentiation)}

\item{subclasses_names}{character: Optional: Names of r2ogs6 subclasses
(r2ogs6 classes without a method for input_add)}
}
\description{
Takes an XML node and turns it into a class object
}
