% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{node_to_r2ogs6_obj}
\alias{node_to_r2ogs6_obj}
\title{node_to_r2ogs6_obj}
\usage{
node_to_r2ogs6_obj(xml_node, subclasses_names = NULL)
}
\arguments{
\item{xml_node}{An XML node (of class xml2::xml_node)}

\item{subclasses_names}{Optional: A character vector containing the names of r2ogs6_*
subclasses (r2ogs6_* classes without a method for input_add)}
}
\description{
Takes an XML node and turns it into a class object
}
