% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{OGS6}
\alias{OGS6}
\title{Constructor for the OGS6 base class}
\description{
Constructor for the OGS6 base class

Constructor for the OGS6 base class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6$new()}}
\item \href{#method-add_sim_input}{\code{OGS6$add_sim_input()}}
\item \href{#method-set_sim_input_obj_param}{\code{OGS6$set_sim_input_obj_param()}}
\item \href{#method-add_sim_output}{\code{OGS6$add_sim_output()}}
\item \href{#method-clone}{\code{OGS6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$new(sim_name, sim_id, sim_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_sim_input"></a>}}
\if{latex}{\out{\hypertarget{method-add_sim_input}{}}}
\subsection{Method \code{add_sim_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_sim_input(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_sim_input_obj_param"></a>}}
\if{latex}{\out{\hypertarget{method-set_sim_input_obj_param}{}}}
\subsection{Method \code{set_sim_input_obj_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$set_sim_input_obj_param(obj_name, obj_param_name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_sim_output"></a>}}
\if{latex}{\out{\hypertarget{method-add_sim_output}{}}}
\subsection{Method \code{add_sim_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_sim_output(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
