% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{ogs6_call_ogs6}
\alias{ogs6_call_ogs6}
\title{ogs6_call_ogs6}
\usage{
ogs6_call_ogs6(ogs6_obj, write_logfile = TRUE, ogs6_bin_path, verbose = F)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{write_logfile}{flag: Should output be written to a logfile? If
\code{FALSE}, output will be written to console. If \code{TRUE}, logfile
directory will be created in \code{ogs6$sim_path} directory}

\item{ogs6_bin_path}{string: Optional: OpenGeoSys 6 bin folder path. Defaults
to \code{options("r2ogs6.default_ogs6_bin_path")}}

\item{verbose}{flag}
}
\description{
Makes system call to OpenGeoSys 6 and retrieves exit code.
}
