% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_chain.R
\name{OGS6_Chain}
\alias{OGS6_Chain}
\title{OGS6_Chain}
\description{
Constructor for the OGS6_Chain base class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_Chain$new()}}
\item \href{#method-run_simulation}{\code{OGS6_Chain$run_simulation()}}
\item \href{#method-print_log}{\code{OGS6_Chain$print_log()}}
\item \href{#method-clone}{\code{OGS6_Chain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_Chain object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Chain$new(ogs6_obj, iter_n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ogs6_obj}}{OGS6: A simulation object.}

\item{\code{iter_n}}{number: Number of iterations (length of the chain)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-run_simulation}{}}}
\subsection{Method \code{run_simulation()}}{
Runs the simulation. This calls r2ogs6::run_simulation() internally.
For chains, output will be written to logfiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Chain$run_simulation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_log"></a>}}
\if{latex}{\out{\hypertarget{method-print_log}{}}}
\subsection{Method \code{print_log()}}{
Reads in logfile of simulation and prints it to the console.
This calls print_log internally. If no index is specified, this will
print the logfile of the last object in the chain.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Chain$print_log(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{number: chain index of a OGS6 object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Chain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
