% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{ogs6_generate_benchmark_script}
\alias{ogs6_generate_benchmark_script}
\title{Generate benchmark script}
\usage{
ogs6_generate_benchmark_script(
  prj_path,
  sim_path,
  ogs6_bin_path,
  script_path,
  read_in_gml,
  read_in_vtu = FALSE
)
}
\arguments{
\item{prj_path}{string: `.prj` file the script will be based on}

\item{sim_path}{string: Path where all simulation files will be saved}

\item{ogs6_bin_path}{string: Path to OpenGeoSys executable. Defaults to
options("r2ogs6.default_ogs6_bin_path").}

\item{script_path}{string: Path where benchmark script will be saved}

\item{read_in_gml}{flag: Optional: Should `.gml` file just be copied or
read in too?}

\item{read_in_vtu}{flag: Optional: Should `.vtu` file(s) just be copied
or read in too?}
}
\description{
Generates an R script from an existing `.prj` benchmark file.
}
