% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{ogs6_generate_benchmark_scripts}
\alias{ogs6_generate_benchmark_scripts}
\title{Generate benchmark scripts}
\usage{
ogs6_generate_benchmark_scripts(
  path,
  sim_path,
  script_path,
  read_in_gml,
  read_in_vtu = FALSE,
  starting_from_prj_path = "",
  skip_prj_paths = character()
)
}
\arguments{
\item{path}{string: Path to a benchmark directory to generate scripts from}

\item{sim_path}{string: Path where all simulation files will be saved}

\item{script_path}{string: Path where benchmark script will be saved}

\item{read_in_gml}{flag: Optional: Should `.gml` file just be copied or
read in too?}

\item{read_in_vtu}{flag: Optional: Should `.vtu` file(s) just be copied
or read in too?}

\item{starting_from_prj_path}{string: Optional: `.prj` path to start from}

\item{skip_prj_paths}{character: Optional: `.prj` paths to skip}
}
\description{
This is a wrapper function for `ogs6_generate_benchmark_script()`.
}
