% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_vtu.R
\name{ogs6_generate_structured_mesh}
\alias{ogs6_generate_structured_mesh}
\title{ogs6_generate_structured_mesh}
\usage{
ogs6_generate_structured_mesh(args_str, ogs6_utils_path)
}
\arguments{
\item{args_str}{string: The arguments the script will be called with}

\item{ogs6_utils_path}{string: Optional: Path to OpenGeoSys6 utils folder.
Defaults to options("r2ogs6.default_ogs6_utils_path")}
}
\value{
string: .vtu file path
}
\description{
Wrapper function to call generateStructuredMesh.exe
  (VTK mesh generator). For full documentation see
  https://www.opengeosys.org/docs/tools/meshing/structured-mesh-generation/
}
