% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_ensemble.R
\name{ogs6_get_combinations}
\alias{ogs6_get_combinations}
\title{ogs6_get_combinations}
\usage{
ogs6_get_combinations(...)
}
\arguments{
\item{...}{vector:}
}
\description{
Gets possible combinations from supplied vectors
}
