% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{ogs6_get_medium_property}
\alias{ogs6_get_medium_property}
\title{ogs6_get_medium_property}
\usage{
ogs6_get_medium_property(medium, name)
}
\arguments{
\item{medium}{prj_medium}

\item{name}{string: The property name}
}
\value{
prj_pr_property
}
\description{
Returns a medium property based on the property name
}
