% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_gml.R
\name{OGS6_gml}
\alias{OGS6_gml}
\title{OGS6_gml}
\description{
Constructor for the OGS6_gml base class
}
\examples{
OGS6_gml$new(
    name = "cube_1x1x1_geometry",
    points = tibble::tibble(
        x = c(0, 0, 0, 0),
        y = c(0, 0, 1, 1),
        z = c(0, 1, 1, 0),
        name = c("origin", "", "", "")
    ),
    polylines = list(polyline = list("front_left",
                                     c(
                                         pnt = 0, pnt = 1
                                     ))),
    surfaces = list(surface = list(
        name = "left",
        element = c(p1 = 0, p2 = 1, p3 = 2),
        element = c(p1 = 0, p2 = 3, p3 = 2)
    ))
)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{gml_path}}{Getter for private parameter '.gml_path'}

\item{\code{name}}{Access to private parameter '.name'}

\item{\code{points}}{Access to private parameter '.points'}

\item{\code{polylines}}{Access to private parameter '.polylines'}

\item{\code{surfaces}}{Access to private parameter '.surfaces'}

\item{\code{is_subclass}}{Getter for private parameter '.is_subclass'}

\item{\code{attr_names}}{Getter for private parameter '.attr_names'}

\item{\code{flatten_on_exp}}{Getter for private parameter '.flatten_on_exp'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_gml$new()}}
\item \href{#method-print}{\code{OGS6_gml$print()}}
\item \href{#method-clone}{\code{OGS6_gml$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_gml object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_gml$new(
  gml_path = NULL,
  name = NULL,
  points = NULL,
  polylines = NULL,
  surfaces = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gml_path}}{string: Optional: Path to .gml file}

\item{\code{name}}{string: Geometry name}

\item{\code{points}}{tibble: Must have 3 vectors named 'x', 'y' and 'z', may
have optional 'name' vector}

\item{\code{polylines}}{list(list("foo", c(1, 2))):}

\item{\code{surfaces}}{list(list("foo", c(1, 2, 3), c(2, 3, 4)))}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Overrides default printing behaviour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_gml$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_gml$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
