% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{OGS6_mesh}
\alias{OGS6_mesh}
\title{OGS6_mesh}
\description{
Constructor for the OGS6_mesh base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mesh_path}}{Access to private parameter '.mesh_path'}

\item{\code{mesh_filename}}{Access to private parameter '.mesh_filename'}

\item{\code{is_subclass}}{Access to private parameter '.is_subclass'}

\item{\code{subclasses_names}}{Access to private parameter '.subclasses_names'}

\item{\code{attr_names}}{Access to private parameter '.attr_names'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_mesh$new()}}
\item \href{#method-clone}{\code{OGS6_mesh$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_mesh object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_mesh$new(mesh_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mesh_path}}{string:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_mesh$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
