% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_msh.R
\name{OGS6_msh}
\alias{OGS6_msh}
\title{OGS6_msh}
\description{
Small class to support \code{MSH} mesh files. Note that
MSH files are a legacy format from OGS5. It is recommended to switch to VTU
 files.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{msh_path}}{Getter/setter for private parameter `.msh_path`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OGS6_msh-new}{\code{OGS6_msh$new()}}
\item \href{#method-OGS6_msh-print}{\code{OGS6_msh$print()}}
\item \href{#method-OGS6_msh-clone}{\code{OGS6_msh$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OGS6_msh-new"></a>}}
\if{latex}{\out{\hypertarget{method-OGS6_msh-new}{}}}
\subsection{Method \code{new()}}{
reates new OGS6_vtu object..
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_msh$new(msh_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msh_path}}{path to \code{*.msh} file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OGS6_msh-print"></a>}}
\if{latex}{\out{\hypertarget{method-OGS6_msh-print}{}}}
\subsection{Method \code{print()}}{
Overrides the default print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_msh$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OGS6_msh-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OGS6_msh-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_msh$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
