% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{OGS6_Piece}
\alias{OGS6_Piece}
\title{OGS6_Piece}
\description{
Constructor for the OGS6_Piece base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{NumberOfPoints}}{Access to private parameter '.NumberOfPoints'}

\item{\code{NumberOfCells}}{Access to private parameter '.NumberOfCells'}

\item{\code{PointData}}{Access to private parameter '.PointData'}

\item{\code{Points}}{Access to private parameter '.Points'}

\item{\code{Cells}}{Access to private parameter '.Cells'}

\item{\code{CellData}}{Access to private parameter '.CellData'}

\item{\code{is_subclass}}{Access to private parameter '.is_subclass'}

\item{\code{subclasses_names}}{Access to private parameter '.subclasses_names'}

\item{\code{attr_names}}{Access to private parameter '.attr_names'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_Piece$new()}}
\item \href{#method-clone}{\code{OGS6_Piece$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_Piece object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Piece$new(
  NumberOfPoints,
  NumberOfCells,
  PointData,
  Points,
  Cells,
  CellData = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NumberOfPoints}}{string | number:}

\item{\code{NumberOfCells}}{string | number:}

\item{\code{PointData}}{list, :}

\item{\code{Points}}{list, :}

\item{\code{Cells}}{list, :}

\item{\code{CellData}}{list, : Optional:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Piece$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
