% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{ogs6_prj_top_level_classes}
\alias{ogs6_prj_top_level_classes}
\title{ogs6_prj_top_level_classes}
\usage{
ogs6_prj_top_level_classes()
}
\value{
character
}
\description{
Returns named character vector of \code{OGS6} top level
\code{.prj} tags (names) represented by \code{r2ogs6} classes along with
their class names (values).
}
