% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{OGS6_pvd}
\alias{OGS6_pvd}
\title{OGS6_pvd}
\description{
Constructor for the OGS6_pvd base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pvd_path}}{Getter for private parameter '.pvd_path'}

\item{\code{datasets}}{Getter for private parameter '.datasets'}

\item{\code{vtu_paths}}{Getter for `datasets` `file`}

\item{\code{abs_vtu_paths}}{Gets absolute .vtu paths, e.g. `dirname(pvd_path)` + `datasets` `file`}

\item{\code{timesteps}}{Gets timesteps from private parameter '.datasets'}

\item{\code{OGS6_vtus}}{Getter for private parameter '.OGS6_vtus'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_pvd$new()}}
\item \href{#method-get_vtu_path_by_timestep}{\code{OGS6_pvd$get_vtu_path_by_timestep()}}
\item \href{#method-get_timestep_by_vtu_path}{\code{OGS6_pvd$get_timestep_by_vtu_path()}}
\item \href{#method-get_PointData_time_tibble}{\code{OGS6_pvd$get_PointData_time_tibble()}}
\item \href{#method-get_PointData_at_timestep}{\code{OGS6_pvd$get_PointData_at_timestep()}}
\item \href{#method-clone}{\code{OGS6_pvd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_pvd object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$new(pvd_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pvd_path}}{string: Path to .pvd file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vtu_path_by_timestep"></a>}}
\if{latex}{\out{\hypertarget{method-get_vtu_path_by_timestep}{}}}
\subsection{Method \code{get_vtu_path_by_timestep()}}{
Returns .vtu path for specified timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_vtu_path_by_timestep(timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timestep}}{string: Timestep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_timestep_by_vtu_path"></a>}}
\if{latex}{\out{\hypertarget{method-get_timestep_by_vtu_path}{}}}
\subsection{Method \code{get_timestep_by_vtu_path()}}{
Returns timestep for specified .vtu path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_timestep_by_vtu_path(vtu_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vtu_path}}{string: .vtu path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_PointData_time_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-get_PointData_time_tibble}{}}}
\subsection{Method \code{get_PointData_time_tibble()}}{
Creates a tibble object from PointData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_PointData_time_tibble(
  point_ids,
  Names,
  start_at_timestep,
  end_at_timestep
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{point_ids}}{numeric: Optional: IDs of the points of interest.
Will default to all point IDs if not defined.}

\item{\code{Names}}{character: `Name` attributes of `DataArray` elements}

\item{\code{start_at_timestep}}{number: Timestep to start at}

\item{\code{end_at_timestep}}{number: Timestep to end at}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_PointData_at_timestep"></a>}}
\if{latex}{\out{\hypertarget{method-get_PointData_at_timestep}{}}}
\subsection{Method \code{get_PointData_at_timestep()}}{
Gets PointData at specified timestep. Calls
`get_PointData_time_tibble` internally with `start_at_timestep` and
`end_at_timestep` both being `timestep`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_PointData_at_timestep(point_ids, Names, timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{point_ids}}{number: Point IDs}

\item{\code{Names}}{character: `Name` attributes of `DataArray` elements}

\item{\code{timestep}}{string: Timestep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
