% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{ogs6_run_simulation}
\alias{ogs6_run_simulation}
\title{ogs6_run_simulation}
\usage{
ogs6_run_simulation(
  ogs6_obj,
  write_logfile = TRUE,
  ogs6_bin_path,
  overwrite = T,
  verbose = F
)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{write_logfile}{flag: Should output be written to a logfile? If
\code{FALSE}, output will be written to console. If \code{TRUE}, logfile
directory will be created in \code{ogs6$sim_path} directory}

\item{ogs6_bin_path}{string: Optional: OpenGeoSys 6 executable path. Defaults
to \code{options("r2ogs6.default_ogs6_bin_path")}}

\item{overwrite}{flag: Should existing files be overwritten?}

\item{verbose}{flag}
}
\description{
Wrapper function that calls \code{ogs6_export_sim_files()},
  \code{ogs6_call_ogs6()} and \code{ogs6_read_output_files()}.
}
