% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{OGS6_UnstructuredGrid}
\alias{OGS6_UnstructuredGrid}
\title{OGS6_UnstructuredGrid}
\description{
Constructor for the OGS6_UnstructuredGrid base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Piece}}{Access to private parameter '.Piece'}

\item{\code{FieldData}}{Access to private parameter '.FieldData'}

\item{\code{is_subclass}}{Access to private parameter '.is_subclass'}

\item{\code{subclasses_names}}{Access to private parameter '.subclasses_names'}

\item{\code{attr_names}}{Access to private parameter '.attr_names'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_UnstructuredGrid$new()}}
\item \href{#method-clone}{\code{OGS6_UnstructuredGrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_UnstructuredGrid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_UnstructuredGrid$new(Piece, FieldData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Piece}}{OGS6_Piece:}

\item{\code{FieldData}}{character, length == 2:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_UnstructuredGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
