% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_vtu.R
\name{OGS6_vtu}
\alias{OGS6_vtu}
\title{OGS6_vtu}
\description{
Constructor for the `OGS6_vtu` base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{vtu_path}}{Getter for private parameter `.vtu_path`}

\item{\code{number_of_points}}{Getter for NumberOfPoints parameter of `.vtkUnstructuredGrid`}

\item{\code{number_of_cells}}{Getter for NumberOfCells parameter of `.vtkUnstructuredGrid`}

\item{\code{points}}{Getter for Points parameter of `.dsa_vtkUnstructuredGrid`}

\item{\code{cells}}{Getter for Cells parameter of `.dsa_vtkUnstructuredGrid`}

\item{\code{field_data}}{Getter for FieldData parameter of `.dsa_vtkUnstructuredGrid`}

\item{\code{point_data}}{Getter for PointData parameter of `.dsa_vtkUnstructuredGrid`}

\item{\code{cell_data}}{Getter for CellData parameter of `.dsa_vtkUnstructuredGrid`}

\item{\code{vtkPointLocator}}{Getter for private parameter `.vtkPointLocator`}

\item{\code{vtkUnstructuredGrid}}{Access to private parameter `.vtkUnstructuredGrid`}

\item{\code{dsa_vtkUnstructuredGrid}}{Getter for private parameter `.dsa_vtkUnstructuredGrid`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_vtu$new()}}
\item \href{#method-print}{\code{OGS6_vtu$print()}}
\item \href{#method-get_field_data}{\code{OGS6_vtu$get_field_data()}}
\item \href{#method-get_point_coords}{\code{OGS6_vtu$get_point_coords()}}
\item \href{#method-get_point_data_at}{\code{OGS6_vtu$get_point_data_at()}}
\item \href{#method-get_point_data}{\code{OGS6_vtu$get_point_data()}}
\item \href{#method-get_cell_data}{\code{OGS6_vtu$get_cell_data()}}
\item \href{#method-clone}{\code{OGS6_vtu$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_vtu object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$new(vtu_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vtu_path}}{string: Path to .vtu file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Overrides default printing behaviour.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_field_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_field_data}{}}}
\subsection{Method \code{get_field_data()}}{
Gets FieldData.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_field_data(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements, defaults to all in `FieldData`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of format list(value_a = 1, value_b = 2), where
the names reference the `Name` attributes of the `DataArray`
elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_point_coords"></a>}}
\if{latex}{\out{\hypertarget{method-get_point_coords}{}}}
\subsection{Method \code{get_point_coords()}}{
Gets coordinates of specific points by their IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_point_coords(point_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{point_ids}}{numeric: Point IDs}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If `point_ids` is a number, a coordinate array. If
point_ids` is a numeric vector with length > 1, a list of coordinate
arrays.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_point_data_at"></a>}}
\if{latex}{\out{\hypertarget{method-get_point_data_at}{}}}
\subsection{Method \code{get_point_data_at()}}{
Gets PointData at specified coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_point_data_at(coordinates, keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coordinates}}{list(numeric): List of coordinates (a coordinate
is a numeric vector of length 3)}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements, defaults to all in `PointData`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_point_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_point_data}{}}}
\subsection{Method \code{get_point_data()}}{
Gets PointData for points with IDs in `point_ids`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_point_data(point_ids, keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{point_ids}}{numeric: Optional: Point IDs, defaults to all}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements, defaults to all in `PointData`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble::tibble where each row represents a point.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cell_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_cell_data}{}}}
\subsection{Method \code{get_cell_data()}}{
Gets CellData for cells with IDs in `cell_ids`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_cell_data(cell_ids, keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell_ids}}{numeric: Optional: Cell IDs, defaults to all}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements, defaults to all in `CellData`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble::tibble where each row represents a cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
