% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{ogs_export_sim_files}
\alias{ogs_export_sim_files}
\title{ogs_export_sim_files}
\usage{
ogs_export_sim_files(ogs6_obj, test_mode = FALSE)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{test_mode}{flag: If \code{TRUE}, Will not check status of
\code{ogs6_obj} before exporting files. Defaults to \code{FALSE}}
}
\description{
Creates \code{ogs6$sim_path} directory if it does not exist yet
and exports and / or copies all simulation files to it.
}
