% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{ogs_get_medium_property}
\alias{ogs_get_medium_property}
\title{ogs_get_medium_property}
\usage{
ogs_get_medium_property(medium, name)
}
\arguments{
\item{medium}{r2ogs6_medium}

\item{name}{string: The property name}
}
\value{
r2ogs6_pr_property
}
\description{
Returns a medium property based on the property name
}
