% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{ogs_read_output_files}
\alias{ogs_read_output_files}
\title{ogs_read_output_files}
\usage{
ogs_read_output_files(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}
}
\description{
Read in generated \code{.pvd} files and add it to ogs6_obj
}
