% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{ogs_run_simulation}
\alias{ogs_run_simulation}
\title{ogs_run_simulation}
\usage{
ogs_run_simulation(ogs6_obj, write_logfile = TRUE, ogs_bin_path, verbose = F)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{write_logfile}{flag: Should output be written to a logfile? If `FALSE`,
output will be written to console. If `TRUE`, logfile directory will be
created in `ogs6$sim_path` directory}

\item{ogs_bin_path}{string: Optional: OpenGeoSys 6 bin folder path. Defaults
to `options("r2ogs6.default_ogs_bin_path")`}

\item{verbose}{flag}
}
\description{
Wrapper function that calls `ogs_export_sim_files`,
`ogs_call_ogs6` and `ogs_read_output_files`.
}
