% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{order_parameters}
\alias{order_parameters}
\title{order_parameters}
\usage{
order_parameters(parameters, class_name)
}
\arguments{
\item{parameters}{list: Parameters}

\item{class_name}{string: The name of a class}
}
\description{
Orders a list of parameters corresponding to the argument order
of a class
}
