% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_from_file.R
\name{pick_file}
\alias{pick_file}
\title{pick_file}
\usage{
pick_file(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}
}
\description{
Lets the user pick a file and adds it to the specified OGS6 class object
}
