% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gml.R
\name{points_to_xml}
\alias{points_to_xml}
\title{points_to_xml}
\usage{
points_to_xml(points)
}
\arguments{
\item{points}{tibble: The specified points}
}
\value{
An XML node containing the points
}
\description{
Turns a tibble of points into an XML node
}
