% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gml.R
\name{polylines_to_xml}
\alias{polylines_to_xml}
\title{polylines_to_xml}
\usage{
polylines_to_xml(polylines)
}
\arguments{
\item{polylines}{A list of polylines}
}
\value{
An XML node containing the polylines
}
\description{
Turns a list of polylines into an XML node
}
