% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_local_coordinate_system.R
\name{prj_basis_vector_1}
\alias{prj_basis_vector_1}
\title{prj_basis_vector_1}
\usage{
prj_basis_vector_1(name = NULL, implicit = NULL)
}
\arguments{
\item{name}{Optional:}

\item{implicit}{Optional:}
}
\description{
tag: basis_vector_1
}
