% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{prj_chemical_reaction}
\alias{prj_chemical_reaction}
\title{prj_chemical_reaction}
\usage{
prj_chemical_reaction(
  stoichiometric_coefficients,
  reaction_type,
  first_order_rate_constant
)
}
\arguments{
\item{stoichiometric_coefficients}{numeric:}

\item{reaction_type}{string:}

\item{first_order_rate_constant}{double:}
}
\description{
tag: chemical_reaction
}
