% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{prj_chemical_system}
\alias{prj_chemical_system}
\title{prj_chemical_system}
\usage{
prj_chemical_system(
  chemical_solver,
  database,
  solution,
  mesh = NULL,
  knobs = NULL,
  kinetic_reactants = NULL,
  rates = NULL,
  equilibrium_reactants = NULL,
  surface = NULL,
  user_punch = NULL,
  linear_solver = NULL,
  exchangers = NULL
)
}
\arguments{
\item{chemical_solver}{string:}

\item{database}{string:}

\item{solution}{prj_solution:}

\item{mesh}{Optional: string:}

\item{knobs}{Optional: list:}

\item{kinetic_reactants}{Optional: list, prj_kinetic_reactant:}

\item{rates}{Optional: list, prj_rate:}

\item{equilibrium_reactants}{Optional: list, prj_phase_component:}

\item{surface}{Optional:}

\item{user_punch}{Optional:}

\item{linear_solver}{Optional:}

\item{exchangers}{Optional: prj_exchangers}
}
\description{
tag: chemical_system
}
\examples{
prj_chemical_system(
    chemical_solver = "Phreeqc",
    database = "PSINA_12_07_110615_DAV_s.dat",
    solution = prj_solution(
        temperature = 25,
        pressure = 1,
        pe = 4,
        components = list(
            component = "C(4)",
            component = "Ca"
        ),
        charge_balance = "pH"
    ),
    mesh = "calcite_ReactiveDomain",
    knobs = list(
        max_iter = "100",
        relative_convergence_tolerance = "1e-12",
        tolerance = "1e-15",
        step_size = "100",
        scaling = "0"
    ),
    equilibrium_reactants = list(
        phase_component = prj_phase_component(
            name = "Calcite",
            saturation_index = 0,
            initial_amount = 0.000207
        )
    )
)
}
