% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{prj_component}
\alias{prj_component}
\title{prj_component}
\usage{
prj_component(name, properties)
}
\arguments{
\item{name}{string:}

\item{properties}{list, prj_com_property:}
}
\description{
tag: component
}
\examples{
prj_component(
    name = "concentration",
    properties = list(
        property = prj_com_property(name = "molecular_diffusion",
                                       type = "Constant",
                                       value = 2e-09)
    )
)
}
