% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{prj_convergence_criterion}
\alias{prj_convergence_criterion}
\title{prj_convergence_criterion}
\usage{
prj_convergence_criterion(
  type,
  norm_type,
  abstol = NULL,
  reltol = NULL,
  abstols = NULL,
  reltols = NULL
)
}
\arguments{
\item{type}{string: Type}

\item{norm_type}{string: ...}

\item{abstol}{string | double: Absolute tolerance}

\item{reltol}{string | double: Relative tolerance}

\item{abstols}{string | numeric: Absolute tolerances}

\item{reltols}{string | numeric: Relative tolerances}
}
\description{
tag: convergence_criterion
}
\examples{
prj_convergence_criterion(
    type = "DeltaX",
    norm_type = "INFINITY_N",
    abstol = 1e-08,
    reltol = 1e-10
)
}
