% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_curve.R
\name{prj_curve}
\alias{prj_curve}
\title{prj_curve}
\usage{
prj_curve(name, coords, values)
}
\arguments{
\item{name}{string: Name of the curve}

\item{coords}{string | numeric: Coordinates at which the curve's values
are given}

\item{values}{string | numeric: Values of the curve at the given coordinates}
}
\description{
tag: curve, a curve
}
\examples{
prj_curve(name = "Dirichlet_temporal",
             coords = "0 3",
             values = "200 200")
}
