% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_prj.R
\name{prj_data_to_xml}
\alias{prj_data_to_xml}
\title{Wrapper function to create a .prj XML document based on the user input data}
\usage{
prj_data_to_xml(prj_obj)
}
\arguments{
\item{prj_obj}{.}
}
\description{
Wrapper function to create a .prj XML document based on the user input data
}
