% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variable.R
\name{prj_deactivated_subdomain}
\alias{prj_deactivated_subdomain}
\title{prj_deactivated_subdomain}
\usage{
prj_deactivated_subdomain(
  material_ids,
  time_interval = NULL,
  time_curve = NULL,
  line_segment = NULL
)
}
\arguments{
\item{material_ids}{string | double:}

\item{time_interval}{Optional: list, numeric:}

\item{time_curve}{Optional:}

\item{line_segment}{Optional:}
}
\description{
tag: deactivated_subdomain
}
\examples{
prj_deactivated_subdomain(material_ids = 0,
                          time_interval = list(start = "0.51",
                                               end = "1.0"))
}
