% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_borehole_heat_exchanger.R
\name{prj_flow_and_temperature_control}
\alias{prj_flow_and_temperature_control}
\title{prj_flow_and_temperature_control}
\usage{
prj_flow_and_temperature_control(
  type,
  flow_rate = NULL,
  temperature_curve = NULL,
  power = NULL,
  power_curve = NULL,
  flow_rate_curve = NULL
)
}
\arguments{
\item{type}{string:}

\item{flow_rate}{Optional: string | double:}

\item{temperature_curve}{Optional: string:}

\item{power}{Optional: string | double:}

\item{power_curve}{Optional: string:}

\item{flow_rate_curve}{Optional: string:}
}
\description{
tag: flow_and_temperature_control
}
\examples{
prj_flow_and_temperature_control(
    type = "TemperatureCurveConstantFlow",
    flow_rate = 2e-04,
    temperature_curve = "inflow_temperature"
)
}
