% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_fracture_model}
\alias{prj_fracture_model}
\title{prj_fracture_model}
\usage{
prj_fracture_model(
  type,
  normal_stiffness,
  shear_stiffness,
  penalty_aperture_cutoff,
  tension_cutoff,
  fracture_toughness = NULL,
  peak_normal_traction = NULL,
  friction_angle = NULL,
  dilatancy_angle = NULL,
  cohesion = NULL,
  nonlinear_solver = NULL
)
}
\arguments{
\item{type}{string:}

\item{normal_stiffness}{string:}

\item{shear_stiffness}{string:}

\item{penalty_aperture_cutoff}{string | double:}

\item{tension_cutoff}{string | double:}

\item{fracture_toughness}{Optional: string:}

\item{peak_normal_traction}{Optional: string:}

\item{friction_angle}{Optional: string:}

\item{dilatancy_angle}{Optional: string:}

\item{cohesion}{Optional: string:}

\item{nonlinear_solver}{Optional: list:}
}
\description{
tag: fracture_model
}
\examples{
prj_fracture_model(type = "LinearElasticIsotropic",
                      normal_stiffness = "Kn",
                      shear_stiffness = "Ks",
                      penalty_aperture_cutoff = 1e-05,
                      tension_cutoff = 1)
}
