% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{prj_geometrical_set}
\alias{prj_geometrical_set}
\title{prj_geometrical_set}
\usage{
prj_geometrical_set(name, geometry)
}
\arguments{
\item{name}{string}

\item{geometry}{string}
}
\description{
tag: geometrical_set
}
