% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_ice_constitutive_relation}
\alias{prj_ice_constitutive_relation}
\title{prj_ice_constitutive_relation}
\usage{
prj_ice_constitutive_relation(
  type,
  youngs_modulus = NULL,
  poissons_ratio = NULL
)
}
\arguments{
\item{type}{string:}

\item{youngs_modulus}{Optional:}

\item{poissons_ratio}{Optional:}
}
\description{
tag: ice_constitutive_relation
}
