% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_insitu.R
\name{prj_insitu}
\alias{prj_insitu}
\title{prj_insitu}
\usage{
prj_insitu(scripts)
}
\arguments{
\item{scripts}{character: Script names}
}
\description{
tag: insitu
}
\examples{
prj_insitu(scripts = c(script = "square_1e1_neumann-insitu.py"))
}
