% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_jacobian_assembler}
\alias{prj_jacobian_assembler}
\title{prj_jacobian_assembler}
\usage{
prj_jacobian_assembler(
  type,
  component_magnitudes = NULL,
  relative_epsilons = NULL
)
}
\arguments{
\item{type}{string:}

\item{component_magnitudes}{Optional: string | double:}

\item{relative_epsilons}{Optional: string | double:}
}
\description{
tag: jacobian_assembler
}
\examples{
prj_jacobian_assembler(
    type = "CentralDifferences",
    component_magnitudes = 1,
    relative_epsilons = 1e-08
)
}
