% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{prj_kinetic_reactant}
\alias{prj_kinetic_reactant}
\title{prj_kinetic_reactant}
\usage{
prj_kinetic_reactant(
  name,
  initial_amount = NULL,
  chemical_formula = NULL,
  fix_amount = NULL
)
}
\arguments{
\item{name}{The component name}

\item{initial_amount}{The initial amount of the component}

\item{chemical_formula}{The chemical formula of the component}

\item{fix_amount}{Should the amount be fixed or not?}
}
\description{
S3 class describing .prj kinetic_reactant
}
\examples{
prj_kinetic_reactant(name = "Productc",
                        initial_amount = 1e-06)
}
