% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_micro_porosity}
\alias{prj_micro_porosity}
\title{prj_micro_porosity}
\usage{
prj_micro_porosity(mass_exchange_coefficient, nonlinear_solver)
}
\arguments{
\item{mass_exchange_coefficient}{Required:}

\item{nonlinear_solver}{Required:}
}
\description{
tag: micro_porosity
}
