% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_nonlinear_solver.R
\name{prj_nonlinear_solver}
\alias{prj_nonlinear_solver}
\title{prj_nonlinear_solver}
\usage{
prj_nonlinear_solver(name, type, max_iter, linear_solver, damping = NULL)
}
\arguments{
\item{name}{string: Name}

\item{type}{string: Type}

\item{max_iter}{string | double: Maximal number of iterations}

\item{linear_solver}{string: Name of corresponding linear_solver}

\item{damping}{Optional: string | double: Damping}
}
\description{
tag: nonlinear_solver
}
\examples{
prj_nonlinear_solver(
    name = "basic_newton",
    type = "Newton",
    max_iter = 50,
    linear_solver = "general_linear_solver"
)
}
