% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{prj_output_mesh}
\alias{prj_output_mesh}
\title{prj_output_mesh}
\usage{
prj_output_mesh(name, material_ids = NULL)
}
\arguments{
\item{name}{string : Name of the mesh}

\item{material_ids}{Optional: string}
}
\description{
tag: mesh
}
