% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_parameter.R
\name{prj_parameter}
\alias{prj_parameter}
\title{prj_parameter}
\usage{
prj_parameter(
  name,
  type,
  value = NULL,
  values = NULL,
  curve = NULL,
  parameter = NULL,
  group_id_property = NULL,
  field_name = NULL,
  mesh = NULL,
  time_series = NULL,
  use_local_coordinate_system = NULL,
  ...
)
}
\arguments{
\item{name}{string:}

\item{type}{string:}

\item{value}{Optional: string | double: Parameter value}

\item{values}{Optional: string | double: Parameter values}

\item{curve}{Optional: string:}

\item{parameter}{Optional: string:}

\item{group_id_property}{Optional: string:}

\item{field_name}{Optional: string:}

\item{mesh}{Optional: string:}

\item{time_series}{Optional: list:}

\item{use_local_coordinate_system}{Optional: string, "true" | "false":}

\item{...}{Optional: for index_values and expression tags (since there can be
multiple)}
}
\description{
tag: parameter
}
\examples{
prj_parameter(name = "zero",
                 type = "Constant",
                 value = 0)

prj_parameter(name = "flux_in",
                 type = "Constant",
                 value = 1e-04)
}
