% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{prj_ph_property}
\alias{prj_ph_property}
\title{prj_ph_property}
\usage{
prj_ph_property(
  name,
  type,
  value = NULL,
  reference_value = NULL,
  offset = NULL,
  exponent = NULL,
  parameter_name = NULL,
  swelling_pressures = NULL,
  exponents = NULL,
  lower_saturation_limit = NULL,
  upper_saturation_limit = NULL,
  tortuosity = NULL,
  curve = NULL,
  ...
)
}
\arguments{
\item{name}{string:}

\item{type}{string:}

\item{value}{Optional:}

\item{reference_value}{Optional:}

\item{offset}{Optional:}

\item{exponent}{Optional:}

\item{parameter_name}{Optional:}

\item{swelling_pressures}{Optional:}

\item{exponents}{Optional:}

\item{lower_saturation_limit}{Optional:}

\item{upper_saturation_limit}{Optional:}

\item{tortuosity}{Optional:}

\item{curve}{Optional:}

\item{...}{independent_variable, dvalue}
}
\description{
tag: property
}
\examples{
prj_ph_property(name = "porosity",
                type = "Constant",
                value = 0.3)
}
