% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{prj_phase}
\alias{prj_phase}
\title{prj_phase}
\usage{
prj_phase(type, properties = NULL, components = NULL)
}
\arguments{
\item{type}{string: Phase type
(get valid types with get_valid_phase_types())}

\item{properties}{list, prj_pr_property: Properties}

\item{components}{list, components}
}
\description{
tag: phase, a coherent material with homogeneous properties
}
\examples{
prj_phase(
    type = "Solid",
    properties = list(
        property = prj_ph_property(name = "porosity",
                                      type = "Constant",
                                      value = 0.3)
    )
)
}
