% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{prj_phase_component}
\alias{prj_phase_component}
\title{prj_phase_component}
\usage{
prj_phase_component(name, saturation_index, initial_amount = NULL)
}
\arguments{
\item{name}{The component name}

\item{saturation_index}{The saturation index of the component}

\item{initial_amount}{optional: The initial amount of the component}
}
\description{
S3 class describing .prj phase_component
}
\examples{
prj_phase_component(
    name = "Calcite",
    saturation_index = 0,
    initial_amount = 0.000207
)
}
