% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_phasefield_parameters}
\alias{prj_phasefield_parameters}
\title{prj_phasefield_parameters}
\usage{
prj_phasefield_parameters(
  residual_stiffness,
  crack_resistance,
  crack_length_scale,
  kinetic_coefficient = NULL,
  history_field = NULL
)
}
\arguments{
\item{residual_stiffness}{string:}

\item{crack_resistance}{string:}

\item{crack_length_scale}{string:}

\item{kinetic_coefficient}{string:}

\item{history_field}{Optional: string:}
}
\description{
tag: phasefield_parameters
}
\examples{
prj_phasefield_parameters(
    residual_stiffness = "k",
    crack_resistance = "gc",
    crack_length_scale = "ls",
    kinetic_coefficient = "M",
    history_field = "H"
)
}
