% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_borehole_heat_exchanger.R
\name{prj_pipes}
\alias{prj_pipes}
\title{prj_pipes}
\usage{
prj_pipes(
  longitudinal_dispersion_length,
  inlet = NULL,
  outlet = NULL,
  distance_between_pipes = NULL,
  outer = NULL,
  inner = NULL
)
}
\arguments{
\item{longitudinal_dispersion_length}{string | double:}

\item{inlet}{Optional: list:}

\item{outlet}{Optional: list:}

\item{distance_between_pipes}{Optional: string | double:}

\item{outer}{Optional: list:}

\item{inner}{Optional: list:}
}
\description{
tag: pipes
}
\examples{
prj_pipes(
    longitudinal_dispersion_length = 0.001,
    inlet = list(
        diameter = " 0.013665",
        wall_thickness = "0.003035",
        wall_thermal_conductivity = "0.39"
    ),
    outlet = list(
        diameter = "0.013665",
        wall_thickness = "0.003035",
        wall_thermal_conductivity = "0.39"
    ),
    distance_between_pipes = 0.053
)
}
