% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_process}
\alias{prj_process}
\title{prj_process}
\usage{
prj_process(
  name,
  type,
  integration_order,
  process_variables,
  secondary_variables = NULL,
  specific_body_force = NULL,
  solid_density = NULL,
  dimension = NULL,
  coupling_scheme = NULL,
  reference_temperature = NULL,
  fracture_model = NULL,
  jacobian_assembler = NULL,
  internal_length = NULL,
  mass_lumping = NULL,
  porosity = NULL,
  calculatesurfaceflux = NULL,
  intrinsic_permeability = NULL,
  specific_storage = NULL,
  fluid_viscosity = NULL,
  biot_coefficient = NULL,
  fluid_density = NULL,
  initial_effective_stress = NULL,
  initial_fracture_effective_stress = NULL,
  phasefield_parameters = NULL,
  deactivate_matrix_in_flow = NULL,
  borehole_heat_exchangers = NULL,
  temperature = NULL,
  reactive_system = NULL,
  fluid_specific_heat_source = NULL,
  fluid_specific_isobaric_heat_capacity = NULL,
  solid_hydraulic_permeability = NULL,
  solid_specific_heat_source = NULL,
  solid_heat_conductivity = NULL,
  solid_specific_isobaric_heat_capacity = NULL,
  tortuosity = NULL,
  diffusion_coefficient = NULL,
  solid_density_dry = NULL,
  solid_density_initial = NULL,
  characteristic_pressure = NULL,
  characteristic_temperature = NULL,
  characteristic_vapour_mass_fraction = NULL,
  output_element_matrices = NULL,
  material_property = NULL,
  diffusion_coeff_component_b = NULL,
  diffusion_coeff_component_a = NULL,
  initial_stress = NULL,
  thermal_parameters = NULL,
  non_advective_form = NULL,
  density_solid = NULL,
  latent_heat_evaporation = NULL,
  micro_porosity = NULL,
  explicit_hm_coupling_in_unsaturated_zone = NULL,
  simplified_elasticity = NULL,
  chemically_induced_porosity_change = NULL,
  use_server_communication = NULL,
  phasefield_model = NULL,
  irreversible_threshold = NULL,
  tabular_file = NULL,
  temperature_field = NULL,
  use_stokes_brinkman_form = NULL,
  ...
)
}
\arguments{
\item{name}{string:}

\item{type}{string:}

\item{integration_order}{string | double:}

\item{process_variables}{list, character:}

\item{secondary_variables}{Optional:}

\item{specific_body_force}{Optional:}

\item{solid_density}{Optional:}

\item{dimension}{Optional:}

\item{coupling_scheme}{Optional:}

\item{reference_temperature}{Optional:}

\item{fracture_model}{Optional:}

\item{jacobian_assembler}{Optional:}

\item{internal_length}{Optional:}

\item{mass_lumping}{Optional:}

\item{porosity}{Optional:}

\item{calculatesurfaceflux}{Optional:}

\item{intrinsic_permeability}{Optional:}

\item{specific_storage}{Optional:}

\item{fluid_viscosity}{Optional:}

\item{biot_coefficient}{Optional:}

\item{fluid_density}{Optional:}

\item{initial_effective_stress}{Optional:}

\item{initial_fracture_effective_stress}{Optional:}

\item{phasefield_parameters}{Optional:}

\item{deactivate_matrix_in_flow}{Optional:}

\item{borehole_heat_exchangers}{Optional:}

\item{temperature}{Optional:}

\item{reactive_system}{Optional:}

\item{fluid_specific_heat_source}{Optional:}

\item{fluid_specific_isobaric_heat_capacity}{Optional:}

\item{solid_hydraulic_permeability}{Optional:}

\item{solid_specific_heat_source}{Optional:}

\item{solid_heat_conductivity}{Optional:}

\item{solid_specific_isobaric_heat_capacity}{Optional:}

\item{tortuosity}{Optional:}

\item{diffusion_coefficient}{Optional:}

\item{solid_density_dry}{Optional:}

\item{solid_density_initial}{Optional:}

\item{characteristic_pressure}{Optional:}

\item{characteristic_temperature}{Optional:}

\item{characteristic_vapour_mass_fraction}{Optional:}

\item{output_element_matrices}{Optional:}

\item{material_property}{Optional:}

\item{diffusion_coeff_component_b}{Optional:}

\item{diffusion_coeff_component_a}{Optional:}

\item{initial_stress}{Optional:}

\item{thermal_parameters}{Optional:}

\item{non_advective_form}{Optional:}

\item{density_solid}{Optional:}

\item{latent_heat_evaporation}{Optional:}

\item{micro_porosity}{Optional:}

\item{explicit_hm_coupling_in_unsaturated_zone}{Optional:}

\item{simplified_elasticity}{Optional: character}

\item{chemically_induced_porosity_change}{Optional: character}

\item{use_server_communication}{Optional:}

\item{phasefield_model}{Optional:}

\item{irreversible_threshold}{Optional: numeric}

\item{tabular_file}{Optional: character}

\item{temperature_field}{Optional: character}

\item{use_stokes_brinkman_form}{Optional: character}

\item{...}{Optional: fracture_properties, constitutive_relation}
}
\description{
tag: process (parent tag: processes)
}
\examples{
prj_process(
    name = "HM",
    type = "HYDRO_MECHANICS",
    integration_order = 3,
    process_variables = list(displacement = "displacement",
                             pressure = "pressure"),
    secondary_variables = list(
        secondary_variable = c(internal_name = "sigma_xx", output_name = "sigma_xx"),
        secondary_variable = c(internal_name = "sigma_yy", output_name = "sigma_yy")
    ),
    specific_body_force = c(0, 0, 0),
    dimension = 3,
    constitutive_relation = prj_constitutive_relation(
        type = "LinearElasticIsotropic",
        youngs_modulus = "E",
        poissons_ratio = "nu"
    )
)
}
