% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variable.R
\name{prj_process_variable}
\alias{prj_process_variable}
\title{prj_process_variable}
\usage{
prj_process_variable(
  name,
  components,
  order,
  initial_condition,
  boundary_conditions = NULL,
  source_terms = NULL,
  mesh = NULL,
  deactivated_subdomains = NULL
)
}
\arguments{
\item{name}{string: The name of the process variable}

\item{components}{string | double:}

\item{order}{string | double:}

\item{initial_condition}{string:}

\item{boundary_conditions}{list, prj_boundary_condition:}

\item{source_terms}{Optional: list, prj_source_term:}

\item{mesh}{Optional: string: list:}

\item{deactivated_subdomains}{Optional: list, prj_deactivated_subdomain:}
}
\description{
tag: process_variable
}
\examples{
prj_process_variable(
    name = "pressure",
    components = 1,
    order = 1,
    initial_condition = "pressure0",
    boundary_conditions = list(
        boundary_condition = prj_boundary_condition(
            type = "Neumann",
            parameter = "flux_in",
            geometrical_set = "cube_1x1x1_geometry",
            geometry = "left",
            component = 0
        )
    )
)
}
