% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_python_script.R
\name{prj_python_script}
\alias{prj_python_script}
\title{prj_python_script}
\usage{
prj_python_script(file_path)
}
\arguments{
\item{file_path}{string: Path or name of python script file (\code{*.py}).}
}
\description{
tag: python_script
}
\examples{
prj_python_script(file_path = "bcs_laplace_eq.py")

}
