% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_raster.R
\name{prj_raster}
\alias{prj_raster}
\title{prj_raster}
\usage{
prj_raster(file, variable, dimension = NULL)
}
\arguments{
\item{file}{string:}

\item{variable}{string:}

\item{dimension}{Optional:}
}
\description{
tag: raster
}
