% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_search_length_algorithm.R
\name{prj_search_length_algorithm}
\alias{prj_search_length_algorithm}
\title{prj_search_length_algorithm}
\usage{
prj_search_length_algorithm(type, value = NULL)
}
\arguments{
\item{type}{string: The type}

\item{value}{string | double: The value}
}
\description{
tag: search_length_algorithm
}
\examples{
prj_search_length_algorithm(type = "fixed",
                               value = 1e-06)
}
