% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variable.R
\name{prj_source_term}
\alias{prj_source_term}
\title{prj_source_term}
\usage{
prj_source_term(
  type,
  parameter = NULL,
  geometrical_set = NULL,
  geometry = NULL,
  mesh = NULL,
  source_term_object = NULL,
  flush_stdout = NULL
)
}
\arguments{
\item{type}{string:}

\item{parameter}{Optional: string:}

\item{geometrical_set}{Optional: string:}

\item{geometry}{Optional: string:}

\item{mesh}{Optional: string:}

\item{source_term_object}{Optional: string:}

\item{flush_stdout}{Optional: string:}
}
\description{
tag: source_term
}
\examples{
prj_source_term(
    type = "Nodal",
    parameter = "sourceterm_heat_transfer_coefficient",
    geometrical_set = "geometry",
    geometry = "po0"
)
}
