% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{prj_submesh_residuum_output}
\alias{prj_submesh_residuum_output}
\title{prj_submesh_residuum_output}
\usage{
prj_submesh_residuum_output(
  type,
  prefix,
  variables,
  suffix = NULL,
  timesteps = NULL,
  compress_output = NULL,
  data_mode = NULL,
  output_iteration_results = NULL,
  meshes = NULL,
  fixed_output_times = NULL,
  hdf = NULL,
  geometrical_sets = NULL
)
}
\arguments{
\item{type}{string:}

\item{prefix}{string:}

\item{variables}{vector:}

\item{suffix}{Optional: string:}

\item{timesteps}{Optional:}

\item{compress_output}{Optional: string: Should the output be compressed?
Either "true" or "false"}

\item{data_mode}{Optional: string:}

\item{output_iteration_results}{Optional: string: Either "true" or "false"}

\item{meshes}{Optional: list: A list of meshes}

\item{fixed_output_times}{Optional: string | numeric:}

\item{hdf}{Optional: numeric}

\item{geometrical_sets}{Optional: list, prj_geometrical_set}
}
\description{
tag: submesh_residuum_output
}
