% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{prj_surface}
\alias{prj_surface}
\title{prj_surface}
\usage{
prj_surface(site_unit = NULL, ...)
}
\arguments{
\item{site_unit}{Optional: string}

\item{...}{site}
}
\description{
tag: surface
}
