% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{prj_tl_process}
\alias{prj_tl_process}
\title{prj_tl_process}
\usage{
prj_tl_process(
  ref,
  nonlinear_solver,
  convergence_criterion,
  time_discretization,
  time_stepping,
  compensate_non_equilibrium_initial_residuum = NULL
)
}
\arguments{
\item{ref}{string: References a prj_process object by name}

\item{nonlinear_solver}{string:}

\item{convergence_criterion}{prj_convergence_criterion:}

\item{time_discretization}{vector:}

\item{time_stepping}{prj_time_stepping:}

\item{compensate_non_equilibrium_initial_residuum}{string: Optional: Either
"true" or "false"}
}
\description{
tag: process (parent: time_loop, NOT processes!)
}
\examples{
prj_tl_process(
    ref = "HM",
    nonlinear_solver = "basic_newton",
    convergence_criterion = prj_convergence_criterion(
        type = "DeltaX",
        norm_type = "NORM2",
        reltol = 1e-08
    ),
    time_discretization = list(type = "BackwardEuler"),
    time_stepping = prj_time_stepping(
        type = "FixedTimeStepping",
        t_initial = 0,
        t_end = 10000,
        timesteps = list(pair = list(1000,
                                     delta_t = 10))
    )
)
}
