% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prj_top_level_tags}
\alias{prj_top_level_tags}
\title{prj_top_level_tags}
\usage{
prj_top_level_tags()
}
\value{
list: List of lists.
}
\description{
Gets top level .prj tags along with info if they are required.
}
