% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_borehole_heat_exchanger.R
\name{r2ogs6_borehole_heat_exchanger}
\alias{r2ogs6_borehole_heat_exchanger}
\title{r2ogs6_borehole_heat_exchanger}
\usage{
r2ogs6_borehole_heat_exchanger(
  type,
  flow_and_temperature_control,
  borehole,
  grout,
  pipes,
  refrigerant,
  use_bhe_pipe_network = NULL
)
}
\arguments{
\item{type}{string:}

\item{flow_and_temperature_control}{r2ogs6_flow_and_temperature_control:}

\item{borehole}{list:}

\item{grout}{list:}

\item{pipes}{r2ogs6_pipes:}

\item{refrigerant}{list:}

\item{use_bhe_pipe_network}{Optional: string ("true" | "false"):}
}
\description{
tag: borehole_heat_exchanger
}
