% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variables.R
\name{r2ogs6_boundary_condition}
\alias{r2ogs6_boundary_condition}
\title{r2ogs6_boundary_condition}
\usage{
r2ogs6_boundary_condition(
  type,
  parameter,
  component = NULL,
  mesh = NULL,
  geometrical_set = NULL,
  geometry = NULL
)
}
\arguments{
\item{type}{...}

\item{parameter}{...}

\item{component}{...}

\item{mesh}{...}

\item{geometrical_set}{...}

\item{geometry}{...}
}
\description{
S3 class describing a .prj boundary condition
}
