% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_porous_medium.R
\name{r2ogs6_capillary_pressure}
\alias{r2ogs6_capillary_pressure}
\title{r2ogs6_capillary_pressure}
\usage{
r2ogs6_capillary_pressure(
  type,
  pd = NULL,
  sr = NULL,
  smax = NULL,
  m = NULL,
  pc_max = NULL,
  has_regularized = NULL,
  curve = NULL
)
}
\arguments{
\item{type}{string:}

\item{pd}{Optional: string | double:}

\item{sr}{Optional: string | double:}

\item{smax}{Optional: string | double:}

\item{m}{Optional: string | double:}

\item{pc_max}{Optional: string | double:}

\item{has_regularized}{Optional: string, "true" | "false":}

\item{curve}{Optional: list:}
}
\description{
tag: capillary_pressure
}
\examples{
r2ogs6_capillary_pressure(
    type = "vanGenuchten",
    pd = 2e+06,
    sr = 0.4,
    smax = 1,
    m = 0.3288590604,
    pc_max = 2e+07,
    has_regularized = "true"
)
}
