% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_com_property}
\alias{r2ogs6_com_property}
\title{r2ogs6_com_property}
\usage{
r2ogs6_com_property(name, type, value = NULL, parameter_name = NULL)
}
\arguments{
\item{name}{string: Property name}

\item{type}{string: Property type}

\item{value}{Optional: string | double: ...}

\item{parameter_name}{Optional:}
}
\description{
tag: property
}
\examples{
r2ogs6_com_property(name = "molecular_diffusion",
                    type = "Constant",
                    value = 2e-09)
}
