% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variable.R
\name{r2ogs6_deactivated_subdomain}
\alias{r2ogs6_deactivated_subdomain}
\title{r2ogs6_deactivated_subdomain}
\usage{
r2ogs6_deactivated_subdomain(time_interval, material_ids)
}
\arguments{
\item{time_interval}{list, numeric:}

\item{material_ids}{string | double:}
}
\description{
tag: deactivated_subdomain
}
\examples{
r2ogs6_deactivated_subdomain(time_interval = list(start = "0",
                                                  end = "1"),
                             material_ids = 1)
}
