% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_borehole_heat_exchanger.R
\name{r2ogs6_flow_and_temperature_control}
\alias{r2ogs6_flow_and_temperature_control}
\title{r2ogs6_flow_and_temperature_control}
\usage{
r2ogs6_flow_and_temperature_control(
  type,
  flow_rate = NULL,
  temperature_curve = NULL,
  power = NULL,
  power_curve = NULL,
  flow_rate_curve = NULL
)
}
\arguments{
\item{type}{string:}

\item{flow_rate}{Optional: string | double:}

\item{temperature_curve}{Optional: string:}

\item{power}{Optional: string | double:}

\item{power_curve}{Optional: string:}

\item{flow_rate_curve}{Optional: string:}
}
\description{
tag: flow_and_temperature_control
}
