% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_material_property.R
\name{r2ogs6_fluid}
\alias{r2ogs6_fluid}
\title{r2ogs6_fluid}
\usage{
r2ogs6_fluid(
  liquid_density,
  gas_density,
  liquid_viscosity,
  gas_viscosity,
  specific_heat_capacity_solid = NULL,
  specific_heat_capacity_water = NULL,
  specific_heat_capacity_air = NULL,
  specific_heat_capacity_water_vapor = NULL,
  thermal_conductivity_dry_solid = NULL,
  thermal_conductivity_wet_solid = NULL
)
}
\arguments{
\item{liquid_density}{list:}

\item{gas_density}{list:}

\item{liquid_viscosity}{list:}

\item{gas_viscosity}{list:}

\item{specific_heat_capacity_solid}{Optional:}

\item{specific_heat_capacity_water}{Optional:}

\item{specific_heat_capacity_air}{Optional:}

\item{specific_heat_capacity_water_vapor}{Optional:}

\item{thermal_conductivity_dry_solid}{Optional:}

\item{thermal_conductivity_wet_solid}{Optional:}
}
\description{
tag: fluid
}
\examples{
r2ogs6_fluid(
    liquid_density = list(type = "Constant",
                          value = " 1.e3 "),
    gas_density = list(type = "IdealGasLaw",
                       molar_mass = " 0.002 "),
    liquid_viscosity = list(type = "Constant",
                            value = " 3.171e-11 "),
    gas_viscosity = list(type = "Constant",
                         value = " 2.8539e-13 ")
)
}
