% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_global_process_coupling}
\alias{r2ogs6_global_process_coupling}
\title{r2ogs6_global_process_coupling}
\usage{
r2ogs6_global_process_coupling(max_iter, convergence_criteria)
}
\arguments{
\item{max_iter}{string | double: Maximal number of iterations}

\item{convergence_criteria}{list, r2ogs6_convergence_criterion:
Convergence criteria}
}
\description{
tag: global_process_coupling
}
\examples{
r2ogs6_global_process_coupling(
    max_iter = 100,
    convergence_criteria = list(
        convergence_criterion = r2ogs6_convergence_criterion(
            type = "DeltaX",
            norm_type = "INFINITY_N",
            abstol = 1e-08,
            reltol = 1e-10
        )
    )
)
}
