% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_insitu.R
\name{r2ogs6_insitu}
\alias{r2ogs6_insitu}
\title{r2ogs6_insitu}
\usage{
r2ogs6_insitu(scripts)
}
\arguments{
\item{scripts}{character: Script names}
}
\description{
tag: insitu
}
\examples{
r2ogs6_insitu(scripts = c(script = "square_1e1_neumann-insitu.py"))
}
