% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{r2ogs6_kinetic_reactant}
\alias{r2ogs6_kinetic_reactant}
\title{r2ogs6_kinetic_reactant}
\usage{
r2ogs6_kinetic_reactant(
  name,
  initial_amount,
  chemical_formula = NULL,
  fix_amount = NULL
)
}
\arguments{
\item{name}{The component name}

\item{initial_amount}{The initial amount of the component}

\item{chemical_formula}{The chemical formula of the component}

\item{fix_amount}{Should the amount be fixed or not?}
}
\description{
S3 class describing .prj kinetic_reactant
}
