% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_linear_solver.R
\name{r2ogs6_linear_solver}
\alias{r2ogs6_linear_solver}
\title{r2ogs6_linear_solver}
\usage{
r2ogs6_linear_solver(name, eigen = NULL, lis = NULL, petsc = NULL)
}
\arguments{
\item{name}{string: The name of the linear solver}

\item{eigen}{Optional: list: ...}

\item{lis}{Optional: string: ...}

\item{petsc}{Optional: character: ...}
}
\description{
tag: linear_solver
}
\examples{
r2ogs6_linear_solver(
    name = "general_linear_solver",
    eigen = r2ogs6_eigen(
        solver_type = "BiCGSTAB",
        precon_type = "ILUT",
        max_iteration_step = 10000,
        error_tolerance = 1e-16
    ),
    lis = "-i bicgstab -p ilu -tol 1e-16 -maxiter 10000"
)
}
