% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_local_coordinate_system.R
\name{r2ogs6_local_coordinate_system}
\alias{r2ogs6_local_coordinate_system}
\title{r2ogs6_local_coordinate_system}
\usage{
r2ogs6_local_coordinate_system(
  basis_vector_0,
  basis_vector_1,
  basis_vector_2 = NULL
)
}
\arguments{
\item{basis_vector_0}{string | double: A basis vector}

\item{basis_vector_1}{string | double: A basis vector}

\item{basis_vector_2}{Optional: string | double: A basis vector}
}
\description{
tag: local_coordinate_system
}
\examples{
r2ogs6_local_coordinate_system(basis_vector_0 = "e0",
                               basis_vector_1 = "e1",
                               basis_vector_2 = "e2")
}
