% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_material_property.R
\name{r2ogs6_material_property}
\alias{r2ogs6_material_property}
\title{r2ogs6_material_property}
\usage{
r2ogs6_material_property(fluid, porous_medium)
}
\arguments{
\item{fluid}{r2ogs6_fluid:}

\item{porous_medium}{r2ogs6_porous_medium:}
}
\description{
tag: material_property
}
