% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_medium}
\alias{r2ogs6_medium}
\title{r2ogs6_medium}
\usage{
r2ogs6_medium(phases, properties = NULL, id = NULL)
}
\arguments{
\item{phases}{list, r2ogs6_phase: Medium phases}

\item{properties}{list, r2ogs6_pr_property: Medium properties}

\item{id}{string | double: ID corresponding to the MaterialIDs}
}
\description{
tag: medium, a specific medium with optional id corresponding
to the MaterialIDs
}
