% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{r2ogs6_medium_phase}
\alias{r2ogs6_medium_phase}
\title{r2ogs6_medium_phase}
\usage{
r2ogs6_medium_phase(type, properties)
}
\arguments{
\item{type}{A string specifying the medium type (one of "Gas", "Solid", "AqueousLiquid" and "NonAqueousLiquid")}

\item{properties}{A list of properties (see ?r2ogs6_medium_property for more info)}
}
\description{
S3 class describing a .prj medium phase (a coherent material with homogeneous properties)
}
