% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{r2ogs6_medium_property}
\alias{r2ogs6_medium_property}
\title{r2ogs6_medium_property}
\usage{
r2ogs6_medium_property(name, type, value = NULL, ...)
}
\arguments{
\item{name}{A string specifying the property name}

\item{type}{A string specifying the property type}

\item{value}{A string ...}

\item{...}{...}
}
\description{
S3 class describing a .prj medium property (some rare parameters are NOT implemented yet!)
}
