% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{r2ogs6_mesh}
\alias{r2ogs6_mesh}
\title{r2ogs6_mesh}
\usage{
r2ogs6_mesh(mesh_ref)
}
\arguments{
\item{mesh_ref}{A reference to a .vtu file}
}
\description{
Wrapper class, so far only has a reference to a .vtu file
}
