% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_nonlinear_solver.R
\name{r2ogs6_nonlinear_solver}
\alias{r2ogs6_nonlinear_solver}
\title{r2ogs6_nonlinear_solver}
\usage{
r2ogs6_nonlinear_solver(name, type, max_iter, linear_solver)
}
\arguments{
\item{name}{The name of the nonlinear solver}

\item{type}{The type of the nonlinear solver}

\item{max_iter}{The maximal number of iterations}

\item{linear_solver}{The name of the linear solver}
}
\description{
S3 class describing a .prj nonlinear solver
}
