% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_output}
\alias{r2ogs6_output}
\title{r2ogs6_output}
\usage{
r2ogs6_output(
  type,
  prefix,
  variables,
  suffix = NULL,
  timesteps = NULL,
  compress_output = NULL,
  data_mode = NULL,
  output_iteration_results = NULL,
  meshes = NULL,
  fixed_output_times = NULL
)
}
\arguments{
\item{type}{string:}

\item{prefix}{string:}

\item{variables}{vector:}

\item{suffix}{Optional: string:}

\item{timesteps}{Optional:}

\item{compress_output}{Optional: string: Should the output be compressed?
Either "true" or "false"}

\item{data_mode}{Optional: string:}

\item{output_iteration_results}{Optional: string: Either "true" or "false"}

\item{meshes}{Optional: character: A vector of mesh names}

\item{fixed_output_times}{Optional: string | numeric:}
}
\description{
tag: output
}
