% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_parameters.R
\name{r2ogs6_parameter}
\alias{r2ogs6_parameter}
\title{r2ogs6_parameter}
\usage{
r2ogs6_parameter(name, type, values)
}
\arguments{
\item{name}{The parameter name}

\item{type}{The parameter type}

\item{values}{The parameter values}
}
\description{
S3 class describing a .prj parameter
}
