% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_ph_property}
\alias{r2ogs6_ph_property}
\title{r2ogs6_ph_property}
\usage{
r2ogs6_ph_property(
  name,
  type,
  value = NULL,
  reference_value = NULL,
  independent_variable = NULL,
  initial_porosity = NULL,
  minimal_porosity = NULL,
  maximal_porosity = NULL,
  parameter_name = NULL,
  exponents = NULL,
  offset = NULL,
  exponent = NULL,
  swelling_pressures = NULL,
  lower_saturation_limit = NULL,
  upper_saturation_limit = NULL,
  intrinsic_permeabilities = NULL
)
}
\arguments{
\item{name}{string: Property name}

\item{type}{string: Property type}

\item{value}{Optional: string | double: ...}

\item{reference_value}{Optional: string | double:}

\item{independent_variable}{Optional: list:}

\item{initial_porosity}{Optional: string:}

\item{minimal_porosity}{Optional: string | double:}

\item{maximal_porosity}{Optional: string | double:}

\item{parameter_name}{Optional: string:}

\item{exponents}{Optional: string | numeric:}

\item{offset}{Optional: string | double:}

\item{exponent}{Optional: list:}

\item{swelling_pressures}{Optional: string | numeric:}

\item{lower_saturation_limit}{Optional: string | double:}

\item{upper_saturation_limit}{Optional: string | double:}

\item{intrinsic_permeabilities}{Optional: string | numeric:}
}
\description{
tag: property
}
