% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_phase}
\alias{r2ogs6_phase}
\title{r2ogs6_phase}
\usage{
r2ogs6_phase(type, properties = NULL, components = NULL)
}
\arguments{
\item{type}{string: Phase type
(get valid types with get_valid_phase_types())}

\item{properties}{list, r2ogs6_pr_property: Properties}

\item{components}{list, components}
}
\description{
tag: phase, a coherent material with homogeneous properties
}
\examples{
r2ogs6_phase(
    type = "Solid",
    properties = list(
        property = r2ogs6_ph_property(name = "porosity",
                                      type = "Constant",
                                      value = 0.3)
    )
)
}
