% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{r2ogs6_phasefield_parameters}
\alias{r2ogs6_phasefield_parameters}
\title{r2ogs6_phasefield_parameters}
\usage{
r2ogs6_phasefield_parameters(
  residual_stiffness,
  crack_resistance,
  crack_length_scale,
  kinetic_coefficient,
  history_field = NULL
)
}
\arguments{
\item{residual_stiffness}{string:}

\item{crack_resistance}{string:}

\item{crack_length_scale}{string:}

\item{kinetic_coefficient}{string:}

\item{history_field}{Optional: string:}
}
\description{
tag: phasefield_parameters
}
