% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_borehole_heat_exchanger.R
\name{r2ogs6_pipes}
\alias{r2ogs6_pipes}
\title{r2ogs6_pipes}
\usage{
r2ogs6_pipes(
  longitudinal_dispersion_length,
  inlet = NULL,
  outlet = NULL,
  distance_between_pipes = NULL,
  outer = NULL,
  inner = NULL
)
}
\arguments{
\item{longitudinal_dispersion_length}{string | double:}

\item{inlet}{Optional: list:}

\item{outlet}{Optional: list:}

\item{distance_between_pipes}{Optional: string | double:}

\item{outer}{Optional: list:}

\item{inner}{Optional: list:}
}
\description{
tag: pipes
}
