% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_porous_medium.R
\name{r2ogs6_porous_medium}
\alias{r2ogs6_porous_medium}
\title{r2ogs6_porous_medium}
\usage{
r2ogs6_porous_medium(
  id,
  permeability,
  porosity,
  storage,
  capillary_pressure,
  relative_permeability
)
}
\arguments{
\item{id}{string | double:}

\item{permeability}{list:}

\item{porosity}{list:}

\item{storage}{list:}

\item{capillary_pressure}{r2ogs6_capillary_pressure:}

\item{relative_permeability}{list:}
}
\description{
tag: porous_medium
}
\examples{
r2ogs6_porous_medium(
    id = 0,
    permeability = list(permeability_tensor_entries = "kappa1",
                        type = "Constant"),
    porosity = list(type = "Constant",
                    porosity_parameter = "constant_porosity_parameter"),
    storage = list(type = "Constant",
                   value = " 0.0 "),
    capillary_pressure = r2ogs6_capillary_pressure(
        type = "vanGenuchten",
        pd = 2e+06,
        sr = 0.4,
        smax = 1,
        m = 0.3288590604,
        pc_max = 2e+07,
        has_regularized = "true"
    ),
    relative_permeability = list(
        relative_permeability = r2ogs6_relative_permeability(
            type = "NonWettingPhaseVanGenuchten",
            sr = 0,
            smax = 0.6,
            m = 0.3288590604,
            krel_min = 0,
            id = "0"
        )
    )
)
}
