% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_pr_property}
\alias{r2ogs6_pr_property}
\title{r2ogs6_pr_property}
\usage{
r2ogs6_pr_property(
  name,
  type,
  value = NULL,
  parameter_name = NULL,
  exponent = NULL,
  residual_liquid_saturation = NULL,
  residual_gas_saturation = NULL,
  p_b = NULL,
  independent_variable = NULL,
  curve = NULL,
  minimum_relative_permeability_liquid = NULL,
  initial_permeability = NULL,
  maximum_permeability = NULL,
  lambda = NULL,
  cutoff_value = NULL,
  intrinsic_permeability = NULL,
  initial_aperture = NULL,
  mean_frac_distance = NULL,
  threshold_strain = NULL,
  fracture_normal = NULL,
  reference_permeability = NULL,
  fitting_factor = NULL,
  cohesion = NULL,
  friction_angle = NULL,
  tensile_strength_parameter = NULL,
  b1 = NULL,
  b2 = NULL,
  b3 = NULL,
  minimum_permeability = NULL,
  initial_porosity = NULL,
  entry_pressure = NULL,
  min_relative_permeability_liquid = NULL,
  min_relative_permeability_gas = NULL
)
}
\arguments{
\item{name}{string:}

\item{type}{string:}

\item{value}{Optional:}

\item{parameter_name}{Optional:}

\item{exponent}{Optional:}

\item{residual_liquid_saturation}{Optional:}

\item{residual_gas_saturation}{Optional:}

\item{p_b}{Optional:}

\item{independent_variable}{Optional:}

\item{curve}{Optional:}

\item{minimum_relative_permeability_liquid}{Optional:}

\item{initial_permeability}{Optional:}

\item{maximum_permeability}{Optional:}

\item{lambda}{Optional:}

\item{cutoff_value}{Optional:}

\item{intrinsic_permeability}{Optional:}

\item{initial_aperture}{Optional:}

\item{mean_frac_distance}{Optional:}

\item{threshold_strain}{Optional:}

\item{fracture_normal}{Optional:}

\item{reference_permeability}{Optional:}

\item{fitting_factor}{Optional:}

\item{cohesion}{Optional:}

\item{friction_angle}{Optional:}

\item{tensile_strength_parameter}{Optional:}

\item{b1}{Optional:}

\item{b2}{Optional:}

\item{b3}{Optional:}

\item{minimum_permeability}{Optional:}

\item{initial_porosity}{Optional:}

\item{entry_pressure}{Optional:}

\item{min_relative_permeability_liquid}{Optional:}

\item{min_relative_permeability_gas}{Optional:}
}
\description{
tag: property
}
