% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{r2ogs6_process}
\alias{r2ogs6_process}
\title{r2ogs6_process}
\usage{
r2ogs6_process(
  name,
  type,
  integration_order,
  dimension,
  constitutive_relation,
  process_variables,
  secondary_variables,
  specific_body_force,
  coupling_scheme = NULL
)
}
\arguments{
\item{name}{The process name}

\item{type}{The process type}

\item{integration_order}{...}

\item{dimension}{...}

\item{constitutive_relation}{...}

\item{process_variables}{...}

\item{secondary_variables}{A list of secondary variables}

\item{specific_body_force}{The specific body force}

\item{coupling_scheme}{Optional:}
}
\description{
S3 class describing a .prj (processes) process
}
