% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{r2ogs6_rate}
\alias{r2ogs6_rate}
\title{r2ogs6_rate}
\usage{
r2ogs6_rate(kinetic_reactant, expression)
}
\arguments{
\item{kinetic_reactant}{string: References a kinetic_reactant object}

\item{expression}{character: Statements}
}
\description{
S3 class describing .prj rate
}
