% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_search_length_algorithm.R
\name{r2ogs6_search_length_algorithm}
\alias{r2ogs6_search_length_algorithm}
\title{r2ogs6_search_length_algorithm}
\usage{
r2ogs6_search_length_algorithm(type, value = NULL)
}
\arguments{
\item{type}{string: The type}

\item{value}{string | double: The value}
}
\description{
tag: search_length_algorithm
}
