% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{r2ogs6_solution}
\alias{r2ogs6_solution}
\title{r2ogs6_solution}
\usage{
r2ogs6_solution(temperature, pressure, pe, components, charge_balance = NULL)
}
\arguments{
\item{temperature}{string | double: Temperature}

\item{pressure}{string | double: Pressure}

\item{pe}{string | double: pe}

\item{components}{list: Components}

\item{charge_balance}{Optional: string: Charge balance}
}
\description{
tag: solution
}
\examples{
r2ogs6_solution(
    temperature = 25,
    pressure = 1,
    pe = 4,
    components = list(
        component = "C(4)",
        component = "Ca"
    ),
    charge_balance = "pH"
)
}
