% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_time_loop}
\alias{r2ogs6_time_loop}
\title{r2ogs6_time_loop}
\usage{
r2ogs6_time_loop(processes, output, global_process_coupling = NULL)
}
\arguments{
\item{processes}{list, r2ogs6_tl_process:}

\item{output}{r2ogs6_output:}

\item{global_process_coupling}{Optional: r2ogs6_global_process_coupling:}
}
\description{
tag: time_loop
}
