% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_time_stepping}
\alias{r2ogs6_time_stepping}
\title{r2ogs6_time_stepping}
\usage{
r2ogs6_time_stepping(
  type,
  t_initial = NULL,
  t_end = NULL,
  timesteps = NULL,
  initial_dt = NULL,
  minimum_dt = NULL,
  maximum_dt = NULL,
  number_iterations = NULL,
  multiplier = NULL,
  dt_guess = NULL,
  dt_min = NULL,
  dt_max = NULL,
  rel_dt_min = NULL,
  rel_dt_max = NULL,
  tol = NULL
)
}
\arguments{
\item{type}{string:}

\item{t_initial}{Optional: string | double:}

\item{t_end}{Optional: string | double:}

\item{timesteps}{Optional: list:}

\item{initial_dt}{Optional: string | double:}

\item{minimum_dt}{Optional: string | double:}

\item{maximum_dt}{Optional: string | double:}

\item{number_iterations}{Optional: string | numeric:}

\item{multiplier}{Optional: string | numeric:}

\item{dt_guess}{Optional: string | double:}

\item{dt_min}{Optional: string | double:}

\item{dt_max}{Optional: string | double:}

\item{rel_dt_min}{Optional: string | double:}

\item{rel_dt_max}{Optional: string | double:}

\item{tol}{Optional: string | double:}
}
\description{
tag: time_stepping
}
