% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_tl_output}
\alias{r2ogs6_tl_output}
\title{r2ogs6_tl_output}
\usage{
r2ogs6_tl_output(
  type,
  prefix,
  suffix,
  timesteps,
  variables,
  compress_output = NULL
)
}
\arguments{
\item{type}{...}

\item{prefix}{...}

\item{suffix}{...}

\item{timesteps}{...}

\item{variables}{...}

\item{compress_output}{Optional: Should the output be compressed?}
}
\description{
S3 class describing a .prj time_loop output
}
