% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_tl_process}
\alias{r2ogs6_tl_process}
\title{r2ogs6_tl_process}
\usage{
r2ogs6_tl_process(
  ref,
  nonlinear_solver,
  convergence_criterion,
  time_discretization,
  time_stepping
)
}
\arguments{
\item{ref}{References a r2ogs6_process object by name}

\item{nonlinear_solver}{...}

\item{convergence_criterion}{...}

\item{time_discretization}{...}

\item{time_stepping}{...}
}
\description{
S3 class describing a .prj time_loop process
}
