% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_vtkdiff.R
\name{r2ogs6_vtkdiff}
\alias{r2ogs6_vtkdiff}
\title{r2ogs6_vtkdiff}
\usage{
r2ogs6_vtkdiff(
  field,
  absolute_tolerance,
  relative_tolerance,
  file = NULL,
  regex = NULL
)
}
\arguments{
\item{field}{string: ...}

\item{absolute_tolerance}{string | double: Absolute tolerance}

\item{relative_tolerance}{string | double: Relative tolerance}

\item{file}{string: Optional: File}

\item{regex}{string: Optional: A regular expression}
}
\description{
tag: vtkdiff
}
\examples{
r2ogs6_vtkdiff(field = "displacement",
               absolute_tolerance = 1e-15,
               relative_tolerance = 0,
               regex = "gravity_pcs_0_ts_.*_t_.*.vtu")
}
