% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_from_file.R
\name{read_gml_points}
\alias{read_gml_points}
\title{Reads in a .tsv file and parses it as a tibble.
A column specifying the name of the points is optional, but the number of columns must be consistent.
If one row has an entry for the point name, all rows need an entry for the point name.
If only some points need to be named, put "" (double quotes) for the points that don't have a name.}
\usage{
read_gml_points(file_name)
}
\arguments{
\item{file_name}{The name of the .tsv file containing the point tibble}
}
\description{
Reads in a .tsv file and parses it as a tibble.
A column specifying the name of the points is optional, but the number of columns must be consistent.
If one row has an entry for the point name, all rows need an entry for the point name.
If only some points need to be named, put "" (double quotes) for the points that don't have a name.
}
