% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_curves}
\alias{read_in_curves}
\title{read_in_curves}
\usage{
read_in_curves(ogs6_obj, prj_path, curve_names = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the curve elements should be read from}

\item{curve_names}{Optional: The names of the curve elements to be read in}
}
\description{
Reads in curve elements from a .prj file
}
