% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_linear_solvers}
\alias{read_in_linear_solvers}
\title{read_in_linear_solvers}
\usage{
read_in_linear_solvers(ogs6_obj, prj_path, linear_solver_names = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the linear_solver elements should be read from}

\item{linear_solver_names}{Optional: The names of the linear_solver elements to be read in}
}
\description{
Reads in linear_solver elements from a .prj file
}
