% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_media}
\alias{read_in_media}
\title{read_in_media}
\usage{
read_in_media(ogs6_obj, prj_path, medium_indices = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the medium elements should be read from}

\item{medium_indices}{Optional: The indices of the medium elements to be read in}
}
\description{
Reads in medium elements from a .prj file
}
