% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_nonlinear_solvers}
\alias{read_in_nonlinear_solvers}
\title{read_in_nonlinear_solvers}
\usage{
read_in_nonlinear_solvers(ogs6_obj, prj_path, nonlinear_solver_names = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the nonlinear_solver elements should be read from}

\item{nonlinear_solver_names}{Optional: The names of the nonlinear_solver elements to be read in}
}
\description{
Reads in nonlinear_solver elements from a .prj file
}
