% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_parameters}
\alias{read_in_parameters}
\title{read_in_parameters}
\usage{
read_in_parameters(ogs6_obj, prj_path, parameter_names = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the parameter elements should be read from}

\item{parameter_names}{Optional: The names of the parameter elements to be read in}
}
\description{
Reads in parameter elements from a .prj file
}
