% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_gml.R
\name{read_in_points}
\alias{read_in_points}
\title{read_in_points}
\usage{
read_in_points(xml_doc)
}
\arguments{
\item{xml_doc}{A parsed XML document (of class \code{xml2::xml_document})}
}
\description{
Reads points from a \code{.gml} file
}
