% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_process_variables}
\alias{read_in_process_variables}
\title{read_in_process_variables}
\usage{
read_in_process_variables(ogs6_obj, prj_path, process_variable_names = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the process_variable elements should be read from}

\item{process_variable_names}{Optional: The names of the process_variable elements to be read in}
}
\description{
Reads in process_variable elements from a .prj file
}
