% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_processes}
\alias{read_in_processes}
\title{read_in_processes}
\usage{
read_in_processes(ogs6_obj, prj_path, process_names = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the process elements should be read from}

\item{process_names}{Optional: The names of the process elements to be read in}
}
\description{
Reads in process elements from a .prj file
}
