% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_vtu.R
\name{read_in_pvd}
\alias{read_in_pvd}
\title{read_in_pvd}
\usage{
read_in_pvd(pvd_path)
}
\arguments{
\item{pvd_path}{string: Path to .pvd file that should be read in}
}
\value{
character: Vector containing .vtu paths
}
\description{
Function to read in a .pvd file
}
