% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_test_definition}
\alias{read_in_test_definition}
\title{read_in_test_definition}
\usage{
read_in_test_definition(ogs6_obj, prj_path, vtkdiff_indices = NULL)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the vtkdiff elements should be read from}

\item{vtkdiff_indices}{Optional: The indices of the vtkdiff elements to be read in}
}
\description{
Reads in vtkdiff elements from a .prj file
}
