% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_time_loop}
\alias{read_in_time_loop}
\title{read_in_time_loop}
\usage{
read_in_time_loop(ogs6_obj, prj_path)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file the time_loop element should be read from}
}
\description{
Reads in time_loop element from a .prj file
}
