% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_timesteps_node}
\alias{read_in_timesteps_node}
\title{read_in_timesteps_node}
\usage{
read_in_timesteps_node(timesteps_node)
}
\arguments{
\item{timesteps_node}{The timesteps node}
}
\description{
Reads in a timesteps node
}
