% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_vtu.R
\name{read_in_vtu}
\alias{read_in_vtu}
\title{read_in_vtu}
\usage{
read_in_vtu(vtu_path)
}
\arguments{
\item{vtu_path}{string: Path to .vtu file that should be read in}
}
\value{
OGS6_vtu: Mesh object
}
\description{
Wrapper function to read in a whole .vtu file
}
