% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{run_all_benchmarks}
\alias{run_all_benchmarks}
\title{run_all_benchmarks}
\usage{
run_all_benchmarks(
  path,
  ogs_processlib_path,
  ogs_bin_path,
  sim_path,
  starting_from_prj_path = "",
  print_results = TRUE
)
}
\arguments{
\item{path}{string: Path to benchmark folder}

\item{ogs_processlib_path}{string: Path to OpenGeoSys 6 ProcessLib folder
which contains relevant Tests.cmake files}

\item{ogs_bin_path}{string:}

\item{sim_path}{string: Path where simulation files will be saved}

\item{starting_from_prj_path}{string: .prj path to start from}

\item{print_results}{flag: Print results in the end?}
}
\description{
Utility function, for quick benchmark runs. Calls
run_benchmark internally.
}
