% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_benchmarks.R
\name{scrape_benchmarks}
\alias{scrape_benchmarks}
\title{scrape_benchmarks}
\usage{
scrape_benchmarks(
  url = "https://gitlab.opengeosys.org/ogs/ogs/-/tree/master/Tests/Data/",
  path = "extdata/benchmarks/"
)
}
\arguments{
\item{url}{Optional: An URL}

\item{path}{Optional: The path to save the benchmark folders to}
}
\description{
Downloads all OGS6 benchmarks (WIP)
}
