% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_fitting_subclass}
\alias{select_fitting_subclass}
\title{select_fitting_subclass}
\usage{
select_fitting_subclass(xpath_expr, subclasses_names)
}
\arguments{
\item{xpath_expr}{string: An XPath expression}

\item{subclasses_names}{character: A named character vector}
}
\description{
Utility function to differentiate which property class to pick
i.e. when dealing with r2ogs6 which has 3 subclasses with the tag name
'property'
}
