% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{setup_logging}
\alias{setup_logging}
\title{setup_logging}
\usage{
setup_logging(sim_name, sim_path, ogs6_call)
}
\arguments{
\item{sim_name}{string: Simulation name}

\item{sim_path}{string: Simulation path}

\item{ogs6_call}{string: Corresponding OGS6 call}
}
\description{
Sets up logging.
}
