% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_vector_to_node}
\alias{simple_vector_to_node}
\title{simple_list_to_node}
\usage{
simple_vector_to_node(parent_name, simple_vector)
}
\arguments{
\item{parent_name}{The name of the parent node}

\item{simple_vector}{The vector to turn into the node structure}
}
\description{
Helper to turn a simple vector into the corresponding node structure
with the vector elements as children. This works for lists too (as they are vectors).
}
