% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_functions.R
\name{strings_equal}
\alias{strings_equal}
\title{strings_equal}
\usage{
strings_equal(strings_vector)
}
\arguments{
\item{strings_vector}{A vector or a list of strings}
}
\description{
Checks if all strings in a vector or list are equal
}
