% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gml.R
\name{surfaces_to_xml}
\alias{surfaces_to_xml}
\title{surfaces_to_xml}
\usage{
surfaces_to_xml(surfaces)
}
\arguments{
\item{surfaces}{A list of surfaces}
}
\value{
An XML node containing the surfaces
}
\description{
Turns a list of surfaces into an XML node
}
