% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_ensemble.R
\name{validate_ensemble_parameters}
\alias{validate_ensemble_parameters}
\title{validate_ensemble_parameters}
\usage{
validate_ensemble_parameters(parameters, ogs6_obj_ref)
}
\arguments{
\item{parameters}{list: The specified parameters}

\item{ogs6_obj_ref}{string: Deparsed argument, for checking if the parameters
came from the same OGS6 object that the OGS6_Ensemble is based on}
}
\description{
Validates the parameters given to an OGS6_Ensemble object
}
