% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gml.R
\name{validate_gml_data}
\alias{validate_gml_data}
\title{validate_gml_data}
\usage{
validate_gml_data(gml_obj)
}
\arguments{
\item{gml_obj}{A gml class object}
}
\description{
Checks if the defined gml class object isn't empty
}
