% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_dir_path}
\alias{validate_is_dir_path}
\title{validate_is_dir_path}
\usage{
validate_is_dir_path(path)
}
\arguments{
\item{path}{string: A path}
}
\description{
Checks if a given path ends on '/'
}
