% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_null_or_class_obj}
\alias{validate_is_null_or_class_obj}
\title{validate_is_null_or_class_obj}
\usage{
validate_is_null_or_class_obj(obj, class_name)
}
\arguments{
\item{obj}{The object to check}

\item{class_name}{The name of the expected class}
}
\description{
Checks if an object is either null or a class object of class
'class_name'
}
