% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_null_or_number}
\alias{validate_is_null_or_number}
\title{validate_is_null_or_number}
\usage{
validate_is_null_or_number(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if an object is either null or a number (helper to save
some typing when validating optional object parameters)
}
