% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_null_or_numeric}
\alias{validate_is_null_or_numeric}
\title{validate_is_null_or_numeric}
\usage{
validate_is_null_or_numeric(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if an object is either null or numeric (helper to save
some typing when validating optional object parameters)
}
