% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_null_or_string}
\alias{validate_is_null_or_string}
\title{validate_is_null_or_string}
\usage{
validate_is_null_or_string(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if an object is either null or a string (helper to save
some typing when validating optional object parameters)
}
