% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_number}
\alias{validate_is_number}
\title{validate_is_string}
\usage{
validate_is_number(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if an object is a number (helper to save
some typing when validating obligatory object parameters)
}
