% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_string}
\alias{validate_is_string}
\title{validate_is_string}
\usage{
validate_is_string(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if an object is a string (helper to save
some typing when validating obligatory object parameters)
}
