% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_is_string_flag}
\alias{validate_is_string_flag}
\title{validate_is_string_flag}
\usage{
validate_is_string_flag(...)
}
\arguments{
\item{...}{Ellipsis}
}
\description{
Checks if an object is a string reading either
"true" or "false"
}
