% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{validate_ogs6}
\alias{validate_ogs6}
\title{Validating functions for an ogs6 class object (to be run before a simulation is started)}
\usage{
validate_ogs6(ogs6_obj)
}
\description{
Validating functions for an ogs6 class object (to be run before a simulation is started)
}
