% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_param_list}
\alias{validate_param_list}
\title{validate_param_list}
\usage{
validate_param_list(param_list, expected_length, possible_names)
}
\arguments{
\item{param_list}{A list of parameters}

\item{expected_length}{The expected list length}

\item{possible_names}{How the list elements may be named (if the user DID name them)}
}
\description{
Validator function for a parameter list
}
