% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_paths}
\alias{validate_paths}
\title{validate_paths}
\usage{
validate_paths(sim_path, ogs_bin_path)
}
\arguments{
\item{sim_path}{The path where all relevant files for the simulation will be saved}

\item{ogs_bin_path}{Path to OpenGeoSys6 /bin directory}
}
\description{
Helper function to pull path validation out of already large class OGS6
}
