% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{validate_pnt_values}
\alias{validate_pnt_values}
\title{validate_pnt_values}
\usage{
validate_pnt_values(element_1, element_2)
}
\arguments{
\item{element_1}{numeric, length = 3}

\item{element_2}{numeric, length = 3}
}
\description{
Checks if two numerical vectors of length 3
(two surface elements) each consist of 3 different elements and have
exactly 2 matching elements between them. Think of the two vectors as
triangles, and the triangles together form a square which is our surface.
}
