% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gml.R
\name{validate_points}
\alias{validate_points}
\title{Checks if the input is a tibble, if this tibble has the right number of elements,
if those elements are named correctly, if the lists in the tibble are of the same length and
if there is any overlapping points or duplicate point names}
\usage{
validate_points(gml_points)
}
\arguments{
\item{gml_points}{A tibble with 3 vectors named 'x', 'y' and 'z' (and an optional 'name' vector)}
}
\description{
Checks if the input is a tibble, if this tibble has the right number of elements,
if those elements are named correctly, if the lists in the tibble are of the same length and
if there is any overlapping points or duplicate point names
}
