% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gml.R
\name{validate_polylines}
\alias{validate_polylines}
\title{validate_polylines}
\usage{
validate_polylines(gml_polylines)
}
\arguments{
\item{gml_polylines}{A list consisting of other lists}
}
\description{
Checks if the input is a list, if this list consists of other lists and
if those child lists have the correct structure (length of 2, first element is a string named
'name', second element is a numeric vector)
}
