% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_processes.R
\name{validate_r2ogs6_process}
\alias{validate_r2ogs6_process}
\title{validate_r2ogs6_process}
\usage{
validate_r2ogs6_process(r2ogs6_process)
}
\description{
Validator for r2ogs6_process class objects, contains more extensive checks than the constructor
}
