% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{validate_read_in_xml}
\alias{validate_read_in_xml}
\title{validate_read_in_xml}
\usage{
validate_read_in_xml(file)
}
\arguments{
\item{file}{A file to be parsed as XML}
}
\value{
The parsed XML file (as class object of type xml2::xml_document)
}
\description{
Utility function, tries parsing the provided file as an XML document
}
