% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{validate_surface_elements}
\alias{validate_surface_elements}
\title{validate_surface_elements}
\usage{
validate_surface_elements(surface_element_1, surface_element_2)
}
\arguments{
\item{surface_element_1}{numeric, length = 3}

\item{surface_element_2}{numeric, length = 3}
}
\description{
Helper function, checks if two numerical vectors of length 3
(two surface elements) each consist of 3 different elements and also have
exactly 2 matching elements between them which means they describe a valid
surface. You can think of the two vectors as two triangles, and the two
triangles together form a square which is our surface.
}
