% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gml.R
\name{validate_surfaces}
\alias{validate_surfaces}
\title{validate_surfaces}
\usage{
validate_surfaces(surfaces)
}
\arguments{
\item{surfaces}{A list consisting of other lists}
}
\description{
Checks if the input is a list, if this list consists of other lists and
if those child lists have the correct structure (length of 3, first element is a string named
'name', second and third element are numeric vectors)
}
