% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_wrapper_list}
\alias{validate_wrapper_list}
\title{validate_wrapper_list}
\usage{
validate_wrapper_list(wrapper_list, expected_element_class)
}
\arguments{
\item{wrapper_list}{The list to check}

\item{expected_class}{The class each element of the wrapper list should have}
}
\description{
Helper function, checks if a lists consists only of elements of a specific class
}
