% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_utils.R
\name{vector_from_nodeset}
\alias{vector_from_nodeset}
\title{vector_from_nodeset}
\usage{
vector_from_nodeset(xml_nodeset)
}
\arguments{
\item{xml_nodeset}{An XML nodeset (of class xml2::xml_nodeset)}
}
\description{
Creates a named vector from a nodeset
}
