% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_vtu.R
\name{vtu_appended_data_to_xml
(WIP)}
\alias{vtu_appended_data_to_xml
(WIP)}
\alias{vtu_appended_data_to_xml}
\title{vtu_appended_data_to_xml}
\usage{
vtu_appended_data_to_xml(appended_data)
}
\arguments{
\item{appended_data}{A list object consisting of 2 elements named 'data' and 'encoding'}
}
\description{
vtu_appended_data_to_xml
}
