% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_vtu.R
\name{vtu_data_to_xml}
\alias{vtu_data_to_xml}
\title{Wrapper function to create a .vtu XML document based on the user input data}
\usage{
vtu_data_to_xml(vtu_obj)
}
\arguments{
\item{vtu_obj}{A vtu class object}
}
\value{
A XML document ready for export to a file
}
\description{
Wrapper function to create a .vtu XML document based on the user input data
}
