% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_vtu.R
\name{vtu_piece_to_xml}
\alias{vtu_piece_to_xml}
\title{(WIP, change input parameter to piece!!)}
\usage{
vtu_piece_to_xml(
  number_of_points,
  number_of_cells,
  point_data,
  cell_data,
  points,
  cells
)
}
\arguments{
\item{piece}{A list consisting of lists of data_array class objects}
}
\description{
(WIP, change input parameter to piece!!)
}
