% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{xpaths_for_classes}
\alias{xpaths_for_classes}
\title{xpaths_for_classes}
\format{
list: A named list of character vectors where the names are r2ogs6
class names and the values are the corresponding xpaths
}
\usage{
xpaths_for_classes
}
\description{
Path to the .xml document containing information about the .prj file
structure
}
\keyword{datasets}
