% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adopt_nodes}
\alias{adopt_nodes}
\title{adopt_nodes}
\usage{
adopt_nodes(parent_name, objs)
}
\arguments{
\item{parent_name}{The name of the new parent node}

\item{objs}{A list of class objects (class must have method for generic function as_node)}
}
\description{
A helper function for creating parent nodes using the generic function as_node
}
