% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_xml.R
\name{analyse_xml}
\alias{analyse_xml}
\title{analyse_xml}
\usage{
analyse_xml(path, pattern, element_name)
}
\arguments{
\item{path}{A path}

\item{pattern}{A regex pattern}

\item{element_name}{The name of the XML element to look for}
}
\description{
This is a helper function which can be used to combat missing documentation. It looks
 for files in a path which match the given pattern and then attempts to parse them as XML.
 For each occurence of the specified element, it documents its attributes and direct children and prints a little
 summary of its findings at the end.
}
