% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_xml.R
\name{analyse_xml}
\alias{analyse_xml}
\title{analyse_xml}
\usage{
analyse_xml(path, pattern, xpath, print_findings = TRUE)
}
\arguments{
\item{path}{string: A path}

\item{pattern}{string: A regex pattern}

\item{xpath}{string: An XPath expression. WARNING: Only works for expressions
that return nodesets, use it to look up tag names.}

\item{print_findings}{Optional: Should the results be printed to the console?}
}
\description{
This is a helper function which can be used to combat missing
documentation. It looks for files in a path which match the given pattern
and then attempts to parse them as XML. For each occurence of the specified
element, it documents its attributes and direct children and prints a little
summary of its findings at the end.
}
