% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_node.default}
\alias{as_node.default}
\title{as_node.default}
\usage{
\method{as_node}{default}(obj)
}
\arguments{
\item{obj}{Any object}
}
\description{
Default implementation for S3 generic function as_node, just returns the object
passed to it (this may seem senseless, but is actually important for add_children
to work when as_node is called on values which are not r2ogs6 classes)
}
