% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_curves.R
\name{as_node.r2ogs6_curve}
\alias{as_node.r2ogs6_curve}
\title{as_node.r2ogs6_curve}
\usage{
\method{as_node}{r2ogs6_curve}(x)
}
\arguments{
\item{x}{A r2ogs6_curve class object}
}
\description{
Implementation of generic function as_node for S3 class r2ogs6_curve
}
