% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_node}
\alias{as_node}
\title{as_node}
\usage{
as_node(x, ...)
}
\arguments{
\item{x}{Any object}

\item{...}{In case arguments need to be passed from one method to another}
}
\description{
S3 generic function, use to turn class data into XML friendly format
}
