% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{build_redux_doc}
\alias{build_redux_doc}
\title{build_redux_doc}
\usage{
build_redux_doc(path, pattern, xpath, export_path)
}
\arguments{
\item{path}{string: See ?analyse_xml}

\item{pattern}{string: See ?analyse_xml}

\item{xpath}{string: See ?analyse_xml}

\item{export_path}{string: Path to export the XML document to}
}
\description{
Builds an XML document based on the findings of analyse_xml.
Calls recursive function `build_redux_tree` internally.
}
