% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{construct_add_call}
\alias{construct_add_call}
\title{construct_add_call}
\usage{
construct_add_call(object, nested_call = FALSE)
}
\arguments{
\item{object}{An object (numeric, character, list, NULL, or r2ogs6 class
object)}

\item{nested_call}{Optional: For recursion purposes, you should leave this as
it is.}
}
\value{
A string representing the code with which the component would be added
to an OGS6 object
}
\description{
Constructs a call based on an OGS6 component. This is a
recursive function, handle with care.
}
