% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{construct_ogs_command}
\alias{construct_ogs_command}
\title{construct_ogs_command}
\usage{
construct_ogs_command(ogs6_bin_path)
}
\arguments{
\item{ogs6_bin_path}{string: Optional: Path to OpenGeoSys 6 executable or
OpenGeoSys container (singularity image) file. Defaults
to \code{options("r2ogs6.default_ogs6_bin_path")}}
}
\value{
string: Call object.
}
\description{
Constructs the call string to for 'system2()'.
}
