% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gml.R
\name{export_gml}
\alias{export_gml}
\title{export_gml}
\usage{
export_gml(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}
}
\description{
Wrapper function to create a .gml XML document based on the user input data
}
