% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml.R
\name{export_xml_to_file}
\alias{export_xml_to_file}
\title{Export function}
\usage{
export_xml_to_file(xml_data, file_name)
}
\arguments{
\item{xml_data}{The .gml data (already in XML friendly format)}

\item{file_name}{The name of the .gml file to be written}
}
\description{
Export function
}
