% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{generate_all_benchmark_scripts}
\alias{generate_all_benchmark_scripts}
\title{generate_all_benchmark_scripts}
\usage{
generate_all_benchmark_scripts(
  path,
  sim_path,
  scripts_path,
  starting_from_prj_path = "",
  skip_prj_paths = character(),
  read_in_vtus = FALSE,
  read_in_gmls = TRUE
)
}
\arguments{
\item{path}{string: Path to a benchmark directory to generate scripts from}

\item{sim_path}{string: Path where all simulation files will be saved}

\item{scripts_path}{string: Path where benchmark scripts will be saved}

\item{starting_from_prj_path}{string: Optional:}

\item{skip_prj_paths}{character: Optional: .prj paths to skip}

\item{read_in_vtus}{flag: Optional: Should .vtu files just be copied or read
in too?}

\item{read_in_gmls}{flag: Optional: Should .gml files just be copied or read
in too?}
}
\description{
Wrapper function to generate benchmark scripts from all .prj
files in a directory
}
