% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{generate_benchmark_script}
\alias{generate_benchmark_script}
\title{generate_benchmark_script}
\usage{
generate_benchmark_script(prj_path, dest_dir = "")
}
\arguments{
\item{prj_path}{The path to the project file the script will be based on}

\item{dest_dir}{Optional: The directory to write the script to}
}
\description{
Generates a benchmark script from an existing .prj file.
}
