% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_benchmark_script.R
\name{generate_benchmark_script}
\alias{generate_benchmark_script}
\title{generate_benchmark_script}
\usage{
generate_benchmark_script(
  prj_path,
  sim_path,
  ogs_bin_path,
  script_path,
  read_in_vtu = FALSE,
  read_in_gml = TRUE
)
}
\arguments{
\item{prj_path}{string: .prj file the script will be based on}

\item{sim_path}{string: Path where all simulation files will be saved}

\item{ogs_bin_path}{string: OpenGeoSys bin folder path}

\item{script_path}{string: Path where benchmark script will be saved}

\item{read_in_vtu}{flag: Optional: Should .vtu file(s) just be copied or read
in too?}

\item{read_in_gml}{flag: Optional: Should .gml file just be copied or read
in too?}
}
\description{
Generates a benchmark script from an existing .prj file.
}
