% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_class.R
\name{generate_R6}
\alias{generate_R6}
\title{generate_R6}
\usage{
generate_R6(params, prefix = "", print_result = TRUE)
}
\arguments{
\item{params}{list: (Return value of analyse_xml())}

\item{prefix}{Optional: For subclasses whose represented elements have
the same tag name as an element for which a class was already specified,
a prefix must be appended to the class name}

\item{print_result}{flag: Should the result be printed to the console?}
}
\description{
Helper function to generate a R6 class out of a tag name
and a flag vector
}
