% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu_mesh.R
\name{generate_structured_mesh}
\alias{generate_structured_mesh}
\title{generate_structured_mesh}
\usage{
generate_structured_mesh(ogs6_obj, element_type, ...)
}
\arguments{
\item{ogs6_obj}{The OGS6 class object (containing the path to the ogs6 bin/ folder)}

\item{element_type}{Element type to be created: line | tri | quad | hex |
prism | tet | pyramid}
}
\value{
The newly generated .vtu file path
}
\description{
Wrapper function to call generateStructuredMesh.exe (VTK mesh generator).
For full documentation see https://www.opengeosys.org/docs/tools/meshing/structured-mesh-generation/
}
