% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{generate_structured_mesh}
\alias{generate_structured_mesh}
\title{generate_structured_mesh}
\usage{
generate_structured_mesh(ogs6_obj, call_str, read_in_vtu)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{call_str}{string: The arguments the script will be called with
(EXCEPT -o output_file_name, this will be generated automatically!)}

\item{read_in_vtu}{flag: Should .vtu file just be copied or read in too?}
}
\value{
string: .vtu file path
}
\description{
Wrapper function to call generateStructuredMesh.exe
(VTK mesh generator). For full documentation see
https://www.opengeosys.org/docs/tools/meshing/structured-mesh-generation/
}
