% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ogs.R
\name{install_ogs}
\alias{install_ogs}
\title{Install specific version of OpenGeoSys 6 (OGS) along with necessary Python 
packages}
\usage{
install_ogs(
  ...,
  ogs_version = "6.4.4",
  envname = "r2ogs6",
  py_version = "3.10:latest",
  py_exe = NULL,
  new_env = identical(envname, "r2ogs6")
)
}
\arguments{
\item{...}{Rest of the arguments passed to "py_install()"}

\item{ogs_version}{The version of OGS to install (default: 6.4.4).}

\item{envname}{The name of the Python virtual environment (default: r2ogs6).}
}
\value{
None.
}
\description{
This function handles the installation of a specific version of OGS and 
ensures that related Python packages such as `vtk` and `numpy` are 
installed in a specified Python virtual environment.
}
\examples{
install_ogs(
 ogs_version = "6.4.0", 
 envname = "r2ogs6"
)

}
