% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obj_is_defined}
\alias{obj_is_defined}
\title{obj_is_defined}
\usage{
obj_is_defined(flag, obj, obj_type = "", is_opt = FALSE)
}
\arguments{
\item{flag}{Boolean flag to keep track of missing components}

\item{obj}{The specified object}

\item{obj_type}{Optional: What kind of object is this?}

\item{is_opt}{flag: Is the element optional i.e. can it be NULL?}
}
\description{
Helper function for get_status() to check if an object was
defined
}
