% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{OGS6}
\alias{OGS6}
\title{OGS6}
\description{
Constructor for the OGS6 base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sim_name}}{Simulation name. `value` must be string}

\item{\code{sim_path}}{Simulation path. `value` must be string}

\item{\code{logfile}}{Logfile path. `value` must be string}

\item{\code{gml}}{.gml. read-only}

\item{\code{geometry}}{.prj `geometry` tag. `value` must be string}

\item{\code{meshes}}{.prj `meshes` tag. `value` must be list of strings}

\item{\code{vtus}}{.vtus. `value` must be list of `OGS_vtu` objects}

\item{\code{python_script}}{.prj `python_script` tag. `value` must be string}

\item{\code{search_length_algorithm}}{.prj `search_length_algorithm` tag. `value` must be
`r2ogs6_search_length_algorithm` object}

\item{\code{processes}}{.prj `processes` tag. `value` must be list of `r2ogs6_process` objects}

\item{\code{time_loop}}{.prj `time_loop` tag. `value` must be `r2ogs6_time_loop` object}

\item{\code{local_coordinate_system}}{.prj `local_coordinate_system` tag. `value` must be
`r2ogs6_local_coordinate_system` object}

\item{\code{media}}{.prj `media` tag. `value` must be list of `r2ogs6_medium` objects}

\item{\code{parameters}}{.prj `parameters` tag. `value` must be list of `r2ogs6_parameter`
objects}

\item{\code{chemical_system}}{.prj `chemical_system` tag. `value` must be `r2ogs6_chemical_system`
object}

\item{\code{curves}}{.prj `curves` tag. `value` must be list of `r2ogs6_curve` objects}

\item{\code{process_variables}}{.prj `process_variables` tag. `value` must be list of
`r2ogs6_process_variable` objects}

\item{\code{nonlinear_solvers}}{.prj `nonlinear_solvers` tag. `value` must be list of
`r2ogs6_nonlinear_solver` objects}

\item{\code{linear_solvers}}{.prj `linear_solvers` tag. `value` must be list of
`r2ogs6_linear_solver` objects}

\item{\code{test_definition}}{.prj `test_definition` tag. `value` must be list of `r2ogs6_vtkdiff`
objects}

\item{\code{insitu}}{.prj `insitu` tag. `value` must be `r2ogs6_insitu` object}

\item{\code{pvds}}{.pvds. `value` must be list of `OGS6_pvd` objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6$new()}}
\item \href{#method-add}{\code{OGS6$add()}}
\item \href{#method-add_gml}{\code{OGS6$add_gml()}}
\item \href{#method-add_vtu}{\code{OGS6$add_vtu()}}
\item \href{#method-get_status}{\code{OGS6$get_status()}}
\item \href{#method-print}{\code{OGS6$print()}}
\item \href{#method-print_log}{\code{OGS6$print_log()}}
\item \href{#method-clear}{\code{OGS6$clear()}}
\item \href{#method-clone}{\code{OGS6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$new(sim_name, sim_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sim_name}}{string: Simulation name}

\item{\code{sim_path}}{string: Path where all files for the simulation will be
saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Adds a .prj simulation component
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An object of any .prj `r2ogs6` class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_gml"></a>}}
\if{latex}{\out{\hypertarget{method-add_gml}{}}}
\subsection{Method \code{add_gml()}}{
Adds a reference to a .gml file and optionally, a OGS6_gml object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_gml(gml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gml}}{string | r2ogs6_gml: Either a path to a file with extension
.gml or a r2ogs6_gml object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_vtu"></a>}}
\if{latex}{\out{\hypertarget{method-add_vtu}{}}}
\subsection{Method \code{add_vtu()}}{
Adds a reference to a .vtu file and optionally, a OGS6_vtu object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_vtu(path, read_in_vtu = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{string:}

\item{\code{read_in_vtu}}{flag: Optional: Should .vtu file just be copied or
read in too?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Checks if the OGS6 object has all necessary parameters for
starting a simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$get_status(print_status = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print_status}}{flag: Should the status be printed to the console?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Overrides default printing behaviour
print_log
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_log"></a>}}
\if{latex}{\out{\hypertarget{method-print_log}{}}}
\subsection{Method \code{print_log()}}{
Prints logfile to console (if it exists)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$print_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
Clears components from the OGS6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clear(which)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{character: The names of the components (all by default).
If you want to delete only some components, run
names(prj_top_level_classes()) for the available options.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
