% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{OGS6}
\alias{OGS6}
\title{OGS6}
\description{
Constructor for the OGS6 base class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6$new()}}
\item \href{#method-add_sim_output}{\code{OGS6$add_sim_output()}}
\item \href{#method-add_mesh}{\code{OGS6$add_mesh()}}
\item \href{#method-add_gml}{\code{OGS6$add_gml()}}
\item \href{#method-add_process}{\code{OGS6$add_process()}}
\item \href{#method-add_time_loop}{\code{OGS6$add_time_loop()}}
\item \href{#method-add_medium}{\code{OGS6$add_medium()}}
\item \href{#method-add_parameter}{\code{OGS6$add_parameter()}}
\item \href{#method-add_curve}{\code{OGS6$add_curve()}}
\item \href{#method-add_process_variable}{\code{OGS6$add_process_variable()}}
\item \href{#method-add_nonlinear_solver}{\code{OGS6$add_nonlinear_solver()}}
\item \href{#method-add_linear_solver}{\code{OGS6$add_linear_solver()}}
\item \href{#method-add_vtkdiff}{\code{OGS6$add_vtkdiff()}}
\item \href{#method-get_status}{\code{OGS6$get_status()}}
\item \href{#method-clear}{\code{OGS6$clear()}}
\item \href{#method-clone}{\code{OGS6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$new(sim_name, sim_id, sim_path, ogs_bin_path, test_mode = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sim_name}}{The name of the simulation}

\item{\code{sim_id}}{The ID of the simulation}

\item{\code{sim_path}}{The path where all relevant files for the simulation will be saved}

\item{\code{ogs_bin_path}}{Path to OpenGeoSys6 /bin directory}

\item{\code{test_mode}}{In test mode, sim_path and ogs_bin_path will not be validated. If you're not
a developer, please leave this variable as it is :)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_sim_output"></a>}}
\if{latex}{\out{\hypertarget{method-add_sim_output}{}}}
\subsection{Method \code{add_sim_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_sim_output(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_mesh"></a>}}
\if{latex}{\out{\hypertarget{method-add_mesh}{}}}
\subsection{Method \code{add_mesh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_mesh(mesh)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_gml"></a>}}
\if{latex}{\out{\hypertarget{method-add_gml}{}}}
\subsection{Method \code{add_gml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_gml(gml)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_process"></a>}}
\if{latex}{\out{\hypertarget{method-add_process}{}}}
\subsection{Method \code{add_process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_process(process)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_time_loop"></a>}}
\if{latex}{\out{\hypertarget{method-add_time_loop}{}}}
\subsection{Method \code{add_time_loop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_time_loop(time_loop)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_medium"></a>}}
\if{latex}{\out{\hypertarget{method-add_medium}{}}}
\subsection{Method \code{add_medium()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_medium(medium)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-add_parameter}{}}}
\subsection{Method \code{add_parameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_parameter(parameter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_curve"></a>}}
\if{latex}{\out{\hypertarget{method-add_curve}{}}}
\subsection{Method \code{add_curve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_curve(curve)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_process_variable"></a>}}
\if{latex}{\out{\hypertarget{method-add_process_variable}{}}}
\subsection{Method \code{add_process_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_process_variable(process_variable)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_nonlinear_solver"></a>}}
\if{latex}{\out{\hypertarget{method-add_nonlinear_solver}{}}}
\subsection{Method \code{add_nonlinear_solver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_nonlinear_solver(nonlinear_solver)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_linear_solver"></a>}}
\if{latex}{\out{\hypertarget{method-add_linear_solver}{}}}
\subsection{Method \code{add_linear_solver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_linear_solver(linear_solver)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_vtkdiff"></a>}}
\if{latex}{\out{\hypertarget{method-add_vtkdiff}{}}}
\subsection{Method \code{add_vtkdiff()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_vtkdiff(vtkdiff)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$get_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clear(
  which = c("meshes", "geometry", "processes", "time_loop", "media", "parameters",
    "curves", "process_variables", "nonlinear_solvers", "linear_solvers",
    "test_definition")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
