% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6.R
\name{OGS6}
\alias{OGS6}
\title{OGS6}
\description{
Constructor for the OGS6 base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sim_name}}{Access to private parameter '.sim_name'}

\item{\code{sim_id}}{Getter for OGS6 private parameter '.sim_id'}

\item{\code{sim_path}}{Access to private parameter '.sim_path'}

\item{\code{logfile}}{Access to private parameter '.logfile'}

\item{\code{ogs_bin_path}}{Getter for OGS6 private parameter '.ogs_bin_path'}

\item{\code{gml}}{Getter for OGS6 private parameter '.gml'}

\item{\code{geometry}}{Access to private parameter '.geometry'}

\item{\code{meshes}}{Access to private parameter '.meshes'}

\item{\code{vtus}}{Access to private parameter '.vtus'}

\item{\code{python_script}}{Access to private parameter '.python_script'}

\item{\code{search_length_algorithm}}{Access to private parameter '.search_length_algorithm'}

\item{\code{processes}}{Access to private parameter '.processes'}

\item{\code{time_loop}}{Access to private parameter '.time_loop'}

\item{\code{local_coordinate_system}}{Access to private parameter '.local_coordinate_system'}

\item{\code{media}}{Access to private parameter '.media'}

\item{\code{parameters}}{Access to private parameter '.parameters'}

\item{\code{curves}}{Access to private parameter '.curves'}

\item{\code{process_variables}}{Access to private parameter '.process_variables'}

\item{\code{nonlinear_solvers}}{Access to private parameter '.nonlinear_solvers'}

\item{\code{linear_solvers}}{Access to private parameter '.linear_solvers'}

\item{\code{test_definition}}{Access to private parameter '.test_definition'}

\item{\code{insitu}}{Access to private parameter '.insitu'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6$new()}}
\item \href{#method-add_component}{\code{OGS6$add_component()}}
\item \href{#method-add_mesh}{\code{OGS6$add_mesh()}}
\item \href{#method-add_gml}{\code{OGS6$add_gml()}}
\item \href{#method-add_vtu}{\code{OGS6$add_vtu()}}
\item \href{#method-add_python_script}{\code{OGS6$add_python_script()}}
\item \href{#method-add_search_length_algorithm}{\code{OGS6$add_search_length_algorithm()}}
\item \href{#method-add_process}{\code{OGS6$add_process()}}
\item \href{#method-add_time_loop}{\code{OGS6$add_time_loop()}}
\item \href{#method-add_local_coordinate_system}{\code{OGS6$add_local_coordinate_system()}}
\item \href{#method-add_medium}{\code{OGS6$add_medium()}}
\item \href{#method-add_parameter}{\code{OGS6$add_parameter()}}
\item \href{#method-add_curve}{\code{OGS6$add_curve()}}
\item \href{#method-add_process_variable}{\code{OGS6$add_process_variable()}}
\item \href{#method-add_nonlinear_solver}{\code{OGS6$add_nonlinear_solver()}}
\item \href{#method-add_linear_solver}{\code{OGS6$add_linear_solver()}}
\item \href{#method-add_vtkdiff}{\code{OGS6$add_vtkdiff()}}
\item \href{#method-add_insitu}{\code{OGS6$add_insitu()}}
\item \href{#method-get_status}{\code{OGS6$get_status()}}
\item \href{#method-print_log}{\code{OGS6$print_log()}}
\item \href{#method-clear}{\code{OGS6$clear()}}
\item \href{#method-clone}{\code{OGS6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$new(sim_name, sim_id, sim_path, ogs_bin_path, test_mode = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sim_name}}{string: Simulation name}

\item{\code{sim_id}}{double: Simulation ID}

\item{\code{sim_path}}{string: Path where all files for the simulation will be
saved}

\item{\code{ogs_bin_path}}{string: Path to OpenGeoSys6 /bin directory}

\item{\code{test_mode}}{In test mode, ogs_bin_path will not be
validated. Unless you're a dev, please don't touch.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_component"></a>}}
\if{latex}{\out{\hypertarget{method-add_component}{}}}
\subsection{Method \code{add_component()}}{
Adds a simulation component (WIP)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_component(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{OGS6_*: An object of a class listed in get_implemented_classes()}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_mesh"></a>}}
\if{latex}{\out{\hypertarget{method-add_mesh}{}}}
\subsection{Method \code{add_mesh()}}{
Adds a reference to a .vtu file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_mesh(mesh)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mesh}}{string: .vtu path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_gml"></a>}}
\if{latex}{\out{\hypertarget{method-add_gml}{}}}
\subsection{Method \code{add_gml()}}{
Adds a r2ogs6_gml object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_gml(gml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gml}}{r2ogs6_gml}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_vtu"></a>}}
\if{latex}{\out{\hypertarget{method-add_vtu}{}}}
\subsection{Method \code{add_vtu()}}{
Adds a r2ogs6_gml object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_vtu(vtu, filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vtu}}{r2ogs6_vtu}

\item{\code{filename}}{string:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_python_script"></a>}}
\if{latex}{\out{\hypertarget{method-add_python_script}{}}}
\subsection{Method \code{add_python_script()}}{
Adds a python script
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_python_script(python_script)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{python_script}}{string: File name of python script}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_search_length_algorithm"></a>}}
\if{latex}{\out{\hypertarget{method-add_search_length_algorithm}{}}}
\subsection{Method \code{add_search_length_algorithm()}}{
Adds a r2ogs6_search_length_algorithm object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_search_length_algorithm(search_length_algorithm)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_length_algorithm}}{r2ogs6_search_length_algorithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_process"></a>}}
\if{latex}{\out{\hypertarget{method-add_process}{}}}
\subsection{Method \code{add_process()}}{
Adds a r2ogs6_process object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_process(process)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{r2ogs6_process}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_time_loop"></a>}}
\if{latex}{\out{\hypertarget{method-add_time_loop}{}}}
\subsection{Method \code{add_time_loop()}}{
Adds a r2ogs6_time_loop object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_time_loop(time_loop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_loop}}{r2ogs6_time_loop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_local_coordinate_system"></a>}}
\if{latex}{\out{\hypertarget{method-add_local_coordinate_system}{}}}
\subsection{Method \code{add_local_coordinate_system()}}{
Adds a r2ogs6_local_coordinate_system object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_local_coordinate_system(local_coordinate_system)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{local_coordinate_system}}{r2ogs6_local_coordinate_system}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_medium"></a>}}
\if{latex}{\out{\hypertarget{method-add_medium}{}}}
\subsection{Method \code{add_medium()}}{
Adds a r2ogs6_medium object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_medium(medium)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{medium}}{r2ogs6_medium}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-add_parameter}{}}}
\subsection{Method \code{add_parameter()}}{
Adds a r2ogs6_parameter object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_parameter(parameter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{r2ogs6_parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_curve"></a>}}
\if{latex}{\out{\hypertarget{method-add_curve}{}}}
\subsection{Method \code{add_curve()}}{
Adds a r2ogs6_curve object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_curve(curve)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{curve}}{r2ogs6_curve}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_process_variable"></a>}}
\if{latex}{\out{\hypertarget{method-add_process_variable}{}}}
\subsection{Method \code{add_process_variable()}}{
Adds a r2ogs6_process_variable object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_process_variable(process_variable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process_variable}}{r2ogs6_process_variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_nonlinear_solver"></a>}}
\if{latex}{\out{\hypertarget{method-add_nonlinear_solver}{}}}
\subsection{Method \code{add_nonlinear_solver()}}{
Adds a r2ogs6_nonlinear_solver object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_nonlinear_solver(nonlinear_solver)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nonlinear_solver}}{r2ogs6_nonlinear_solver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_linear_solver"></a>}}
\if{latex}{\out{\hypertarget{method-add_linear_solver}{}}}
\subsection{Method \code{add_linear_solver()}}{
Adds a r2ogs6_linear_solver object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_linear_solver(linear_solver)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linear_solver}}{r2ogs6_linear_solver}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_vtkdiff"></a>}}
\if{latex}{\out{\hypertarget{method-add_vtkdiff}{}}}
\subsection{Method \code{add_vtkdiff()}}{
Adds a r2ogs6_vtkdiff object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_vtkdiff(vtkdiff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vtkdiff}}{r2ogs6_vtkdiff}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_insitu"></a>}}
\if{latex}{\out{\hypertarget{method-add_insitu}{}}}
\subsection{Method \code{add_insitu()}}{
Adds a r2ogs6_insitu object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$add_insitu(insitu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{insitu}}{r2ogs6_insitu}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Checks if the OGS6 object has all necessary parameters for
starting a simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$get_status(print_status = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print_status}}{flag: Should the status be printed to the console?
print_log}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_log"></a>}}
\if{latex}{\out{\hypertarget{method-print_log}{}}}
\subsection{Method \code{print_log()}}{
Prints logfile to console (if it exists)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$print_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
Clears components from the OGS6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clear(which = names(get_implemented_classes()))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{character: The names of the components (all by default).
If you want to delete only some components, run
names(get_implemented_classes()) for the available options.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
