% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_ensemble.R
\name{OGS6_Ensemble}
\alias{OGS6_Ensemble}
\title{OGS6_Ensemble}
\description{
Constructor for the OGS6_Ensemble base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dp_parameters}}{Getter for private parameter '.dp_parameters'}

\item{\code{parameter_percs}}{Getter for private parameter '.parameter_percs'}

\item{\code{parameter_values}}{Getter for private parameter '.parameter_values'}

\item{\code{ens_path}}{Getter for private parameter '.ens_path'}

\item{\code{ensemble}}{Getter for private parameter '.ensemble'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_Ensemble$new()}}
\item \href{#method-print}{\code{OGS6_Ensemble$print()}}
\item \href{#method-run_simulation}{\code{OGS6_Ensemble$run_simulation()}}
\item \href{#method-get_point_data}{\code{OGS6_Ensemble$get_point_data()}}
\item \href{#method-clone}{\code{OGS6_Ensemble$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_Ensemble object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$new(
  ogs6_obj,
  parameters,
  sequential_mode = FALSE,
  percentages_mode = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ogs6_obj}}{OGS6: A simulation object.}

\item{\code{parameters}}{list(sublist, length(sublist) == 2): The first element
of a sublist references an OGS6 parameter, the second one is a list or
vector of values. Note that the second elements of the sublists must
have the same length.}

\item{\code{sequential_mode}}{flag: Defaults to `FALSE`}

\item{\code{percentages_mode}}{flag: Defaults to `TRUE`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Overrides default printing behaviour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-run_simulation}{}}}
\subsection{Method \code{run_simulation()}}{
Runs the simulation. This calls r2ogs6::ogs_run_simulation()
internally. For ensembles, output will always be written to logfiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$run_simulation(parallel = FALSE, verbose = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parallel}}{flag: Should the function be run in parallel?
This is implementented via the 'parallel' package.}

\item{\code{verbose}}{flag}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_point_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_point_data}{}}}
\subsection{Method \code{get_point_data()}}{
Wrapper for `OGS6_pvd$get_point_data()` Returns combined dataframe
with extra `name` and / or `perc` column depending on if
`sequential_mode` and / or `percentages_mode` were used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$get_point_data(
  pvd_id = 1,
  point_ids,
  keys,
  start_at_timestep,
  end_at_timestep
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pvd_id}}{number: Optional: Which .pvd to consider in OGS6$pvds.
Defaults to 1.}

\item{\code{point_ids}}{numeric: Optional: Point IDs. Defaults to all.}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements. Defaults to all.}

\item{\code{start_at_timestep}}{number: Optional: Timestep to start at.
Defaults to first timestep.}

\item{\code{end_at_timestep}}{number: Optional: Timestep to end at. Defaults
to last timestep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
