% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_ensemble.R
\name{OGS6_Ensemble}
\alias{OGS6_Ensemble}
\title{OGS6_Ensemble}
\description{
Constructor for the OGS6_Ensemble base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameters}}{Getter for private parameter '.parameters'}

\item{\code{ens_path}}{Getter for private parameter '.ens_path'}

\item{\code{ensemble}}{Getter for private parameter '.ensemble'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_Ensemble$new()}}
\item \href{#method-run_simulation}{\code{OGS6_Ensemble$run_simulation()}}
\item \href{#method-clone}{\code{OGS6_Ensemble$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_Ensemble object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$new(ogs6_obj, parameters, ens_dir_name = "Ensemble")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ogs6_obj}}{OGS6: A simulation object.}

\item{\code{parameters}}{list(sublist, length(sublist) == 2): The first element
of a sublist references an OGS6 parameter, the second one is a list or
vector of values. Note that the second elements of the sublists must
have the same length.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-run_simulation}{}}}
\subsection{Method \code{run_simulation()}}{
Runs the simulation. This calls r2ogs6::run_simulation() internally.
For ensembles, output will be written to logfiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$run_simulation(parallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parallel}}{flag: Should the function be run in parallel?
This is implementented via the 'parallel' package.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_Ensemble$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
