% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_pvd.R
\name{OGS6_pvd}
\alias{OGS6_pvd}
\title{OGS6_pvd}
\description{
Constructor for the OGS6_pvd base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pvd_path}}{Getter for private parameter '.pvd_path'}

\item{\code{datasets}}{Getter for private parameter '.datasets'}

\item{\code{vtu_paths}}{Getter for `datasets` `file`}

\item{\code{abs_vtu_paths}}{Gets absolute .vtu paths, e.g. `dirname(pvd_path)` + `datasets` `file`}

\item{\code{last_timestep}}{Gets last timestep}

\item{\code{timesteps}}{Gets timesteps from private parameter '.datasets'}

\item{\code{OGS6_vtus}}{Getter for private parameter '.OGS6_vtus'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_pvd$new()}}
\item \href{#method-print}{\code{OGS6_pvd$print()}}
\item \href{#method-vtu_by_timestep}{\code{OGS6_pvd$vtu_by_timestep()}}
\item \href{#method-timestep_by_vtu}{\code{OGS6_pvd$timestep_by_vtu()}}
\item \href{#method-get_point_data_at}{\code{OGS6_pvd$get_point_data_at()}}
\item \href{#method-get_point_data}{\code{OGS6_pvd$get_point_data()}}
\item \href{#method-get_cell_data}{\code{OGS6_pvd$get_cell_data()}}
\item \href{#method-clone}{\code{OGS6_pvd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_pvd object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$new(pvd_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pvd_path}}{string: Path to .pvd file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Overrides default printing behaviour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vtu_by_timestep"></a>}}
\if{latex}{\out{\hypertarget{method-vtu_by_timestep}{}}}
\subsection{Method \code{vtu_by_timestep()}}{
Returns .vtu path for specified timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$vtu_by_timestep(timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timestep}}{string: Timestep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-timestep_by_vtu"></a>}}
\if{latex}{\out{\hypertarget{method-timestep_by_vtu}{}}}
\subsection{Method \code{timestep_by_vtu()}}{
Returns timestep for specified .vtu path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$timestep_by_vtu(vtu_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vtu_path}}{string: .vtu path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_point_data_at"></a>}}
\if{latex}{\out{\hypertarget{method-get_point_data_at}{}}}
\subsection{Method \code{get_point_data_at()}}{
Returns a tibble containing point data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_point_data_at(
  coordinates,
  keys,
  start_at_timestep,
  end_at_timestep
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coordinates}}{list(numeric): List of coordinates (a coordinate
is a numeric vector of length 3)}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements. Defaults to all.}

\item{\code{start_at_timestep}}{number: Optional: Timestep to start at.
Defaults to first timestep.}

\item{\code{end_at_timestep}}{number: Optional: Timestep to end at. Defaults
to last timestep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_point_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_point_data}{}}}
\subsection{Method \code{get_point_data()}}{
Returns a tibble containing point data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_point_data(point_ids, keys, start_at_timestep, end_at_timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{point_ids}}{numeric: Optional: Point IDs. Defaults to all.}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements. Defaults to all.}

\item{\code{start_at_timestep}}{number: Optional: Timestep to start at.
Defaults to first timestep.}

\item{\code{end_at_timestep}}{number: Optional: Timestep to end at. Defaults
to last timestep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cell_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_cell_data}{}}}
\subsection{Method \code{get_cell_data()}}{
Returns a tibble containing cell data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$get_cell_data(cell_ids, keys, start_at_timestep, end_at_timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell_ids}}{numeric: Optional: Cell IDs. Defaults to all.}

\item{\code{keys}}{character: Optional: `Name` attributes of `DataArray`
elements. Defaults to all.}

\item{\code{start_at_timestep}}{number: Optional: Timestep to start at.
Defaults to first timestep.}

\item{\code{end_at_timestep}}{number: Optional: Timestep to end at. Defaults
to last timestep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_pvd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
