% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{OGS6_vtu}
\alias{OGS6_vtu}
\title{OGS6_vtu}
\description{
Constructor for the OGS6_vtu base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{vtu_path}}{Getter for private parameter '.vtu_path'}

\item{\code{vtkUnstructuredGrid}}{Access to private parameter '.vtkUnstructuredGrid'}

\item{\code{dsa_wrapped_vtkUnstructuredGrid}}{Getter for private parameter '.dsa_wrapped_vtkUnstructuredGrid'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_vtu$new()}}
\item \href{#method-get_PointData_for_point}{\code{OGS6_vtu$get_PointData_for_point()}}
\item \href{#method-get_PointData}{\code{OGS6_vtu$get_PointData()}}
\item \href{#method-get_number_of_points}{\code{OGS6_vtu$get_number_of_points()}}
\item \href{#method-clone}{\code{OGS6_vtu$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_vtu object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$new(vtu_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vtu_path}}{string: Path to .vtu file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_PointData_for_point"></a>}}
\if{latex}{\out{\hypertarget{method-get_PointData_for_point}{}}}
\subsection{Method \code{get_PointData_for_point()}}{
Gets PointData for point with ID `point_id`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_PointData_for_point(point_id, Names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{point_id}}{number: Point ID}

\item{\code{Names}}{character: Optional: `Name` attributes of `DataArray`
elements, defaults to all in `PointData`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_PointData"></a>}}
\if{latex}{\out{\hypertarget{method-get_PointData}{}}}
\subsection{Method \code{get_PointData()}}{
Gets PointData of `DataArray` element with `Name` attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_PointData(Name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Name}}{string: Optional: `Name` attribute of `DataArray`
elements, defaults to all in `PointData`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_number_of_points"></a>}}
\if{latex}{\out{\hypertarget{method-get_number_of_points}{}}}
\subsection{Method \code{get_number_of_points()}}{
Gets number of points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$get_number_of_points()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
number: The number of points
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
