% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtu.R
\name{OGS6_vtu}
\alias{OGS6_vtu}
\title{OGS6_vtu}
\description{
Constructor for the OGS6_vtu base class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{Access to private parameter '.type'}

\item{\code{version}}{Access to private parameter '.version'}

\item{\code{byte_order}}{Access to private parameter '.byte_order'}

\item{\code{UnstructuredGrid}}{Access to private parameter '.UnstructuredGrid'}

\item{\code{header_type}}{Access to private parameter '.header_type'}

\item{\code{compressor}}{Access to private parameter '.compressor'}

\item{\code{AppendedData}}{Access to private parameter '.AppendedData '}

\item{\code{is_subclass}}{Access to private parameter '.is_subclass'}

\item{\code{attr_names}}{Access to private parameter '.attr_names'}

\item{\code{flatten_on_exp}}{Access to private parameter '.flatten_on_exp'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OGS6_vtu$new()}}
\item \href{#method-clone}{\code{OGS6_vtu$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates new OGS6_vtu object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$new(
  type,
  version,
  byte_order,
  UnstructuredGrid,
  header_type = NULL,
  compressor = NULL,
  AppendedData = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{string:}

\item{\code{version}}{string:}

\item{\code{byte_order}}{string:}

\item{\code{UnstructuredGrid}}{OGS6_UnstructuredGrid:}

\item{\code{header_type}}{string: Optional:}

\item{\code{compressor}}{string: Optional:}

\item{\code{AppendedData}}{string: Optional:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OGS6_vtu$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
