% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gml.R
\name{polylines_to_xml}
\alias{polylines_to_xml}
\title{Turns a list of polylines into an XML node}
\usage{
polylines_to_xml(polyline_list)
}
\arguments{
\item{polyline_list}{The specified list}
}
\value{
An XML node containing the polylines
}
\description{
Turns a list of polylines into an XML node
}
\examples{
my_list <- list(list(name = "front_left", c(0, 1)), list(name = "front_right", c(4, 5)))
polylines_node <- polylines_to_xml(my_list)
}
