% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{prj_chemical_system}
\alias{prj_chemical_system}
\title{prj_chemical_system}
\usage{
prj_chemical_system(
  chemical_solver,
  mesh,
  linear_solver = NULL,
  database = NULL,
  solution = NULL,
  knobs = NULL,
  kinetic_reactants = NULL,
  rates = NULL,
  equilibrium_reactants = NULL,
  surface = NULL,
  user_punch = NULL,
  use_high_precision = NULL,
  exchangers = NULL,
  chemical_reactions = NULL,
  number_of_components = NULL
)
}
\arguments{
\item{chemical_solver}{string: Either "Phreeqc" or "SelfContained".}

\item{mesh}{string:}

\item{linear_solver}{Optional: string:}

\item{database}{string: Required if chemical_solver == "Phreeqc" .}

\item{solution}{prj_solution: Required if chemical_solver == "Phreeqc".}

\item{knobs}{Optional: list:}

\item{kinetic_reactants}{Optional: list, prj_kinetic_reactant:}

\item{rates}{Optional: list, prj_rate:}

\item{equilibrium_reactants}{Optional: list, prj_phase_component:}

\item{surface}{Optional:}

\item{user_punch}{Optional:}

\item{use_high_precision}{Optional:}

\item{exchangers}{Optional: prj_exchangers}

\item{chemical_reactions}{list: Required if
chemical_solver == "SelfContained".}

\item{number_of_components}{integer: Required if
chemical_solver == "SelfContained".}
}
\description{
tag: chemical_system
}
\examples{
prj_chemical_system(
    chemical_solver = "Phreeqc",
    linear_solver = "general_linear_solver",
    database = "PSINA_12_07_110615_DAV_s.dat",
    solution = prj_solution(
        temperature = 25,
        pressure = 1,
        pe = 4,
        components = list(
            component = "C(4)",
            component = "Ca"
        ),
        charge_balance = "pH"
    ),
    mesh = "calcite_ReactiveDomain",
    knobs = list(
        max_iter = "100",
        relative_convergence_tolerance = "1e-12",
        tolerance = "1e-15",
        step_size = "100",
        scaling = "0"
    ),
    equilibrium_reactants = list(
        phase_component = prj_phase_component(
            name = "Calcite",
            saturation_index = 0,
            initial_amount = 0.000207
        )
    )
)
}
