% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{prj_com_property}
\alias{prj_com_property}
\title{prj_com_property}
\usage{
prj_com_property(
  name,
  type,
  value = NULL,
  parameter_name = NULL,
  reference_diffusion = NULL,
  activation_energy = NULL,
  reference_temperature = NULL,
  triple_temperature = NULL,
  triple_pressure = NULL,
  critical_temperature = NULL,
  critical_pressure = NULL,
  reference_pressure = NULL
)
}
\arguments{
\item{name}{string: Property name}

\item{type}{string: Property type}

\item{value}{Optional: string | double: ...}

\item{parameter_name}{Optional:}

\item{reference_diffusion}{Optional: character}

\item{activation_energy}{Optional: string | double}

\item{reference_temperature}{Optional: numeric}

\item{triple_temperature}{Optional: numeric}

\item{triple_pressure}{Optional: numeric}

\item{critical_temperature}{Optional: numeric}

\item{critical_pressure}{Optional: numeric}

\item{reference_pressure}{Optional: numeric}
}
\description{
tag: property
}
\examples{
prj_com_property(name = "molecular_diffusion",
                    type = "Constant",
                    value = 2e-09)
}
