% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_local_coordinate_system.R
\name{prj_local_coordinate_system}
\alias{prj_local_coordinate_system}
\title{prj_local_coordinate_system}
\usage{
prj_local_coordinate_system(
  basis_vector_0,
  basis_vector_1,
  basis_vector_2 = NULL
)
}
\arguments{
\item{basis_vector_0}{prj_basis_vector_0: A basis vector}

\item{basis_vector_1}{prj_basis_vector_1: A basis vector}

\item{basis_vector_2}{Optional prj_basis_vector_2: A basis vector}
}
\description{
tag: local_coordinate_system
}
\examples{
e0 <- prj_basis_vector_0(name = "e0")
e1 <- prj_basis_vector_1(name = "e1")
e2 <- prj_basis_vector_2(implicit = "true")

prj_local_coordinate_system(basis_vector_0 = e0,
                            basis_vector_1 = e1,
                            basis_vector_2 = e2)
}
