% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{prj_process}
\alias{prj_process}
\title{prj_process}
\usage{
prj_process(
  name,
  type,
  integration_order,
  process_variables,
  secondary_variables = NULL,
  specific_body_force = NULL,
  solid_density = NULL,
  dimension = NULL,
  coupling_scheme = NULL,
  reference_solid_density = NULL,
  linear_thermal_expansion_coefficient = NULL,
  specific_heat_capacity = NULL,
  thermal_conductivity = NULL,
  darcy_gravity = NULL,
  reference_temperature = NULL,
  fracture_model = NULL,
  jacobian_assembler = NULL,
  internal_length = NULL,
  mass_lumping = NULL,
  porosity = NULL,
  calculatesurfaceflux = NULL,
  intrinsic_permeability = NULL,
  specific_storage = NULL,
  fluid_viscosity = NULL,
  biot_coefficient = NULL,
  fluid_density = NULL,
  initial_effective_stress = NULL,
  initial_fracture_effective_stress = NULL,
  phasefield_parameters = NULL,
  deactivate_matrix_in_flow = NULL,
  borehole_heat_exchangers = NULL,
  temperature = NULL,
  reactive_system = NULL,
  fluid_specific_heat_source = NULL,
  fluid_specific_isobaric_heat_capacity = NULL,
  solid_hydraulic_permeability = NULL,
  solid_specific_heat_source = NULL,
  solid_heat_conductivity = NULL,
  solid_specific_isobaric_heat_capacity = NULL,
  tortuosity = NULL,
  diffusion_coefficient = NULL,
  solid_density_dry = NULL,
  solid_density_initial = NULL,
  characteristic_pressure = NULL,
  characteristic_temperature = NULL,
  characteristic_vapour_mass_fraction = NULL,
  output_element_matrices = NULL,
  material_property = NULL,
  diffusion_coeff_component_b = NULL,
  diffusion_coeff_component_a = NULL,
  hydro_crack_scheme = NULL,
  at_num = NULL,
  initial_stress = NULL,
  split_method = NULL,
  reg_param = NULL,
  thermal_parameters = NULL,
  non_advective_form = NULL,
  fluid = NULL,
  porous_medium = NULL,
  decay_rate = NULL,
  fluid_reference_density = NULL,
  retardation_factor = NULL,
  solute_dispersivity_longitudinal = NULL,
  solute_dispersivity_transverse = NULL,
  molecular_diffusion_coefficient = NULL,
  density_solid = NULL,
  latent_heat_evaporation = NULL,
  pf_irrv = NULL,
  ...
)
}
\arguments{
\item{name}{string:}

\item{type}{string:}

\item{integration_order}{string | double:}

\item{process_variables}{list, character:}

\item{secondary_variables}{Optional:}

\item{specific_body_force}{Optional:}

\item{solid_density}{Optional:}

\item{dimension}{Optional:}

\item{coupling_scheme}{Optional:}

\item{reference_solid_density}{Optional:}

\item{linear_thermal_expansion_coefficient}{Optional:}

\item{specific_heat_capacity}{Optional:}

\item{thermal_conductivity}{Optional:}

\item{darcy_gravity}{Optional:}

\item{reference_temperature}{Optional:}

\item{fracture_model}{Optional:}

\item{jacobian_assembler}{Optional:}

\item{internal_length}{Optional:}

\item{mass_lumping}{Optional:}

\item{porosity}{Optional:}

\item{calculatesurfaceflux}{Optional:}

\item{intrinsic_permeability}{Optional:}

\item{specific_storage}{Optional:}

\item{fluid_viscosity}{Optional:}

\item{biot_coefficient}{Optional:}

\item{fluid_density}{Optional:}

\item{initial_effective_stress}{Optional:}

\item{initial_fracture_effective_stress}{Optional:}

\item{phasefield_parameters}{Optional:}

\item{deactivate_matrix_in_flow}{Optional:}

\item{borehole_heat_exchangers}{Optional:}

\item{temperature}{Optional:}

\item{reactive_system}{Optional:}

\item{fluid_specific_heat_source}{Optional:}

\item{fluid_specific_isobaric_heat_capacity}{Optional:}

\item{solid_hydraulic_permeability}{Optional:}

\item{solid_specific_heat_source}{Optional:}

\item{solid_heat_conductivity}{Optional:}

\item{solid_specific_isobaric_heat_capacity}{Optional:}

\item{tortuosity}{Optional:}

\item{diffusion_coefficient}{Optional:}

\item{solid_density_dry}{Optional:}

\item{solid_density_initial}{Optional:}

\item{characteristic_pressure}{Optional:}

\item{characteristic_temperature}{Optional:}

\item{characteristic_vapour_mass_fraction}{Optional:}

\item{output_element_matrices}{Optional:}

\item{material_property}{Optional:}

\item{diffusion_coeff_component_b}{Optional:}

\item{diffusion_coeff_component_a}{Optional:}

\item{hydro_crack_scheme}{Optional:}

\item{at_num}{Optional:}

\item{initial_stress}{Optional:}

\item{split_method}{Optional:}

\item{reg_param}{Optional:}

\item{thermal_parameters}{Optional:}

\item{non_advective_form}{Optional:}

\item{fluid}{Optional:}

\item{porous_medium}{Optional:}

\item{decay_rate}{Optional:}

\item{fluid_reference_density}{Optional:}

\item{retardation_factor}{Optional:}

\item{solute_dispersivity_longitudinal}{Optional:}

\item{solute_dispersivity_transverse}{Optional:}

\item{molecular_diffusion_coefficient}{Optional:}

\item{density_solid}{Optional:}

\item{latent_heat_evaporation}{Optional:}

\item{pf_irrv}{Optional:}

\item{...}{Optional: fracture_properties, constitutive_relation}
}
\description{
tag: process (parent tag: processes)
}
\examples{
prj_process(
    name = "HM",
    type = "HYDRO_MECHANICS",
    integration_order = 3,
    process_variables = list(displacement = "displacement",
                             pressure = "pressure"),
    secondary_variables = list(
        secondary_variable = c(internal_name = "sigma_xx", output_name = "sigma_xx"),
        secondary_variable = c(internal_name = "sigma_yy", output_name = "sigma_yy")
    ),
    specific_body_force = c(0, 0, 0),
    dimension = 3,
    constitutive_relation = prj_constitutive_relation(
        type = "LinearElasticIsotropic",
        youngs_modulus = "E",
        poissons_ratio = "nu"
    )
)
}
