% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{r2ogs6_chemical_system}
\alias{r2ogs6_chemical_system}
\title{r2ogs6_chemical_system}
\usage{
r2ogs6_chemical_system(
  chemical_solver,
  database,
  solution,
  mesh = NULL,
  knobs = NULL,
  kinetic_reactants = NULL,
  rates = NULL,
  equilibrium_reactants = NULL,
  surface = NULL,
  user_punch = NULL
)
}
\arguments{
\item{chemical_solver}{string:}

\item{database}{string:}

\item{solution}{r2ogs6_solution:}

\item{mesh}{Optional: string:}

\item{knobs}{Optional: list:}

\item{kinetic_reactants}{Optional: list, r2ogs6_kinetic_reactant:}

\item{rates}{Optional: list, r2ogs6_rate:}

\item{equilibrium_reactants}{Optional: list, r2ogs6_phase_component:}

\item{surface}{Optional:}

\item{user_punch}{Optional:}
}
\description{
tag: chemical_system
}
