% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_component}
\alias{r2ogs6_component}
\title{r2ogs6_component}
\usage{
r2ogs6_component(name, properties)
}
\arguments{
\item{name}{string:}

\item{properties}{list, r2ogs6_com_property:}
}
\description{
tag: component
}
\examples{
r2ogs6_component(
    name = "concentration",
    properties = list(
        property = r2ogs6_com_property(name = "molecular_diffusion",
                                       type = "Constant",
                                       value = 2e-09)
    )
)
}
