% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_curves.R
\name{r2ogs6_curve}
\alias{r2ogs6_curve}
\title{r2ogs6_curve}
\usage{
r2ogs6_curve(name, coords, values)
}
\arguments{
\item{name}{The name of the curve}

\item{coords}{Coordinates at which the curve's values are given}

\item{values}{Values of the curve at the given coordinates}
}
\description{
S3 class describing a .prj curve
}
