% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process.R
\name{r2ogs6_fracture_properties}
\alias{r2ogs6_fracture_properties}
\title{r2ogs6_fracture_properties}
\usage{
r2ogs6_fracture_properties(
  material_id,
  initial_aperture,
  specific_storage = NULL,
  biot_coefficient = NULL,
  permeability_model = NULL
)
}
\arguments{
\item{material_id}{string | double:}

\item{initial_aperture}{string:}

\item{specific_storage}{Optional: string:}

\item{biot_coefficient}{Optional: string:}

\item{permeability_model}{Optional: list:}
}
\description{
tag: fracture_properties
}
\examples{
r2ogs6_fracture_properties(
    material_id = 0,
    initial_aperture = "aperture0",
    specific_storage = "S_f",
    biot_coefficient = "biot_f",
    permeability_model = list(type = "CubicLaw")
)
}
