% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_global_process_coupling}
\alias{r2ogs6_global_process_coupling}
\title{r2ogs6_global_process_coupling}
\usage{
r2ogs6_global_process_coupling(max_iter, convergence_criteria)
}
\arguments{
\item{max_iter}{string | double: Maximal number of iterations}

\item{convergence_criteria}{list, r2ogs6_convergence_criterion:
Convergence criteria}
}
\description{
tag: global_process_coupling
}
