% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_linear_solver.R
\name{r2ogs6_linear_solver}
\alias{r2ogs6_linear_solver}
\title{r2ogs6_linear_solver}
\usage{
r2ogs6_linear_solver(name, eigen, lis = NULL, petsc = NULL)
}
\arguments{
\item{name}{The name of the linear solver}

\item{eigen}{...}

\item{lis}{...}

\item{petsc}{...}
}
\description{
S3 class describing a .prj linear solver
}
