% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_medium}
\alias{r2ogs6_medium}
\title{r2ogs6_medium}
\usage{
r2ogs6_medium(phases = NULL, properties = NULL, id = NULL)
}
\arguments{
\item{phases}{list, r2ogs6_phase: Optional: Medium phases}

\item{properties}{list, r2ogs6_pr_property: Optional: Medium properties}

\item{id}{string | double: Optional: ID corresponding to the MaterialIDs}
}
\description{
tag: medium, a specific medium with optional id corresponding
  to the MaterialIDs
}
\examples{
r2ogs6_medium(
    phases = list(
        phase = r2ogs6_phase(
            type = "Gas",
            properties = list(
                property = r2ogs6_ph_property(name = "viscosity",
                                              type = "Constant",
                                              value = 1e-05)
            )
        ),
        phase = r2ogs6_phase(
            type = "Solid",
            properties = list(
                property = r2ogs6_ph_property(name = "porosity",
                                              type = "Constant",
                                              value = 0.3)
            )
        )
    ),
    properties = list(
        property = r2ogs6_pr_property(name = "reference_temperature",
                                      type = "Constant",
                                      value = 293.15)
    )
)
}
