% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_media.R
\name{r2ogs6_medium}
\alias{r2ogs6_medium}
\title{r2ogs6_medium}
\usage{
r2ogs6_medium(phases, properties, id = NULL)
}
\arguments{
\item{phases}{A list of medium phases (r2ogs6_medium_phase objects)}

\item{properties}{A list of medium properties (r2ogs6_medium_property objects)}

\item{id}{Optional: ID corresponding to the MaterialIDs}
}
\description{
A specific medium with optional id corresponding to the MaterialIDs
}
