% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_nonlinear_solver.R
\name{r2ogs6_nonlinear_solver}
\alias{r2ogs6_nonlinear_solver}
\title{r2ogs6_nonlinear_solver}
\usage{
r2ogs6_nonlinear_solver(name, type, max_iter, linear_solver, damping = NULL)
}
\arguments{
\item{name}{string: Name}

\item{type}{string: Type}

\item{max_iter}{string | double: Maximal number of iterations}

\item{linear_solver}{string: Name of corresponding linear_solver}

\item{damping}{Optional: string | double: Damping}
}
\description{
tag: nonlinear_solver
}
