% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_parameter.R
\name{r2ogs6_parameter}
\alias{r2ogs6_parameter}
\title{r2ogs6_parameter}
\usage{
r2ogs6_parameter(name, type, values = NULL, value = NULL)
}
\arguments{
\item{name}{The parameter name}

\item{type}{The parameter type}

\item{values}{Optional: The parameter values}

\item{value}{Optional: The parameter value}
}
\description{
S3 class describing a .prj parameter
}
