% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_chemical_system.R
\name{r2ogs6_phase_component}
\alias{r2ogs6_phase_component}
\title{r2ogs6_phase_component}
\usage{
r2ogs6_phase_component(name, initial_amount, saturation_index)
}
\arguments{
\item{name}{The component name}

\item{initial_amount}{The initial amount of the component}

\item{saturation_index}{The saturation index of the component}
}
\description{
S3 class describing .prj phase_component
}
