% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_porous_medium.R
\name{r2ogs6_porous_medium}
\alias{r2ogs6_porous_medium}
\title{r2ogs6_porous_medium}
\usage{
r2ogs6_porous_medium(
  id,
  permeability,
  porosity,
  storage,
  capillary_pressure,
  relative_permeability
)
}
\arguments{
\item{id}{string | double:}

\item{permeability}{list:}

\item{porosity}{list:}

\item{storage}{list:}

\item{capillary_pressure}{r2ogs6_capillary_pressure:}

\item{relative_permeability}{list:}
}
\description{
tag: porous_medium
}
