% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_medium.R
\name{r2ogs6_pr_property}
\alias{r2ogs6_pr_property}
\title{r2ogs6_pr_property}
\usage{
r2ogs6_pr_property(
  name,
  type,
  value = NULL,
  parameter_name = NULL,
  exponent = NULL,
  residual_liquid_saturation = NULL,
  residual_gas_saturation = NULL,
  p_b = NULL,
  independent_variable = NULL,
  curve = NULL,
  minimum_relative_permeability_liquid = NULL,
  lambda = NULL,
  cutoff_value = NULL,
  initial_permeability = NULL,
  reference_permeability = NULL,
  fitting_factor = NULL,
  cohesion = NULL,
  friction_angle = NULL,
  maximum_permeability = NULL,
  tensile_strength_parameter = NULL,
  entry_pressure = NULL,
  min_relative_permeability_liquid = NULL,
  min_relative_permeability_gas = NULL
)
}
\arguments{
\item{name}{string: Property name}

\item{type}{string: Property type}

\item{value}{Optional: string | double: ...}

\item{parameter_name}{Optional: string:}

\item{exponent}{Optional: string | double:}

\item{residual_liquid_saturation}{Optional: string | double:}

\item{residual_gas_saturation}{Optional: string | double:}

\item{p_b}{Optional: string | double: string | double:}

\item{independent_variable}{Optional: string:}

\item{curve}{Optional: string:}

\item{minimum_relative_permeability_liquid}{Optional:}

\item{lambda}{Optional: string | double:}

\item{cutoff_value}{Optional: string | double:}

\item{initial_permeability}{Optional: string:}

\item{reference_permeability}{Optional: string | double:}

\item{fitting_factor}{Optional: string | double:}

\item{cohesion}{Optional: string | double:}

\item{friction_angle}{Optional: string | double:}

\item{maximum_permeability}{Optional: string | double:}

\item{tensile_strength_parameter}{Optional: string | double:}

\item{entry_pressure}{Optional: string | double:}

\item{min_relative_permeability_liquid}{Optional: string | double:}

\item{min_relative_permeability_gas}{Optional: string | double:}
}
\description{
tag: property, a constitutive property
}
