% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variable.R
\name{r2ogs6_process_variable}
\alias{r2ogs6_process_variable}
\title{r2ogs6_process_variable}
\usage{
r2ogs6_process_variable(
  name,
  components,
  order,
  initial_condition,
  boundary_conditions = NULL,
  source_terms = NULL,
  mesh = NULL,
  deactivated_subdomains = NULL
)
}
\arguments{
\item{name}{string: The name of the process variable}

\item{components}{string | double:}

\item{order}{string | double:}

\item{initial_condition}{string:}

\item{boundary_conditions}{list, r2ogs6_boundary_condition:}

\item{source_terms}{Optional: list, r2ogs6_source_term:}

\item{mesh}{Optional: string: list:}

\item{deactivated_subdomains}{Optional: list, r2ogs6_deactivated_subdomain:}
}
\description{
tag: process_variable
}
