% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_process_variables.R
\name{r2ogs6_process_variable}
\alias{r2ogs6_process_variable}
\title{r2ogs6_process_variable}
\usage{
r2ogs6_process_variable(
  name,
  components,
  order,
  initial_condition,
  boundary_conditions
)
}
\arguments{
\item{name}{The name of the process variable}

\item{components}{The components of the process variable}

\item{order}{The order of the process variable}

\item{initial_condition}{The initial condition of the process variable}

\item{boundary_conditions}{The boundary conditions of the process variable}
}
\description{
S3 class describing a .prj process variable
}
