% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_time_loop}
\alias{r2ogs6_time_loop}
\title{r2ogs6_time_loop}
\usage{
r2ogs6_time_loop(processes, output, global_processes_coupling = NULL)
}
\arguments{
\item{processes}{...}

\item{output}{...}

\item{global_processes_coupling}{...}
}
\description{
S3 class describing a .prj time_loop
}
