% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_time_loop.R
\name{r2ogs6_tl_process}
\alias{r2ogs6_tl_process}
\title{r2ogs6_tl_process}
\usage{
r2ogs6_tl_process(
  ref,
  nonlinear_solver,
  convergence_criterion,
  time_discretization,
  time_stepping,
  compensate_non_equilibrium_initial_residuum = NULL
)
}
\arguments{
\item{ref}{string: References a r2ogs6_process object by name}

\item{nonlinear_solver}{string:}

\item{convergence_criterion}{r2ogs6_convergence_criterion:}

\item{time_discretization}{vector:}

\item{time_stepping}{r2ogs6_time_stepping:}

\item{compensate_non_equilibrium_initial_residuum}{string: Optional: Either
"true" or "false"}
}
\description{
tag: process (parent: time_loop, NOT processes!)
}
