% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_vtkdiff.R
\name{r2ogs6_vtkdiff}
\alias{r2ogs6_vtkdiff}
\title{r2ogs6_vtkdiff}
\usage{
r2ogs6_vtkdiff(regex, field, absolute_tolerance, relative_tolerance)
}
\arguments{
\item{regex}{...}

\item{field}{...}

\item{absolute_tolerance}{...}

\item{relative_tolerance}{...}
}
\description{
S3 class describing a .prj test_definition vtkdiff
}
