% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_vtkdiff.R
\name{r2ogs6_vtkdiff}
\alias{r2ogs6_vtkdiff}
\title{r2ogs6_vtkdiff}
\usage{
r2ogs6_vtkdiff(regex, field, absolute_tolerance, relative_tolerance)
}
\arguments{
\item{regex}{string: A regular expression}

\item{field}{string: ...}

\item{absolute_tolerance}{string | double: Absolute tolerance}

\item{relative_tolerance}{string | double: Relative tolerance}
}
\description{
tag: vtkdiff
}
