% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_gml.R
\name{read_in_gml}
\alias{read_in_gml}
\title{read_in_gml}
\usage{
read_in_gml(ogs6_obj, gml_path)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{gml_path}{The path to the geometry file that should be read in}
}
\description{
Wrapper function to read in a whole .gml file
}
