% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{read_in_output}
\alias{read_in_output}
\title{read_in_output}
\usage{
read_in_output(ogs6_obj)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}
}
\description{
After a OGS6 simulation was run, reads in the generated .vtu
files as new input for
the .prj file
}
