% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_prj}
\alias{read_in_prj}
\title{read_in_prj}
\usage{
read_in_prj(ogs6_obj, prj_path)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{prj_path}{The path to the project file that should be read in}
}
\description{
Wrapper function to read in a whole .prj file
}
