% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_prj.R
\name{read_in_prj}
\alias{read_in_prj}
\title{read_in_prj}
\usage{
read_in_prj(ogs6_obj, prj_path, read_in_vtu = FALSE, read_in_gml = TRUE)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{prj_path}{string: Path to the project file that should be read in}

\item{read_in_vtu}{flag: Should .vtu file just be copied or read in too?}

\item{read_in_gml}{flag: Should .gml file just be copied or read in too?}
}
\description{
Wrapper function to read in a whole .prj file
}
