% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_vtu.R
\name{read_in_vtu}
\alias{read_in_vtu}
\title{read_in_vtu}
\usage{
read_in_vtu(ogs6_obj, vtu_path)
}
\arguments{
\item{ogs6_obj}{OGS6: Simulation object}

\item{vtu_path}{The path to the mesh file that should be read in}
}
\description{
Wrapper function to read in a whole .vtu file
}
