% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{run_all_benchmarks}
\alias{run_all_benchmarks}
\title{run_all_benchmarks}
\usage{
run_all_benchmarks(
  path,
  ogs_bin_path,
  sim_path,
  starting_from_prj_path = "",
  print_failed_prj_paths = TRUE
)
}
\arguments{
\item{path}{string:}

\item{ogs_bin_path}{string:}

\item{sim_path}{string:}

\item{starting_from_prj_path}{string: .prj path to start from}

\item{print_failed_prj_paths}{flag: Output paths where `read_in_prj()` failed?}
}
\description{
Utility function, for quick benchmark runs. Calls
run_benchmark internally.
}
