% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{run_benchmark}
\alias{run_benchmark}
\title{run_benchmark}
\usage{
run_benchmark(prj_path, ogs_bin_path, sim_path = "D:/OGS_all_simulations/")
}
\arguments{
\item{prj_path}{string:}

\item{ogs_bin_path}{string:}

\item{sim_path}{string: Optional: Path where simulation files will be saved.
Change this to fit your system!}
}
\description{
Utility function for quick benchmark runs
}
