% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(ogs6_obj, iter_n = 1, output_to_log_file = TRUE)
}
\arguments{
\item{ogs6_obj}{A OGS6 class object}

\item{iter_n}{The number of iterations (for simulation chains)}

\item{output_to_log_file}{Should the output be written to a log file?}
}
\description{
Calls OGS6 object validator functions, exports all necessary
files and starts OpenGeoSys6
}
