% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_param_list}
\alias{validate_param_list}
\title{validate_param_list}
\usage{
validate_param_list(param_list, default_names)
}
\arguments{
\item{param_list}{A list (or vector) of parameters}

\item{default_names}{How the list elements will be named as per default}
}
\description{
Validator function for a parameter list or vector
}
