% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{validate_points}
\alias{validate_points}
\title{validate_points}
\usage{
validate_points(points)
}
\arguments{
\item{points}{tibble: Must have 3 vectors named 'x', 'y' and 'z', may have
optional 'name' vector}
}
\description{
Checks if the input is a tibble, if this tibble has the right
number of elements, if those elements are named correctly and if there are
any overlapping points or duplicate point names
}
