% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs6_gml.R
\name{validate_polylines}
\alias{validate_polylines}
\title{validate_polylines}
\usage{
validate_polylines(polylines)
}
\arguments{
\item{polylines}{list(list("foo", c(1, 2))):}
}
\description{
Checks if the input is a list, if this list consists of other
lists and if those lists have the correct structure (length of 2, first
element is a string named 'name', second element is a numeric vector)
}
