% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{validate_r2ogs6_gml}
\alias{validate_r2ogs6_gml}
\title{validate_r2ogs6_gml}
\usage{
validate_r2ogs6_gml(r2ogs6_gml)
}
\arguments{
\item{r2ogs6_gml}{r2ogs6_gml:}
}
\description{
Validator for class r2ogs6_gml. Checks if the defined polylines
and surfaces reference existing points.
}
