% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gml.R
\name{validate_surfaces}
\alias{validate_surfaces}
\title{validate_surfaces}
\usage{
validate_surfaces(surfaces)
}
\arguments{
\item{surfaces}{list(list("foo", c(1, 2, 3), c(2, 3, 4))):}
}
\description{
Checks if the input is a list, if this list consists of other
lists and if those lists have the correct structure (length of 3, first
element is a string named 'name', second and third element are numeric
vectors)
}
